Documentation of Epidemiology API fields

Documents fields included when extracting data from outbreak.info

epidemiologyDataDictionary()

Value

dataframe

Examples

df = epidemiologyDataDictionary()
knitr::kable(df)
#> 
#> 
#> |API Field                         |Documentation                                                                                                                                            |
#> |:---------------------------------|:--------------------------------------------------------------------------------------------------------------------------------------------------------|
#> |admin_level                       |Administrative level (World Bank regions = -1, countries = 0, states/provinces = 1, metropolitan areas = 1.5, counties = 2)                              |
#> |cbsa                              |Metropolitan area FIPS code                                                                                                                              |
#> |confirmed                         |Total number of confirmed COVID-19 cases                                                                                                                 |
#> |confirmed_doublingRate            |Doubling rate of confirmed COVID-19 cases (number of days for COVID-19 cases to double)                                                                  |
#> |confirmed_firstDate               |Date of first confirmed COVID-19 case                                                                                                                    |
#> |confirmed_newToday                |T if new COVID-19 cases reported, F if none                                                                                                              |
#> |confirmed_numIncrease             |Number of new confirmed COVID-19 cases                                                                                                                   |
#> |confirmed_pctIncrease             |Percent increase in confirmed COVID-19 cases                                                                                                             |
#> |confirmed_per_100k                |Total number of confirmed COVID-19 cases per 100,000 persons                                                                                             |
#> |confirmed_rolling                 |Weekly rolling average of new confirmed COVID-19 cases                                                                                                   |
#> |confirmed_rolling_14days_ago      |Weekly rolling average of new confirmed COVID-19 cases 14 days prior                                                                                     |
#> |confirmed_rolling_14days_ago_diff |Difference between a weekly rolling average of new confirmed COVID-19 cases and the weekly rolling average of new confirmed COVID-19 cases 14 days prior |
#> |confirmed_rolling_per_100k        |Weekly rolling average of new confirmed COVID-19 cases per 100,000 persons                                                                               |
#> |country_gdp_per_capita            |Country GDP per capita                                                                                                                                   |
#> |country_iso3                      |Country ISO3                                                                                                                                             |
#> |country_name                      |Country name                                                                                                                                             |
#> |country_population                |Total population of country                                                                                                                              |
#> |date                              |Date                                                                                                                                                     |
#> |daysSince100Cases                 |Days since 100 new confirmed cases of COVID-19 reported                                                                                                  |
#> |daysSince10Deaths                 |Days since 10 new deaths due to COVID-19 reported                                                                                                        |
#> |daysSince50Deaths                 |Days since 50 new deaths due to COVID-19 reported                                                                                                        |
#> |dead                              |Total number of deaths due to COVID-19                                                                                                                   |
#> |dead_doublingRate                 |Doubling rate of deaths due to COVID-19 (number of days for deaths due to COVID-19 to double)                                                            |
#> |dead_firstDate                    |Date of first death due to COVID-19                                                                                                                      |
#> |dead_newToday                     |T if new deaths due to COVID-19 reported, F if none                                                                                                      |
#> |dead_numIncrease                  |Number of new deaths due to COVID-19                                                                                                                     |
#> |dead_pctIncrease                  |Percent increase in deaths due to COVID-19                                                                                                               |
#> |dead_per_100k                     |Total number of deaths due to COVID-19 per 100,000 persons                                                                                               |
#> |dead_rolling                      |Weekly rolling average of new deaths due to COVID-19                                                                                                     |
#> |dead_rolling_14days_ago           |Weekly rolling average of new deaths due to COVID-19 14 days prior                                                                                       |
#> |dead_rolling_14days_ago_diff      |Difference between a weekly rolling average of new deaths due to COVID-19 and the weekly rolling average of new deaths due to COVID-19 14 days prior     |
#> |dead_rolling_per_100k             |Weekly rolling average of new deaths due to COVID-19 per 100,000 persons                                                                                 |
#> |first_dead-first_confirmed        |Number of days between first confirmed case of COVID-19 and first death due to COVID-19                                                                  |
#> |gdp_last_updated                  |Year that GDP was last updated                                                                                                                           |
#> |gdp_per_capita                    |GDP per capita                                                                                                                                           |
#> |iso3                              |ISO3 code                                                                                                                                                |
#> |lat                               |Latitude                                                                                                                                                 |
#> |location_id                       |Location code                                                                                                                                            |
#> |long                              |Longitude                                                                                                                                                |
#> |mostRecent                        |T for most recent row of data, F for all others                                                                                                          |
#> |name                              |Location name                                                                                                                                            |
#> |num_subnational                   |Number of administrative divisions                                                                                                                       |
#> |population                        |Total population                                                                                                                                         |
#> |state_iso3                        |State ISO3 code                                                                                                                                          |
#> |state_name                        |State name                                                                                                                                               |
#> |sub_parts                         |County name, county FIPS code, state name                                                                                                                |
#> |wb_region                         |World Bank region                                                                                                                                        |