Retrieve mutations by lineage

Retrieve all mutations in a specified lineage above a threshold

getMutationsByLineage(pangolin_lineage, frequency = 0.8)

Arguments

pangolin_lineage:
PANGO lineage name or list

frequency:
a number between 0 and 1 specifying the frequency threshold above which to return mutations (default=0.8)

Value

dataframe

Examples

getMutationsByLineage(pangolin_lineage="P.1", frequency=0.8)

#> Retrieving data... 

#>              mutation mutation_count lineage_count lineage  gene
#> 1             s:d614g          71709         72479     P.1     S
#> 2            s:v1176f          71660         72479     P.1     S
#> 3        orf1a:k1795q          71562         72479     P.1 ORF1a
#> 4         orf1b:p314l          71527         72479     P.1 ORF1b
#> 5              n:p80r          71480         72479     P.1     N
#> 6        orf1b:e1264d          71390         72479     P.1 ORF1b
#> 7         orf3a:s253p          71327         72479     P.1 ORF3a
#> 8             s:h655y          71208         72479     P.1     S
#> 9           orf8:e92k          70980         72479     P.1  ORF8
#> 10       orf1a:s1188l          70949         72479     P.1 ORF1a
#> 11             s:l18f          70740         72479     P.1     S
#> 12             s:p26s          70633         72479     P.1     S
#> 13             s:t20n          70368         72479     P.1     S
#> 14 orf1a:del3675/3677          70364         72479     P.1 ORF1a
#> 15           s:t1027i          70134         72479     P.1     S
#> 16            s:d138y          69857         72479     P.1     S
#> 17            s:n501y          69321         72479     P.1     S
#> 18            s:e484k          69173         72479     P.1     S
#> 19            s:k417t          69137         72479     P.1     S
#> 20            n:g204r          68165         72479     P.1     N
#> 21            n:r203k          67782         72479     P.1     N
#> 22            s:r190s          67516         72479     P.1     S
#>                ref_aa       alt_aa codon_num codon_end         type prevalence
#> 1                   D            G       614      None substitution  0.9893762
#> 2                   V            F      1176      None substitution  0.9887002
#> 3                   K            Q      1795      None substitution  0.9873481
#> 4                   P            L       314      None substitution  0.9868652
#> 5                   P            R        80      None substitution  0.9862167
#> 6                   E            D      1264      None substitution  0.9849750
#> 7                   S            P       253      None substitution  0.9841057
#> 8                   H            Y       655      None substitution  0.9824639
#> 9                   E            K        92      None substitution  0.9793181
#> 10                  S            L      1188      None substitution  0.9788904
#> 11                  L            F        18      None substitution  0.9760068
#> 12                  P            S        26      None substitution  0.9745306
#> 13                  T            N        20      None substitution  0.9708743
#> 14 ORF1A:DEL3675/3677 DEL3675/3677      3675      3677     deletion  0.9708191
#> 15                  T            I      1027      None substitution  0.9676458
#> 16                  D            Y       138      None substitution  0.9638240
#> 17                  N            Y       501      None substitution  0.9564288
#> 18                  E            K       484      None substitution  0.9543868
#> 19                  K            T       417      None substitution  0.9538901
#> 20                  G            R       204      None substitution  0.9404793
#> 21                  R            K       203      None substitution  0.9351950
#> 22                  R            S       190      None substitution  0.9315250
#>    change_length_nt query_key
#> 1              None       P.1
#> 2              None       P.1
#> 3              None       P.1
#> 4              None       P.1
#> 5              None       P.1
#> 6              None       P.1
#> 7              None       P.1
#> 8              None       P.1
#> 9              None       P.1
#> 10             None       P.1
#> 11             None       P.1
#> 12             None       P.1
#> 13             None       P.1
#> 14                9       P.1
#> 15             None       P.1
#> 16             None       P.1
#> 17             None       P.1
#> 18             None       P.1
#> 19             None       P.1
#> 20             None       P.1
#> 21             None       P.1
#> 22             None       P.1

# get all delta lineages
getMutationsByLineage(pangolin_lineage="P.1", frequency=0.8)

#> Retrieving data... 

#>              mutation mutation_count lineage_count lineage  gene
#> 1             s:d614g          71709         72479     P.1     S
#> 2            s:v1176f          71660         72479     P.1     S
#> 3        orf1a:k1795q          71562         72479     P.1 ORF1a
#> 4         orf1b:p314l          71527         72479     P.1 ORF1b
#> 5              n:p80r          71480         72479     P.1     N
#> 6        orf1b:e1264d          71390         72479     P.1 ORF1b
#> 7         orf3a:s253p          71327         72479     P.1 ORF3a
#> 8             s:h655y          71208         72479     P.1     S
#> 9           orf8:e92k          70980         72479     P.1  ORF8
#> 10       orf1a:s1188l          70949         72479     P.1 ORF1a
#> 11             s:l18f          70740         72479     P.1     S
#> 12             s:p26s          70633         72479     P.1     S
#> 13             s:t20n          70368         72479     P.1     S
#> 14 orf1a:del3675/3677          70364         72479     P.1 ORF1a
#> 15           s:t1027i          70134         72479     P.1     S
#> 16            s:d138y          69857         72479     P.1     S
#> 17            s:n501y          69321         72479     P.1     S
#> 18            s:e484k          69173         72479     P.1     S
#> 19            s:k417t          69137         72479     P.1     S
#> 20            n:g204r          68165         72479     P.1     N
#> 21            n:r203k          67782         72479     P.1     N
#> 22            s:r190s          67516         72479     P.1     S
#>                ref_aa       alt_aa codon_num codon_end         type prevalence
#> 1                   D            G       614      None substitution  0.9893762
#> 2                   V            F      1176      None substitution  0.9887002
#> 3                   K            Q      1795      None substitution  0.9873481
#> 4                   P            L       314      None substitution  0.9868652
#> 5                   P            R        80      None substitution  0.9862167
#> 6                   E            D      1264      None substitution  0.9849750
#> 7                   S            P       253      None substitution  0.9841057
#> 8                   H            Y       655      None substitution  0.9824639
#> 9                   E            K        92      None substitution  0.9793181
#> 10                  S            L      1188      None substitution  0.9788904
#> 11                  L            F        18      None substitution  0.9760068
#> 12                  P            S        26      None substitution  0.9745306
#> 13                  T            N        20      None substitution  0.9708743
#> 14 ORF1A:DEL3675/3677 DEL3675/3677      3675      3677     deletion  0.9708191
#> 15                  T            I      1027      None substitution  0.9676458
#> 16                  D            Y       138      None substitution  0.9638240
#> 17                  N            Y       501      None substitution  0.9564288
#> 18                  E            K       484      None substitution  0.9543868
#> 19                  K            T       417      None substitution  0.9538901
#> 20                  G            R       204      None substitution  0.9404793
#> 21                  R            K       203      None substitution  0.9351950
#> 22                  R            S       190      None substitution  0.9315250
#>    change_length_nt query_key
#> 1              None       P.1
#> 2              None       P.1
#> 3              None       P.1
#> 4              None       P.1
#> 5              None       P.1
#> 6              None       P.1
#> 7              None       P.1
#> 8              None       P.1
#> 9              None       P.1
#> 10             None       P.1
#> 11             None       P.1
#> 12             None       P.1
#> 13             None       P.1
#> 14                9       P.1
#> 15             None       P.1
#> 16             None       P.1
#> 17             None       P.1
#> 18             None       P.1
#> 19             None       P.1
#> 20             None       P.1
#> 21             None       P.1
#> 22             None       P.1