Retrieve daily prevalence of a lineage by location

Retrieves the daily prevalence of a PANGO lineage(s) by location

getPrevalenceByLocation(
  pangolin_lineage = NULL,
  location = NULL,
  mutations = NULL,
  cumulative = NULL
)

Arguments

pangolin_lineage:
(optional) PANGO lineage name or vector of PANGO lineage names. Either `pangolin_lineage` or `mutations` needs to be specified. A list of lineages will return a long dataframe with `lineage` as a variable; if you want to calculate the prevalence of lineage1 or lineage2 together, enter the lineages separated by " OR ". For instance, to calculate the prevalence of Delta, you'll need to supply `"B.1.617.2 OR AY.1 OR AY.2 OR ..."` **Be sure to include the space around "OR" and it must be capitalized.**

mutations:
(optional) a `vector` of mutation(s). Either `pangolin_lineage` or `mutations` needs to be specified.

location:
(optional) a location name

cumulative:
(optional) `Boolean` (T/F), T returns cumulative prevalence since first day of detection

Value

dataframe

Examples

# lineage: P.1 in Brazil
getPrevalenceByLocation(pangolin_lineage = "P.1", location = "Brazil") %>% head()

#> Retrieving data... 
#> Retrieving data... 

#>         date total_count lineage_count total_count_rolling
#> 1 2020-09-11          17             1            14.66667
#> 2 2020-09-12           8             0            14.66667
#> 3 2020-09-13           2             0            12.85714
#> 4 2020-09-14          17             0            14.71429
#> 5 2020-09-15          35             1            16.28571
#> 6 2020-09-16          18             0            16.42857
#>   lineage_count_rolling  proportion proportion_ci_lower proportion_ci_upper
#> 1             0.1666667 0.011363636        3.219414e-05           0.1518167
#> 2             0.1666667 0.011363636        3.219414e-05           0.1518167
#> 3             0.1428571 0.011111111        3.705193e-05           0.1726194
#> 4             0.1428571 0.009708738        3.219414e-05           0.1518167
#> 5             0.2857143 0.017543860        3.021348e-05           0.1431841
#> 6             0.2857143 0.017391304        3.021348e-05           0.1431841
#>   lineage
#> 1     P.1
#> 2     P.1
#> 3     P.1
#> 4     P.1
#> 5     P.1
#> 6     P.1

# AY.4 and B.1.617.2 in Brazil (AY.4, AY.34, and B.1.617.2 separately)
getPrevalenceByLocation(pangolin_lineage = c("AY.4", "AY.34", "B.1.617.2"), location = "Brazil") %>% filter(date == "2021-09-01")

#> Retrieving data... 
#> Retrieving data... 

#>         date total_count lineage_count total_count_rolling
#> 1 2021-09-01          28             0                  33
#> 2 2021-09-01          28            16                  33
#>   lineage_count_rolling proportion proportion_ci_lower proportion_ci_upper
#> 1               0.00000  0.0000000        1.476743e-05          0.07275871
#> 2              13.33333  0.4040404        2.418367e-01          0.56375817
#>     lineage
#> 1      AY.4
#> 2 B.1.617.2

# AY.4 OR B.1.617.2 in Brazil (AY.4, AY.34 and B.1.617.2 combined)
getPrevalenceByLocation(pangolin_lineage = "AY.4 OR AY.34 OR B.1.617.2", location = "Brazil") %>% filter(date == "2021-09-01")

#> Retrieving data... 
#> Retrieving data... 

#>         date total_count lineage_count total_count_rolling
#> 1 2021-09-01          28            16                  33
#>   lineage_count_rolling proportion proportion_ci_lower proportion_ci_upper
#> 1              13.33333  0.4040404           0.2418367           0.5637582
#>                      lineage
#> 1 AY.4 OR AY.34 OR B.1.617.2

# S:E484K mutation prevalence worldwide
getPrevalenceByLocation(mutations = c("S:E484K"))

#> Retrieving data... 

#>           date total_count lineage_count total_count_rolling
#> 1   2020-02-01          50             1            48.71429
#> 2   2020-02-02          48             0            48.14286
#> 3   2020-02-03          41             0            48.71429
#> 4   2020-02-04          34             0            48.42857
#> 5   2020-02-05          45             0            45.42857
#> 6   2020-02-06          33             0            41.28571
#> 7   2020-02-07          31             0            40.28571
#> 8   2020-02-08          43             0            39.28571
#> 9   2020-02-09          48             0            39.28571
#> 10  2020-02-10          30             0            37.71429
#> 11  2020-02-11          26             0            36.57143
#> 12  2020-02-12          15             0            32.28571
#> 13  2020-02-13          24             0            31.00000
#> 14  2020-02-14          25             0            30.14286
#> 15  2020-02-15         113             2            40.14286
#> 16  2020-02-16          39             0            38.85714
#> 17  2020-02-17          60             0            43.14286
#> 18  2020-02-18          26             0            43.14286
#> 19  2020-02-19          34             0            45.85714
#> 20  2020-02-20          48             0            49.28571
#> 21  2020-02-21          58             0            54.00000
#> 22  2020-02-22          58             0            46.14286
#> 23  2020-02-23          48             0            47.42857
#> 24  2020-02-24          77             1            49.85714
#> 25  2020-02-25          81             0            57.71429
#> 26  2020-02-26          93             0            66.14286
#> 27  2020-02-27         128             0            77.57143
#> 28  2020-02-28         199             0            97.71429
#> 29  2020-02-29         151             0           111.00000
#> 30  2020-03-01         223             1           136.00000
#> 31  2020-03-02         343             0           174.00000
#> 32  2020-03-03         416             0           221.85714
#> 33  2020-03-04         476             0           276.57143
#> 34  2020-03-05         483             0           327.28571
#> 35  2020-03-06         483             0           367.85714
#> 36  2020-03-07         447             0           410.14286
#> 37  2020-03-08         467             0           445.00000
#> 38  2020-03-09        1036             0           544.00000
#> 39  2020-03-10         963             1           622.14286
#> 40  2020-03-11        1091             0           710.00000
#> 41  2020-03-12        1286             0           824.71429
#> 42  2020-03-13        1436             0           960.85714
#> 43  2020-03-14        1310             0          1084.14286
#> 44  2020-03-15        3038             2          1451.42857
#> 45  2020-03-16        2269             0          1627.57143
#> 46  2020-03-17        2228             0          1808.28571
#> 47  2020-03-18        2292             0          1979.85714
#> 48  2020-03-19        2156             0          2104.14286
#> 49  2020-03-20        2237             0          2218.57143
#> 50  2020-03-21        1862             1          2297.42857
#> 51  2020-03-22        1981             0          2146.42857
#> 52  2020-03-23        2993             0          2249.85714
#> 53  2020-03-24        2915             0          2348.00000
#> 54  2020-03-25        2866             0          2430.00000
#> 55  2020-03-26        2873             0          2532.42857
#> 56  2020-03-27        2957             1          2635.28571
#> 57  2020-03-28        2329             0          2702.00000
#> 58  2020-03-29        2042             0          2710.71429
#> 59  2020-03-30        3163             0          2735.00000
#> 60  2020-03-31        2916             1          2735.14286
#> 61  2020-04-01        2992             0          2753.14286
#> 62  2020-04-02        2511             0          2701.42857
#> 63  2020-04-03        2591             0          2649.14286
#> 64  2020-04-04        1856             0          2581.57143
#> 65  2020-04-05        1546             0          2510.71429
#> 66  2020-04-06        2691             0          2443.28571
#> 67  2020-04-07        2326             2          2359.00000
#> 68  2020-04-08        2350             0          2267.28571
#> 69  2020-04-09        1974             0          2190.57143
#> 70  2020-04-10        1957             0          2100.00000
#> 71  2020-04-11        1636             0          2068.57143
#> 72  2020-04-12        1294             0          2032.57143
#> 73  2020-04-13        2190             2          1961.00000
#> 74  2020-04-14        2013             0          1916.28571
#> 75  2020-04-15        3446             1          2072.85714
#> 76  2020-04-16        1789             0          2046.42857
#> 77  2020-04-17        1780             0          2021.14286
#> 78  2020-04-18        1262             0          1967.71429
#> 79  2020-04-19         923             0          1914.71429
#> 80  2020-04-20        1978             1          1884.42857
#> 81  2020-04-21        1615             1          1827.57143
#> 82  2020-04-22        1609             0          1565.14286
#> 83  2020-04-23        1501             1          1524.00000
#> 84  2020-04-24        1380             0          1466.85714
#> 85  2020-04-25         957             0          1423.28571
#> 86  2020-04-26         835             0          1410.71429
#> 87  2020-04-27        1569             0          1352.28571
#> 88  2020-04-28        1340             0          1313.00000
#> 89  2020-04-29        1471             0          1293.28571
#> 90  2020-04-30        1481             1          1290.42857
#> 91  2020-05-01        1361             0          1287.71429
#> 92  2020-05-02         897             0          1279.14286
#> 93  2020-05-03         738             0          1265.28571
#> 94  2020-05-04        1393             0          1240.14286
#> 95  2020-05-05        1120             0          1208.71429
#> 96  2020-05-06        1049             0          1148.42857
#> 97  2020-05-07        1060             1          1088.28571
#> 98  2020-05-08         985             0          1034.57143
#> 99  2020-05-09         723             1          1009.71429
#> 100 2020-05-10         520             0           978.57143
#> 101 2020-05-11        1080             0           933.85714
#> 102 2020-05-12         998             0           916.42857
#> 103 2020-05-13        1036             0           914.57143
#> 104 2020-05-14        1156             0           928.28571
#> 105 2020-05-15        2022             2          1076.42857
#> 106 2020-05-16         592             0          1057.71429
#> 107 2020-05-17         543             0          1061.00000
#> 108 2020-05-18        1265             1          1087.42857
#> 109 2020-05-19        1105             1          1102.71429
#> 110 2020-05-20         840             0          1074.71429
#> 111 2020-05-21         794             0          1023.00000
#>     lineage_count_rolling   proportion proportion_ci_lower proportion_ci_upper
#> 1               0.1428571 2.932551e-03        9.970067e-06        0.0497233853
#> 2               0.1428571 2.967359e-03        1.017669e-05        0.0507272634
#> 3               0.1428571 2.932551e-03        9.970067e-06        0.0497233853
#> 4               0.1428571 2.949853e-03        1.017669e-05        0.0507272634
#> 5               0.1428571 3.144654e-03        1.085137e-05        0.0539977097
#> 6               0.1428571 3.460208e-03        1.190358e-05        0.0590757099
#> 7               0.1428571 3.546099e-03        1.219931e-05        0.0604979752
#> 8               0.0000000 0.000000e+00        1.251010e-05        0.0619903843
#> 9               0.0000000 0.000000e+00        1.251010e-05        0.0619903843
#> 10              0.0000000 0.000000e+00        1.283715e-05        0.0635582550
#> 11              0.0000000 0.000000e+00        1.318175e-05        0.0652074564
#> 12              0.0000000 0.000000e+00        1.522530e-05        0.0749277634
#> 13              0.0000000 0.000000e+00        1.571248e-05        0.0772300357
#> 14              0.0000000 0.000000e+00        1.623186e-05        0.0796781738
#> 15              0.2857143 7.117438e-03        1.219931e-05        0.0604979752
#> 16              0.2857143 7.352941e-03        1.251010e-05        0.0619903843
#> 17              0.2857143 6.622517e-03        1.135315e-05        0.0564227244
#> 18              0.2857143 6.622517e-03        1.135315e-05        0.0564227244
#> 19              0.2857143 6.230530e-03        1.061675e-05        0.0528617046
#> 20              0.2857143 5.797101e-03        9.970067e-06        0.0497233853
#> 21              0.2857143 5.291005e-03        9.051191e-06        0.0452462104
#> 22              0.0000000 0.000000e+00        1.061675e-05        0.0528617046
#> 23              0.0000000 0.000000e+00        1.039207e-05        0.0517725002
#> 24              0.1428571 2.865330e-03        9.771663e-06        0.0487584594
#> 25              0.1428571 2.475248e-03        8.429665e-06        0.0422058745
#> 26              0.1428571 2.159827e-03        7.411762e-06        0.0372056103
#> 27              0.1428571 1.841621e-03        6.275150e-06        0.0315913009
#> 28              0.1428571 1.461988e-03        4.997779e-06        0.0252425708
#> 29              0.1428571 1.287001e-03        4.413773e-06        0.0223261109
#> 30              0.2857143 2.100840e-03        3.603911e-06        0.0182672895
#> 31              0.1428571 8.210181e-04        2.817982e-06        0.0143122939
#> 32              0.1428571 6.439150e-04        2.209376e-06        0.0112386808
#> 33              0.1428571 5.165289e-04        1.771087e-06        0.0090192797
#> 34              0.1428571 4.364906e-04        1.500486e-06        0.0076465242
#> 35              0.1428571 3.883495e-04        1.333426e-06        0.0067980819
#> 36              0.1428571 3.483107e-04        1.196914e-06        0.0061042468
#> 37              0.0000000 0.000000e+00        1.102828e-06        0.0056257599
#> 38              0.0000000 0.000000e+00        9.022219e-07        0.0046047854
#> 39              0.1428571 2.296211e-04        7.891271e-07        0.0040287331
#> 40              0.1428571 2.012072e-04        6.913543e-07        0.0035304554
#> 41              0.1428571 1.732202e-04        5.950129e-07        0.0030392289
#> 42              0.1428571 1.486768e-04        5.108290e-07        0.0026097931
#> 43              0.1428571 1.317697e-04        4.528793e-07        0.0023140746
#> 44              0.4285714 2.952756e-04        3.383528e-07        0.0017293852
#> 45              0.4285714 2.633196e-04        3.015719e-07        0.0015415362
#> 46              0.2857143 1.580028e-04        2.715523e-07        0.0013881925
#> 47              0.2857143 1.443106e-04        2.479659e-07        0.0012676937
#> 48              0.2857143 1.357865e-04        2.333537e-07        0.0011930352
#> 49              0.2857143 1.287830e-04        2.212615e-07        0.0011312479
#> 50              0.4285714 1.865440e-04        2.137488e-07        0.0010928589
#> 51              0.1428571 6.655574e-05        2.287872e-07        0.0011697023
#> 52              0.1428571 6.349609e-05        2.182133e-07        0.0011156723
#> 53              0.1428571 6.084205e-05        2.091066e-07        0.0010691366
#> 54              0.1428571 5.878895e-05        2.020510e-07        0.0010330811
#> 55              0.1428571 5.641112e-05        1.939123e-07        0.0009914888
#> 56              0.2857143 1.084187e-04        1.863332e-07        0.0009527545
#> 57              0.1428571 5.287089e-05        1.817132e-07        0.0009291427
#> 58              0.1428571 5.270092e-05        1.811100e-07        0.0009260599
#> 59              0.1428571 5.223296e-05        1.795209e-07        0.0009179380
#> 60              0.2857143 1.044605e-04        1.795209e-07        0.0009179380
#> 61              0.2857143 1.037775e-04        1.783472e-07        0.0009119395
#> 62              0.2857143 1.057641e-04        1.817805e-07        0.0009294865
#> 63              0.1428571 5.392580e-05        1.853485e-07        0.0009477220
#> 64              0.1428571 5.533728e-05        1.901576e-07        0.0009723000
#> 65              0.1428571 5.689900e-05        1.955339e-07        0.0009997759
#> 66              0.1428571 5.846927e-05        2.009760e-07        0.0010275871
#> 67              0.2857143 1.211167e-04        2.081316e-07        0.0010641544
#> 68              0.2857143 1.260160e-04        2.165771e-07        0.0011073115
#> 69              0.2857143 1.304291e-04        2.240888e-07        0.0011456948
#> 70              0.2857143 1.360544e-04        2.337981e-07        0.0011953060
#> 71              0.2857143 1.381215e-04        2.373007e-07        0.0012132024
#> 72              0.2857143 1.405679e-04        2.415023e-07        0.0012346696
#> 73              0.5714286 2.913965e-04        5.502725e-05        0.0023810455
#> 74              0.2857143 1.490980e-04        2.562476e-07        0.0013100051
#> 75              0.4285714 2.067540e-04        2.368429e-07        0.0012108631
#> 76              0.4285714 2.094241e-04        2.399680e-07        0.0012268305
#> 77              0.4285714 2.120441e-04        2.429361e-07        0.0012419952
#> 78              0.4285714 2.178017e-04        2.494777e-07        0.0012754177
#> 79              0.4285714 2.238305e-04        2.563814e-07        0.0013106887
#> 80              0.2857143 1.516185e-04        2.605994e-07        0.0013322380
#> 81              0.4285714 2.345032e-04        2.685817e-07        0.0013730169
#> 82              0.2857143 1.825484e-04        3.137099e-07        0.0016035320
#> 83              0.4285714 2.812148e-04        3.221483e-07        0.0016466290
#> 84              0.4285714 2.921698e-04        3.346631e-07        0.0017105428
#> 85              0.4285714 3.011141e-04        3.450093e-07        0.0017633779
#> 86              0.4285714 3.037975e-04        3.479429e-07        0.0017783587
#> 87              0.2857143 2.112825e-04        3.631240e-07        0.0018558782
#> 88              0.1428571 1.088021e-04        3.739078e-07        0.0019109401
#> 89              0.1428571 1.104606e-04        3.796903e-07        0.0019404640
#> 90              0.1428571 1.107052e-04        3.805731e-07        0.0019449714
#> 91              0.1428571 1.109385e-04        3.811639e-07        0.0019479880
#> 92              0.1428571 1.116819e-04        3.838456e-07        0.0019616794
#> 93              0.1428571 1.129050e-04        3.880928e-07        0.0019833638
#> 94              0.1428571 1.151941e-04        3.959157e-07        0.0020233025
#> 95              0.1428571 1.181893e-04        4.060653e-07        0.0020751174
#> 96              0.1428571 1.243936e-04        4.276372e-07        0.0021852361
#> 97              0.1428571 1.312680e-04        4.512147e-07        0.0023055787
#> 98              0.1428571 1.380834e-04        4.743148e-07        0.0024234705
#> 99              0.2857143 2.829655e-04        4.860524e-07        0.0024833680
#> 100             0.2857143 2.919708e-04        5.014393e-07        0.0025618829
#> 101             0.2857143 3.059507e-04        5.255921e-07        0.0026851152
#> 102             0.2857143 3.117693e-04        5.359175e-07        0.0027377928
#> 103             0.2857143 3.124024e-04        5.365030e-07        0.0027407800
#> 104             0.1428571 1.538935e-04        5.289894e-07        0.0027024477
#> 105             0.4285714 3.981420e-04        4.562457e-07        0.0023312555
#> 106             0.2857143 2.701243e-04        4.640061e-07        0.0023708613
#> 107             0.2857143 2.692877e-04        4.626944e-07        0.0023641672
#> 108             0.4285714 3.941146e-04        4.516297e-07        0.0023076968
#> 109             0.5714286 5.182018e-04        9.783937e-05        0.0042297077
#> 110             0.5714286 5.317028e-04        1.003882e-04        0.0043396630
#> 111             0.5714286 5.585812e-04        1.054920e-04        0.0045598021
#>     lineage
#> 1   S:E484K
#> 2   S:E484K
#> 3   S:E484K
#> 4   S:E484K
#> 5   S:E484K
#> 6   S:E484K
#> 7   S:E484K
#> 8   S:E484K
#> 9   S:E484K
#> 10  S:E484K
#> 11  S:E484K
#> 12  S:E484K
#> 13  S:E484K
#> 14  S:E484K
#> 15  S:E484K
#> 16  S:E484K
#> 17  S:E484K
#> 18  S:E484K
#> 19  S:E484K
#> 20  S:E484K
#> 21  S:E484K
#> 22  S:E484K
#> 23  S:E484K
#> 24  S:E484K
#> 25  S:E484K
#> 26  S:E484K
#> 27  S:E484K
#> 28  S:E484K
#> 29  S:E484K
#> 30  S:E484K
#> 31  S:E484K
#> 32  S:E484K
#> 33  S:E484K
#> 34  S:E484K
#> 35  S:E484K
#> 36  S:E484K
#> 37  S:E484K
#> 38  S:E484K
#> 39  S:E484K
#> 40  S:E484K
#> 41  S:E484K
#> 42  S:E484K
#> 43  S:E484K
#> 44  S:E484K
#> 45  S:E484K
#> 46  S:E484K
#> 47  S:E484K
#> 48  S:E484K
#> 49  S:E484K
#> 50  S:E484K
#> 51  S:E484K
#> 52  S:E484K
#> 53  S:E484K
#> 54  S:E484K
#> 55  S:E484K
#> 56  S:E484K
#> 57  S:E484K
#> 58  S:E484K
#> 59  S:E484K
#> 60  S:E484K
#> 61  S:E484K
#> 62  S:E484K
#> 63  S:E484K
#> 64  S:E484K
#> 65  S:E484K
#> 66  S:E484K
#> 67  S:E484K
#> 68  S:E484K
#> 69  S:E484K
#> 70  S:E484K
#> 71  S:E484K
#> 72  S:E484K
#> 73  S:E484K
#> 74  S:E484K
#> 75  S:E484K
#> 76  S:E484K
#> 77  S:E484K
#> 78  S:E484K
#> 79  S:E484K
#> 80  S:E484K
#> 81  S:E484K
#> 82  S:E484K
#> 83  S:E484K
#> 84  S:E484K
#> 85  S:E484K
#> 86  S:E484K
#> 87  S:E484K
#> 88  S:E484K
#> 89  S:E484K
#> 90  S:E484K
#> 91  S:E484K
#> 92  S:E484K
#> 93  S:E484K
#> 94  S:E484K
#> 95  S:E484K
#> 96  S:E484K
#> 97  S:E484K
#> 98  S:E484K
#> 99  S:E484K
#> 100 S:E484K
#> 101 S:E484K
#> 102 S:E484K
#> 103 S:E484K
#> 104 S:E484K
#> 105 S:E484K
#> 106 S:E484K
#> 107 S:E484K
#> 108 S:E484K
#> 109 S:E484K
#> 110 S:E484K
#> 111 S:E484K
#>  [ reached 'max' / getOption("max.print") -- omitted 473 rows ]

# B.1.1.7 + S:E484K mutation worldwide
getPrevalenceByLocation(pangolin_lineage = "B.1.1.7", mutations = c("S:E484K"))

#> Retrieving data... 

#>           date total_count lineage_count total_count_rolling
#> 1   2020-12-17        4559             1            4817.286
#> 2   2020-12-18        4521             1            4887.429
#> 3   2020-12-19        2920             0            4957.429
#> 4   2020-12-20        2858             0            5040.429
#> 5   2020-12-21        8632             0            4856.857
#> 6   2020-12-22        4020             0            4533.000
#> 7   2020-12-23        3656             0            4452.286
#> 8   2020-12-24        3347             0            4279.143
#> 9   2020-12-25        1728             0            3880.143
#> 10  2020-12-26        4335             0            4082.286
#> 11  2020-12-27        3996             0            4244.857
#> 12  2020-12-28        9959             0            4434.429
#> 13  2020-12-29        6581             1            4800.286
#> 14  2020-12-30        4863             0            4972.714
#> 15  2020-12-31        4142             0            5086.286
#> 16  2021-01-01        2584             0            5208.571
#> 17  2021-01-02        4986             0            5301.571
#> 18  2021-01-03        4795             1            5415.714
#> 19  2021-01-04       13679             2            5947.143
#> 20  2021-01-05        7808             0            6122.429
#> 21  2021-01-06        5750             1            6249.143
#> 22  2021-01-07        6072             0            6524.857
#> 23  2021-01-08        6118             3            7029.714
#> 24  2021-01-09        5464             0            7098.000
#> 25  2021-01-10        4187             0            7011.143
#> 26  2021-01-11       12847             1            6892.286
#> 27  2021-01-12        8727             0            7023.571
#> 28  2021-01-13        7384             5            7257.000
#> 29  2021-01-14        7945             0            7524.571
#> 30  2021-01-15       13201             2            8536.429
#> 31  2021-01-16        6028             3            8617.000
#> 32  2021-01-17        5030             1            8737.429
#> 33  2021-01-18       13491             1            8829.429
#> 34  2021-01-19        9594             2            8953.286
#> 35  2021-01-20        8063             0            9050.286
#> 36  2021-01-21        7215             3            8946.000
#> 37  2021-01-22        7328             1            8107.000
#> 38  2021-01-23        4517             0            7891.143
#> 39  2021-01-24        4110             2            7759.714
#> 40  2021-01-25       13583             5            7772.857
#> 41  2021-01-26        9632             1            7778.286
#> 42  2021-01-27        9893             1            8039.714
#> 43  2021-01-28        9626             0            8384.143
#> 44  2021-01-29        9290             1            8664.429
#> 45  2021-01-30        6345             2            8925.571
#> 46  2021-01-31        5025             1            9056.286
#> 47  2021-02-01       12299             0            8872.857
#> 48  2021-02-02        9609             2            8869.571
#> 49  2021-02-03        8783             1            8711.000
#> 50  2021-02-04        8116             1            8495.286
#> 51  2021-02-05        7346             0            8217.571
#> 52  2021-02-06        5112             2            8041.429
#> 53  2021-02-07        4788             2            8007.571
#> 54  2021-02-08       13615             0            8195.571
#> 55  2021-02-09       10569             3            8332.714
#> 56  2021-02-10       10296             1            8548.857
#> 57  2021-02-11        9142             1            8695.429
#> 58  2021-02-12        8852             0            8910.571
#> 59  2021-02-13        5444             0            8958.000
#> 60  2021-02-14        5260             2            9025.429
#> 61  2021-02-15       25980             3           10791.857
#> 62  2021-02-16       11089             2           10866.143
#> 63  2021-02-17        9638             4           10772.143
#> 64  2021-02-18        9161             3           10774.857
#> 65  2021-02-19        8921             1           10784.714
#> 66  2021-02-20        6057             2           10872.286
#> 67  2021-02-21        5847             6           10956.143
#> 68  2021-02-22       15551             2            9466.286
#> 69  2021-02-23       10530             4            9386.429
#> 70  2021-02-24       10987             2            9579.143
#> 71  2021-02-25       11216             4            9872.714
#> 72  2021-02-26        9678             8            9980.857
#> 73  2021-02-27        6059             0            9981.143
#> 74  2021-02-28        5892             0            9987.571
#> 75  2021-03-01       15858             5           10031.429
#> 76  2021-03-02       12836             4           10360.857
#> 77  2021-03-03       12382             1           10560.143
#> 78  2021-03-04       10783             4           10498.286
#> 79  2021-03-05       10681             1           10641.571
#> 80  2021-03-06        6774             2           10743.714
#> 81  2021-03-07        5947             3           10751.571
#> 82  2021-03-08       18169             7           11081.714
#> 83  2021-03-09       13494            14           11175.714
#> 84  2021-03-10       13876            19           11389.143
#> 85  2021-03-11       12865             6           11686.571
#> 86  2021-03-12       12708            14           11976.143
#> 87  2021-03-13        8583            16           12234.571
#> 88  2021-03-14        7603             9           12471.143
#> 89  2021-03-15       30003            67           14161.714
#> 90  2021-03-16       16232            25           14552.857
#> 91  2021-03-17       13396            28           14484.286
#> 92  2021-03-18       14224            28           14678.429
#> 93  2021-03-19       11920            22           14565.857
#> 94  2021-03-20        9040             4           14631.143
#> 95  2021-03-21        8783            19           14799.714
#> 96  2021-03-22       21485            20           13582.857
#> 97  2021-03-23       14998            16           13406.571
#> 98  2021-03-24       14848            28           13614.000
#> 99  2021-03-25       14909            25           13711.857
#> 100 2021-03-26       13754            36           13973.857
#> 101 2021-03-27        9369            28           14020.857
#> 102 2021-03-28        7788             5           13878.714
#> 103 2021-03-29       21238            27           13843.429
#> 104 2021-03-30       16710            17           14088.000
#> 105 2021-03-31       14901            19           14095.571
#> 106 2021-04-01       14639            26           14057.000
#> 107 2021-04-02       11614            15           13751.286
#> 108 2021-04-03        9023             7           13701.857
#> 109 2021-04-04        7803             3           13704.000
#> 110 2021-04-05       17327            15           13145.286
#> 111 2021-04-06       15582            17           12984.143
#>     lineage_count_rolling   proportion proportion_ci_lower proportion_ci_upper
#> 1               0.1428571 2.965511e-05        1.019325e-07        0.0005213116
#> 2               0.2857143 5.845902e-05        1.004726e-07        0.0005138468
#> 3               0.2857143 5.763357e-05        9.905382e-08        0.0005065927
#> 4               0.2857143 5.668452e-05        9.742265e-08        0.0004982525
#> 5               0.2857143 5.882699e-05        1.010931e-07        0.0005170196
#> 6               0.2857143 6.302984e-05        1.083184e-07        0.0005539618
#> 7               0.2857143 6.417250e-05        1.102891e-07        0.0005640372
#> 8               0.1428571 3.338452e-05        1.147478e-07        0.0005868332
#> 9               0.0000000 0.000000e+00        1.265471e-07        0.0006471568
#> 10              0.0000000 0.000000e+00        1.202853e-07        0.0006151437
#> 11              0.0000000 0.000000e+00        1.156668e-07        0.0005915317
#> 12              0.0000000 0.000000e+00        1.107368e-07        0.0005663261
#> 13              0.1428571 2.976013e-05        1.022935e-07        0.0005231573
#> 14              0.1428571 2.872820e-05        9.873514e-08        0.0005049633
#> 15              0.1428571 2.808673e-05        9.654156e-08        0.0004937474
#> 16              0.1428571 2.742732e-05        9.426204e-08        0.0004820920
#> 17              0.1428571 2.694619e-05        9.260871e-08        0.0004736382
#> 18              0.2857143 5.275653e-05        9.065950e-08        0.0004636715
#> 19              0.5714286 9.608455e-05        1.814380e-05        0.0007857006
#> 20              0.4285714 7.000023e-05        8.020491e-08        0.0004102132
#> 21              0.5714286 9.144111e-05        1.726693e-05        0.0007477422
#> 22              0.5714286 8.757718e-05        1.653653e-05        0.0007161237
#> 23              1.0000000 1.422533e-04        1.534860e-05        0.0006646962
#> 24              1.0000000 1.408848e-04        1.520155e-05        0.0006583302
#> 25              0.8571429 1.222544e-04        1.539020e-05        0.0006664970
#> 26              0.7142857 1.036355e-04        1.565594e-05        0.0006780015
#> 27              0.7142857 1.016984e-04        1.536171e-05        0.0006652638
#> 28              1.2857143 1.771688e-04        1.486848e-05        0.0006439104
#> 29              1.2857143 1.708688e-04        1.433893e-05        0.0006209841
#> 30              1.1428571 1.338800e-04        1.264060e-05        0.0005474530
#> 31              1.5714286 1.823638e-04        4.823395e-05        0.0007443914
#> 32              1.7142857 1.962003e-04        4.757143e-05        0.0007341703
#> 33              1.7142857 1.941559e-04        4.707570e-05        0.0007265222
#> 34              2.0000000 2.233817e-04        4.642365e-05        0.0007164626
#> 35              1.2857143 1.420634e-04        1.192265e-05        0.0005163673
#> 36              1.7142857 1.916259e-04        4.645998e-05        0.0007170231
#> 37              1.5714286 1.938360e-04        5.126848e-05        0.0007912059
#> 38              1.1428571 1.448278e-04        1.367385e-05        0.0005921893
#> 39              1.2857143 1.656909e-04        1.390469e-05        0.0006021836
#> 40              1.8571429 2.389267e-04        5.347161e-05        0.0008251926
#> 41              1.7142857 2.203938e-04        5.343723e-05        0.0008246623
#> 42              1.8571429 2.309961e-04        5.169575e-05        0.0007977972
#> 43              1.4285714 1.703897e-04        1.286977e-05        0.0005573758
#> 44              1.4285714 1.648777e-04        1.245384e-05        0.0005393670
#> 45              1.7142857 1.920645e-04        4.656409e-05        0.0007186292
#> 46              1.5714286 1.735180e-04        4.589562e-05        0.0007083160
#> 47              0.8571429 9.660280e-05        1.216049e-05        0.0005266654
#> 48              1.0000000 1.127450e-04        1.216460e-05        0.0005268435
#> 49              1.0000000 1.147974e-04        1.238665e-05        0.0005364576
#> 50              1.1428571 1.345284e-04        1.270160e-05        0.0005500946
#> 51              1.0000000 1.216905e-04        1.312974e-05        0.0005686316
#> 52              1.0000000 1.243560e-04        1.341876e-05        0.0005811452
#> 53              1.1428571 1.427221e-04        1.347406e-05        0.0005835394
#> 54              1.1428571 1.394481e-04        1.316499e-05        0.0005701576
#> 55              1.2857143 1.542972e-04        1.294854e-05        0.0005607862
#> 56              1.2857143 1.503960e-04        1.262137e-05        0.0005466208
#> 57              1.2857143 1.478609e-04        1.240944e-05        0.0005374445
#> 58              1.2857143 1.442909e-04        1.210863e-05        0.0005244201
#> 59              1.0000000 1.116321e-04        1.204510e-05        0.0005216692
#> 60              1.0000000 1.107981e-04        1.195568e-05        0.0005177973
#> 61              1.4285714 1.323749e-04        9.998110e-06        0.0004330335
#> 62              1.2857143 1.183230e-04        9.930019e-06        0.0004300850
#> 63              1.7142857 1.591406e-04        3.858399e-05        0.0005955057
#> 64              2.0000000 1.856173e-04        3.857325e-05        0.0005953399
#> 65              2.1428571 1.986939e-04        3.853748e-05        0.0005947880
#> 66              2.4285714 2.233727e-04        3.822908e-05        0.0005900295
#> 67              3.0000000 2.738190e-04        7.712654e-05        0.0007305924
#> 68              2.8571429 3.018230e-04        8.926774e-05        0.0008455582
#> 69              3.1428571 3.348299e-04        9.002867e-05        0.0008527630
#> 70              2.8571429 2.982671e-04        8.821459e-05        0.0008355863
#> 71              3.0000000 3.038678e-04        8.558750e-05        0.0008107110
#> 72              4.0000000 4.007672e-04        1.352911e-04        0.0009526621
#> 73              3.7142857 3.721303e-04        1.352911e-04        0.0009526621
#> 74              2.8571429 2.860698e-04        8.460198e-05        0.0008013792
#> 75              3.2857143 3.275420e-04        8.423929e-05        0.0007979449
#> 76              3.2857143 3.171277e-04        8.155604e-05        0.0007725369
#> 77              3.1428571 2.976150e-04        8.001902e-05        0.0007579825
#> 78              3.1428571 2.993686e-04        8.049164e-05        0.0007624579
#> 79              2.1428571 2.013666e-04        3.905535e-05        0.0006027785
#> 80              2.4285714 2.260458e-04        3.868455e-05        0.0005970573
#> 81              2.8571429 2.657419e-04        7.859000e-05        0.0007444505
#> 82              3.1428571 2.836075e-04        7.624957e-05        0.0007222878
#> 83              4.5714286 4.090502e-04        1.707315e-04        0.0009803917
#> 84              7.1428571 6.271637e-04        2.749611e-04        0.0012064119
#> 85              7.4285714 6.356502e-04        2.679490e-04        0.0011756599
#> 86              9.2857143 7.753510e-04        3.719114e-04        0.0013711363
#> 87             11.2857143 9.224446e-04        4.777596e-04        0.0015554697
#> 88             12.1428571 9.736764e-04        5.261134e-04        0.0016290631
#> 89             20.7142857 1.462696e-03        9.459151e-04        0.0022230582
#> 90             22.2857143 1.531364e-03        9.748261e-04        0.0022464432
#> 91             23.5714286 1.627379e-03        1.089593e-03        0.0024231657
#> 92             26.7142857 1.819969e-03        1.240237e-03        0.0026348388
#> 93             27.8571429 1.912496e-03        1.305706e-03        0.0027364515
#> 94             26.1428571 1.786795e-03        1.188776e-03        0.0025620634
#> 95             27.5714286 1.862970e-03        1.285056e-03        0.0026932041
#> 96             20.8571429 1.535549e-03        9.862470e-04        0.0023177817
#> 97             19.5714286 1.459838e-03        9.405909e-04        0.0022576349
#> 98             19.5714286 1.437596e-03        9.262856e-04        0.0022233213
#> 99             19.1428571 1.396081e-03        8.627510e-04        0.0021184976
#> 100            21.1428571 1.513029e-03        9.586443e-04        0.0022529545
#> 101            24.5714286 1.752491e-03        1.182910e-03        0.0025884534
#> 102            22.5714286 1.626334e-03        1.079484e-03        0.0024422785
#> 103            23.5714286 1.702716e-03        1.140061e-03        0.0025353232
#> 104            23.7142857 1.683297e-03        1.120229e-03        0.0024912504
#> 105            22.4285714 1.591179e-03        1.006439e-03        0.0023192451
#> 106            22.5714286 1.605707e-03        1.065811e-03        0.0024113651
#> 107            19.5714286 1.423244e-03        9.170548e-04        0.0022011792
#> 108            16.5714286 1.209429e-03        7.507679e-04        0.0019407289
#> 109            16.2857143 1.188391e-03        6.950823e-04        0.0018500739
#> 110            14.5714286 1.108491e-03        6.672713e-04        0.0018339387
#> 111            14.5714286 1.122248e-03        6.755472e-04        0.0018566709
#>                     lineage
#> 1   (B.1.1.7) AND (S:E484K)
#> 2   (B.1.1.7) AND (S:E484K)
#> 3   (B.1.1.7) AND (S:E484K)
#> 4   (B.1.1.7) AND (S:E484K)
#> 5   (B.1.1.7) AND (S:E484K)
#> 6   (B.1.1.7) AND (S:E484K)
#> 7   (B.1.1.7) AND (S:E484K)
#> 8   (B.1.1.7) AND (S:E484K)
#> 9   (B.1.1.7) AND (S:E484K)
#> 10  (B.1.1.7) AND (S:E484K)
#> 11  (B.1.1.7) AND (S:E484K)
#> 12  (B.1.1.7) AND (S:E484K)
#> 13  (B.1.1.7) AND (S:E484K)
#> 14  (B.1.1.7) AND (S:E484K)
#> 15  (B.1.1.7) AND (S:E484K)
#> 16  (B.1.1.7) AND (S:E484K)
#> 17  (B.1.1.7) AND (S:E484K)
#> 18  (B.1.1.7) AND (S:E484K)
#> 19  (B.1.1.7) AND (S:E484K)
#> 20  (B.1.1.7) AND (S:E484K)
#> 21  (B.1.1.7) AND (S:E484K)
#> 22  (B.1.1.7) AND (S:E484K)
#> 23  (B.1.1.7) AND (S:E484K)
#> 24  (B.1.1.7) AND (S:E484K)
#> 25  (B.1.1.7) AND (S:E484K)
#> 26  (B.1.1.7) AND (S:E484K)
#> 27  (B.1.1.7) AND (S:E484K)
#> 28  (B.1.1.7) AND (S:E484K)
#> 29  (B.1.1.7) AND (S:E484K)
#> 30  (B.1.1.7) AND (S:E484K)
#> 31  (B.1.1.7) AND (S:E484K)
#> 32  (B.1.1.7) AND (S:E484K)
#> 33  (B.1.1.7) AND (S:E484K)
#> 34  (B.1.1.7) AND (S:E484K)
#> 35  (B.1.1.7) AND (S:E484K)
#> 36  (B.1.1.7) AND (S:E484K)
#> 37  (B.1.1.7) AND (S:E484K)
#> 38  (B.1.1.7) AND (S:E484K)
#> 39  (B.1.1.7) AND (S:E484K)
#> 40  (B.1.1.7) AND (S:E484K)
#> 41  (B.1.1.7) AND (S:E484K)
#> 42  (B.1.1.7) AND (S:E484K)
#> 43  (B.1.1.7) AND (S:E484K)
#> 44  (B.1.1.7) AND (S:E484K)
#> 45  (B.1.1.7) AND (S:E484K)
#> 46  (B.1.1.7) AND (S:E484K)
#> 47  (B.1.1.7) AND (S:E484K)
#> 48  (B.1.1.7) AND (S:E484K)
#> 49  (B.1.1.7) AND (S:E484K)
#> 50  (B.1.1.7) AND (S:E484K)
#> 51  (B.1.1.7) AND (S:E484K)
#> 52  (B.1.1.7) AND (S:E484K)
#> 53  (B.1.1.7) AND (S:E484K)
#> 54  (B.1.1.7) AND (S:E484K)
#> 55  (B.1.1.7) AND (S:E484K)
#> 56  (B.1.1.7) AND (S:E484K)
#> 57  (B.1.1.7) AND (S:E484K)
#> 58  (B.1.1.7) AND (S:E484K)
#> 59  (B.1.1.7) AND (S:E484K)
#> 60  (B.1.1.7) AND (S:E484K)
#> 61  (B.1.1.7) AND (S:E484K)
#> 62  (B.1.1.7) AND (S:E484K)
#> 63  (B.1.1.7) AND (S:E484K)
#> 64  (B.1.1.7) AND (S:E484K)
#> 65  (B.1.1.7) AND (S:E484K)
#> 66  (B.1.1.7) AND (S:E484K)
#> 67  (B.1.1.7) AND (S:E484K)
#> 68  (B.1.1.7) AND (S:E484K)
#> 69  (B.1.1.7) AND (S:E484K)
#> 70  (B.1.1.7) AND (S:E484K)
#> 71  (B.1.1.7) AND (S:E484K)
#> 72  (B.1.1.7) AND (S:E484K)
#> 73  (B.1.1.7) AND (S:E484K)
#> 74  (B.1.1.7) AND (S:E484K)
#> 75  (B.1.1.7) AND (S:E484K)
#> 76  (B.1.1.7) AND (S:E484K)
#> 77  (B.1.1.7) AND (S:E484K)
#> 78  (B.1.1.7) AND (S:E484K)
#> 79  (B.1.1.7) AND (S:E484K)
#> 80  (B.1.1.7) AND (S:E484K)
#> 81  (B.1.1.7) AND (S:E484K)
#> 82  (B.1.1.7) AND (S:E484K)
#> 83  (B.1.1.7) AND (S:E484K)
#> 84  (B.1.1.7) AND (S:E484K)
#> 85  (B.1.1.7) AND (S:E484K)
#> 86  (B.1.1.7) AND (S:E484K)
#> 87  (B.1.1.7) AND (S:E484K)
#> 88  (B.1.1.7) AND (S:E484K)
#> 89  (B.1.1.7) AND (S:E484K)
#> 90  (B.1.1.7) AND (S:E484K)
#> 91  (B.1.1.7) AND (S:E484K)
#> 92  (B.1.1.7) AND (S:E484K)
#> 93  (B.1.1.7) AND (S:E484K)
#> 94  (B.1.1.7) AND (S:E484K)
#> 95  (B.1.1.7) AND (S:E484K)
#> 96  (B.1.1.7) AND (S:E484K)
#> 97  (B.1.1.7) AND (S:E484K)
#> 98  (B.1.1.7) AND (S:E484K)
#> 99  (B.1.1.7) AND (S:E484K)
#> 100 (B.1.1.7) AND (S:E484K)
#> 101 (B.1.1.7) AND (S:E484K)
#> 102 (B.1.1.7) AND (S:E484K)
#> 103 (B.1.1.7) AND (S:E484K)
#> 104 (B.1.1.7) AND (S:E484K)
#> 105 (B.1.1.7) AND (S:E484K)
#> 106 (B.1.1.7) AND (S:E484K)
#> 107 (B.1.1.7) AND (S:E484K)
#> 108 (B.1.1.7) AND (S:E484K)
#> 109 (B.1.1.7) AND (S:E484K)
#> 110 (B.1.1.7) AND (S:E484K)
#> 111 (B.1.1.7) AND (S:E484K)
#>  [ reached 'max' / getOption("max.print") -- omitted 153 rows ]