Retrieve sequence count data

Retrieves number of sequences by location, by date of sample collection. See getSubmissionDateByLocation to view similar information by date of sequence submission.

getSeqCounts(location = NULL, cumulative = NULL, subadmin = NULL)

Arguments

location

(optional) a location name (if not specified, global total counts returned)

cumulative

(optional) Boolean (T/F), T returns cumulative number of sequences to date

subadmin

(optional) Boolean (T/F), subadmin=T and cumulative=T returns cumulative number of sequences for next administrative level

Value

dataframe

Examples

# Retrieves the number of samples sequenced in the U.S. each day, by date of sample collection
getSeqCounts(location="United States")
#> Retrieving data... 
#>           date total_count
#> 1   2020-01-26           3
#> 2   2020-01-27           1
#> 3   2020-01-28           2
#> 4   2020-01-29           4
#> 5   2020-01-31           1
#> 6   2020-02-02           1
#> 7   2020-02-06           1
#> 8   2020-02-08           1
#> 9   2020-02-10           1
#> 10  2020-02-11           1
#> 11  2020-02-17           9
#> 12  2020-02-18           6
#> 13  2020-02-20           2
#> 14  2020-02-21           8
#> 15  2020-02-22           3
#> 16  2020-02-23           1
#> 17  2020-02-24           7
#> 18  2020-02-25           4
#> 19  2020-02-26           7
#> 20  2020-02-27           9
#> 21  2020-02-28          27
#> 22  2020-02-29          37
#> 23  2020-03-01          26
#> 24  2020-03-02          49
#> 25  2020-03-03          43
#> 26  2020-03-04          68
#> 27  2020-03-05         105
#> 28  2020-03-06          84
#> 29  2020-03-07          94
#> 30  2020-03-08          99
#> 31  2020-03-09         192
#> 32  2020-03-10         190
#> 33  2020-03-11         240
#> 34  2020-03-12         381
#> 35  2020-03-13         550
#> 36  2020-03-14         531
#> 37  2020-03-15         999
#> 38  2020-03-16         933
#> 39  2020-03-17         771
#> 40  2020-03-18         830
#> 41  2020-03-19         685
#> 42  2020-03-20         594
#> 43  2020-03-21         459
#> 44  2020-03-22         386
#> 45  2020-03-23         629
#> 46  2020-03-24         721
#> 47  2020-03-25         733
#> 48  2020-03-26         700
#> 49  2020-03-27         675
#> 50  2020-03-28         473
#> 51  2020-03-29         349
#> 52  2020-03-30         747
#> 53  2020-03-31         674
#> 54  2020-04-01         835
#> 55  2020-04-02         707
#> 56  2020-04-03         736
#> 57  2020-04-04         450
#> 58  2020-04-05         375
#> 59  2020-04-06         646
#> 60  2020-04-07         535
#> 61  2020-04-08         607
#> 62  2020-04-09         453
#> 63  2020-04-10         510
#> 64  2020-04-11         403
#> 65  2020-04-12         255
#> 66  2020-04-13         701
#> 67  2020-04-14         472
#> 68  2020-04-15         837
#> 69  2020-04-16         414
#> 70  2020-04-17         461
#> 71  2020-04-18         329
#> 72  2020-04-19         210
#> 73  2020-04-20         533
#> 74  2020-04-21         550
#> 75  2020-04-22         433
#> 76  2020-04-23         436
#> 77  2020-04-24         453
#> 78  2020-04-25         250
#> 79  2020-04-26         182
#> 80  2020-04-27         538
#> 81  2020-04-28         521
#> 82  2020-04-29         610
#> 83  2020-04-30         654
#> 84  2020-05-01         557
#> 85  2020-05-02         305
#> 86  2020-05-03         190
#> 87  2020-05-04         582
#> 88  2020-05-05         537
#> 89  2020-05-06         420
#> 90  2020-05-07         491
#> 91  2020-05-08         431
#> 92  2020-05-09         279
#> 93  2020-05-10         174
#> 94  2020-05-11         465
#> 95  2020-05-12         452
#> 96  2020-05-13         497
#> 97  2020-05-14         604
#> 98  2020-05-15         484
#> 99  2020-05-16         177
#> 100 2020-05-17         133
#> 101 2020-05-18         350
#> 102 2020-05-19         398
#> 103 2020-05-20         372
#> 104 2020-05-21         434
#> 105 2020-05-22         431
#> 106 2020-05-23         160
#> 107 2020-05-24         145
#> 108 2020-05-25         158
#> 109 2020-05-26         469
#> 110 2020-05-27         446
#> 111 2020-05-28         356
#> 112 2020-05-29         451
#> 113 2020-05-30         208
#> 114 2020-05-31         179
#> 115 2020-06-01         425
#> 116 2020-06-02         361
#> 117 2020-06-03         458
#> 118 2020-06-04         442
#> 119 2020-06-05         425
#> 120 2020-06-06         185
#> 121 2020-06-07         138
#> 122 2020-06-08         462
#> 123 2020-06-09         457
#> 124 2020-06-10         420
#> 125 2020-06-11         421
#> 126 2020-06-12         352
#> 127 2020-06-13         376
#> 128 2020-06-14         238
#> 129 2020-06-15         702
#> 130 2020-06-16         664
#> 131 2020-06-17         476
#> 132 2020-06-18         722
#> 133 2020-06-19         648
#> 134 2020-06-20         377
#> 135 2020-06-21         430
#> 136 2020-06-22         765
#> 137 2020-06-23         641
#> 138 2020-06-24         759
#> 139 2020-06-25         698
#> 140 2020-06-26         686
#> 141 2020-06-27         343
#> 142 2020-06-28         359
#> 143 2020-06-29         831
#> 144 2020-06-30         724
#> 145 2020-07-01         845
#> 146 2020-07-02         865
#> 147 2020-07-03         553
#> 148 2020-07-04         335
#> 149 2020-07-05         347
#> 150 2020-07-06         814
#> 151 2020-07-07        1001
#> 152 2020-07-08        1146
#> 153 2020-07-09        1118
#> 154 2020-07-10         960
#> 155 2020-07-11         448
#> 156 2020-07-12         390
#> 157 2020-07-13         936
#> 158 2020-07-14         934
#> 159 2020-07-15         974
#> 160 2020-07-16         726
#> 161 2020-07-17         846
#> 162 2020-07-18         459
#> 163 2020-07-19         410
#> 164 2020-07-20         862
#> 165 2020-07-21         940
#> 166 2020-07-22         761
#> 167 2020-07-23         695
#> 168 2020-07-24         554
#> 169 2020-07-25         361
#> 170 2020-07-26         232
#> 171 2020-07-27         796
#> 172 2020-07-28         702
#> 173 2020-07-29         625
#> 174 2020-07-30         699
#> 175 2020-07-31         690
#> 176 2020-08-01         389
#> 177 2020-08-02         210
#> 178 2020-08-03         736
#> 179 2020-08-04         674
#> 180 2020-08-05         638
#> 181 2020-08-06         600
#> 182 2020-08-07         527
#> 183 2020-08-08         284
#> 184 2020-08-09         232
#> 185 2020-08-10         586
#> 186 2020-08-11         699
#> 187 2020-08-12         546
#> 188 2020-08-13         533
#> 189 2020-08-14         494
#> 190 2020-08-15         268
#> 191 2020-08-16         220
#> 192 2020-08-17         627
#> 193 2020-08-18         572
#> 194 2020-08-19         543
#> 195 2020-08-20         539
#> 196 2020-08-21         448
#> 197 2020-08-22         279
#> 198 2020-08-23         176
#> 199 2020-08-24         572
#> 200 2020-08-25         604
#> 201 2020-08-26         872
#> 202 2020-08-27         566
#> 203 2020-08-28         470
#> 204 2020-08-29         239
#> 205 2020-08-30         172
#> 206 2020-08-31         596
#> 207 2020-09-01         529
#> 208 2020-09-02         540
#> 209 2020-09-03         516
#> 210 2020-09-04         436
#> 211 2020-09-05         193
#> 212 2020-09-06         131
#> 213 2020-09-07         124
#> 214 2020-09-08         577
#> 215 2020-09-09         757
#> 216 2020-09-10         527
#> 217 2020-09-11         589
#> 218 2020-09-12         245
#> 219 2020-09-13         149
#> 220 2020-09-14         694
#> 221 2020-09-15         781
#> 222 2020-09-16         785
#> 223 2020-09-17         621
#> 224 2020-09-18         597
#> 225 2020-09-19         290
#> 226 2020-09-20         189
#> 227 2020-09-21         599
#> 228 2020-09-22         622
#> 229 2020-09-23         993
#> 230 2020-09-24         713
#> 231 2020-09-25         553
#> 232 2020-09-26         351
#> 233 2020-09-27         205
#> 234 2020-09-28         686
#> 235 2020-09-29         700
#> 236 2020-09-30         735
#> 237 2020-10-01         636
#> 238 2020-10-02         612
#> 239 2020-10-03         263
#> 240 2020-10-04         283
#> 241 2020-10-05         914
#> 242 2020-10-06         902
#> 243 2020-10-07        1028
#> 244 2020-10-08        1226
#> 245 2020-10-09         888
#> 246 2020-10-10         463
#> 247 2020-10-11         270
#> 248 2020-10-12         618
#> 249 2020-10-13         768
#> 250 2020-10-14         682
#> 251 2020-10-15         840
#> 252 2020-10-16         785
#> 253 2020-10-17         501
#> 254 2020-10-18         246
#> 255 2020-10-19        1110
#> 256 2020-10-20         799
#> 257 2020-10-21         831
#> 258 2020-10-22        1035
#> 259 2020-10-23         707
#> 260 2020-10-24         360
#> 261 2020-10-25         264
#> 262 2020-10-26        1344
#> 263 2020-10-27        1085
#> 264 2020-10-28        1100
#> 265 2020-10-29        1243
#> 266 2020-10-30        1017
#> 267 2020-10-31         415
#> 268 2020-11-01         573
#> 269 2020-11-02        1547
#> 270 2020-11-03        1496
#> 271 2020-11-04        1980
#> 272 2020-11-05        1882
#> 273 2020-11-06        1675
#> 274 2020-11-07         760
#> 275 2020-11-08         531
#> 276 2020-11-09        1702
#> 277 2020-11-10        1593
#> 278 2020-11-11        1586
#> 279 2020-11-12        1389
#> 280 2020-11-13        1266
#> 281 2020-11-14         792
#> 282 2020-11-15         817
#> 283 2020-11-16        1768
#> 284 2020-11-17        1795
#> 285 2020-11-18        1440
#> 286 2020-11-19        1074
#> 287 2020-11-20        1093
#> 288 2020-11-21         424
#> 289 2020-11-22         331
#> 290 2020-11-23        1430
#> 291 2020-11-24        1413
#> 292 2020-11-25        1067
#> 293 2020-11-26         210
#> 294 2020-11-27         849
#> 295 2020-11-28         483
#> 296 2020-11-29         472
#> 297 2020-11-30        2395
#> 298 2020-12-01        2247
#> 299 2020-12-02        1888
#> 300 2020-12-03        1439
#> 301 2020-12-04        1231
#> 302 2020-12-05         431
#> 303 2020-12-06         351
#> 304 2020-12-07        1358
#> 305 2020-12-08        1298
#> 306 2020-12-09        2143
#> 307 2020-12-10        2746
#> 308 2020-12-11        2765
#> 309 2020-12-12        1749
#> 310 2020-12-13         863
#> 311 2020-12-14        2572
#> 312 2020-12-15        3274
#> 313 2020-12-16        2240
#> 314 2020-12-17        1313
#> 315 2020-12-18        1248
#> 316 2020-12-19         980
#> 317 2020-12-20         686
#> 318 2020-12-21        1781
#> 319 2020-12-22        1325
#> 320 2020-12-23        1311
#> 321 2020-12-24        1083
#> 322 2020-12-25         359
#> 323 2020-12-26        1900
#> 324 2020-12-27        1431
#> 325 2020-12-28        2651
#> 326 2020-12-29        2532
#> 327 2020-12-30        1898
#> 328 2020-12-31        1552
#> 329 2021-01-01         835
#> 330 2021-01-02        1525
#> 331 2021-01-03        1595
#> 332 2021-01-04        4844
#> 333 2021-01-05        4226
#> 334 2021-01-06        3038
#> 335 2021-01-07        2703
#> 336 2021-01-08        2256
#> 337 2021-01-09        2107
#> 338 2021-01-10        1193
#> 339 2021-01-11        3446
#> 340 2021-01-12        3541
#> 341 2021-01-13        3135
#> 342 2021-01-14        3492
#> 343 2021-01-15        4275
#> 344 2021-01-16        2663
#> 345 2021-01-17        1756
#> 346 2021-01-18        3273
#> 347 2021-01-19        4611
#> 348 2021-01-20        2982
#> 349 2021-01-21        2850
#> 350 2021-01-22        2188
#> 351 2021-01-23        1269
#> 352 2021-01-24        1293
#> 353 2021-01-25        2672
#> 354 2021-01-26        2739
#> 355 2021-01-27        3117
#> 356 2021-01-28        3053
#> 357 2021-01-29        2369
#> 358 2021-01-30        1682
#> 359 2021-01-31        1783
#> 360 2021-02-01        3920
#> 361 2021-02-02        4229
#> 362 2021-02-03        3809
#> 363 2021-02-04        2842
#> 364 2021-02-05        2454
#> 365 2021-02-06        1653
#> 366 2021-02-07        1412
#> 367 2021-02-08        3661
#> 368 2021-02-09        4158
#> 369 2021-02-10        3873
#> 370 2021-02-11        3330
#> 371 2021-02-12        2624
#> 372 2021-02-13        1887
#> 373 2021-02-14        1456
#> 374 2021-02-15        3876
#> 375 2021-02-16        3158
#> 376 2021-02-17        3008
#> 377 2021-02-18        2371
#> 378 2021-02-19        2395
#> 379 2021-02-20        1888
#> 380 2021-02-21        1465
#> 381 2021-02-22        3392
#> 382 2021-02-23        3086
#> 383 2021-02-24        3252
#> 384 2021-02-25        2895
#> 385 2021-02-26        2779
#> 386 2021-02-27        2094
#> 387 2021-02-28        1616
#> 388 2021-03-01        3520
#> 389 2021-03-02        3546
#> 390 2021-03-03        3947
#> 391 2021-03-04        3529
#> 392 2021-03-05        3035
#> 393 2021-03-06        2126
#> 394 2021-03-07        1981
#> 395 2021-03-08        4238
#> 396 2021-03-09        3720
#> 397 2021-03-10        3781
#> 398 2021-03-11        3295
#> 399 2021-03-12        4134
#> 400 2021-03-13        2778
#> 401 2021-03-14        2239
#> 402 2021-03-15        6325
#> 403 2021-03-16        4677
#> 404 2021-03-17        4475
#> 405 2021-03-18        4647
#> 406 2021-03-19        4310
#> 407 2021-03-20        3296
#> 408 2021-03-21        2913
#> 409 2021-03-22        5487
#> 410 2021-03-23        5177
#> 411 2021-03-24        5735
#> 412 2021-03-25        5305
#> 413 2021-03-26        4494
#> 414 2021-03-27        3632
#> 415 2021-03-28        3011
#> 416 2021-03-29        6518
#> 417 2021-03-30        5733
#> 418 2021-03-31        6578
#> 419 2021-04-01        6933
#> 420 2021-04-02        5620
#> 421 2021-04-03        4076
#> 422 2021-04-04        2760
#> 423 2021-04-05        7321
#> 424 2021-04-06        6420
#> 425 2021-04-07        6830
#> 426 2021-04-08        6316
#> 427 2021-04-09        6650
#> 428 2021-04-10        4860
#> 429 2021-04-11        3850
#> 430 2021-04-12        7405
#> 431 2021-04-13        6898
#> 432 2021-04-14        6962
#> 433 2021-04-15        8027
#> 434 2021-04-16        5945
#> 435 2021-04-17        4122
#> 436 2021-04-18        2886
#> 437 2021-04-19        6807
#> 438 2021-04-20        5992
#> 439 2021-04-21        5790
#> 440 2021-04-22        5567
#> 441 2021-04-23        5080
#> 442 2021-04-24        3189
#> 443 2021-04-25        2840
#> 444 2021-04-26        6594
#> 445 2021-04-27        5997
#> 446 2021-04-28        5781
#> 447 2021-04-29        5687
#> 448 2021-04-30        5338
#> 449 2021-05-01        3282
#> 450 2021-05-02        2599
#> 451 2021-05-03        6613
#> 452 2021-05-04        5191
#> 453 2021-05-05        5439
#> 454 2021-05-06        4500
#> 455 2021-05-07        4091
#> 456 2021-05-08        2310
#> 457 2021-05-09        1929
#> 458 2021-05-10        4943
#> 459 2021-05-11        4310
#> 460 2021-05-12        3933
#> 461 2021-05-13        3522
#> 462 2021-05-14        2848
#> 463 2021-05-15        2744
#> 464 2021-05-16        1485
#> 465 2021-05-17        3945
#> 466 2021-05-18        3346
#> 467 2021-05-19        3106
#> 468 2021-05-20        2820
#> 469 2021-05-21        2427
#> 470 2021-05-22        1398
#> 471 2021-05-23         996
#> 472 2021-05-24        2718
#> 473 2021-05-25        2351
#> 474 2021-05-26        2140
#> 475 2021-05-27        2016
#> 476 2021-05-28        1700
#> 477 2021-05-29         882
#> 478 2021-05-30         746
#> 479 2021-05-31         539
#> 480 2021-06-01        2430
#> 481 2021-06-02        1891
#> 482 2021-06-03        1798
#> 483 2021-06-04        1571
#> 484 2021-06-05         932
#> 485 2021-06-06         835
#> 486 2021-06-07        1962
#> 487 2021-06-08        1809
#> 488 2021-06-09        1732
#> 489 2021-06-10        1789
#> 490 2021-06-11        1725
#> 491 2021-06-12        1041
#> 492 2021-06-13        1016
#> 493 2021-06-14        2159
#> 494 2021-06-15        2038
#> 495 2021-06-16        1771
#> 496 2021-06-17        1725
#> 497 2021-06-18        1706
#> 498 2021-06-19        1104
#> 499 2021-06-20         885
#> 500 2021-06-21        2359
#> 501 2021-06-22        2102
#> 502 2021-06-23        2008
#> 503 2021-06-24        2017
#> 504 2021-06-25        1867
#> 505 2021-06-26        1268
#> 506 2021-06-27        1250
#> 507 2021-06-28        2975
#> 508 2021-06-29        2632
#> 509 2021-06-30        2480
#> 510 2021-07-01        2617
#> 511 2021-07-02        2448
#> 512 2021-07-03        1503
#> 513 2021-07-04         674
#> 514 2021-07-05        2459
#> 515 2021-07-06        4377
#> 516 2021-07-07        4396
#> 517 2021-07-08        4202
#> 518 2021-07-09        4087
#> 519 2021-07-10        3162
#> 520 2021-07-11        2678
#> 521 2021-07-12        5992
#> 522 2021-07-13        5230
#> 523 2021-07-14        5336
#> 524 2021-07-15        7208
#> 525 2021-07-16        5807
#> 526 2021-07-17        4333
#> 527 2021-07-18        3794
#> 528 2021-07-19        8520
#> 529 2021-07-20        8603
#> 530 2021-07-21        8710
#> 531 2021-07-22        8509
#> 532 2021-07-23        8434
#> 533 2021-07-24        6345
#> 534 2021-07-25        5003
#> 535 2021-07-26       12001
#> 536 2021-07-27       10855
#> 537 2021-07-28       11070
#> 538 2021-07-29       12160
#> 539 2021-07-30       11535
#> 540 2021-07-31        8142
#> 541 2021-08-01        6891
#> 542 2021-08-02       14135
#> 543 2021-08-03       14035
#> 544 2021-08-04       12940
#> 545 2021-08-05       13151
#> 546 2021-08-06       12992
#> 547 2021-08-07        8232
#> 548 2021-08-08        5704
#> 549 2021-08-09       13714
#> 550 2021-08-10       12649
#> 551 2021-08-11       11373
#> 552 2021-08-12       11001
#> 553 2021-08-13        9537
#> 554 2021-08-14        6076
#> 555 2021-08-15        8564
#> 556 2021-08-16       12932
#> 557 2021-08-17       12016
#> 558 2021-08-18       11228
#> 559 2021-08-19       10226
#> 560 2021-08-20       10543
#> 561 2021-08-21        5867
#> 562 2021-08-22        4101
#> 563 2021-08-23       12127
#> 564 2021-08-24       12506
#> 565 2021-08-25       11780
#> 566 2021-08-26       12262
#> 567 2021-08-27       10267
#> 568 2021-08-28        5436
#> 569 2021-08-29        4177
#> 570 2021-08-30       12952
#> 571 2021-08-31       14338
#> 572 2021-09-01       12468
#> 573 2021-09-02       12422
#> 574 2021-09-03       10581
#> 575 2021-09-04        6194
#> 576 2021-09-05        5276
#> 577 2021-09-06        4771
#> 578 2021-09-07       16700
#> 579 2021-09-08       13528
#> 580 2021-09-09       13325
#> 581 2021-09-10       12609
#> 582 2021-09-11        7025
#> 583 2021-09-12        6155
#> 584 2021-09-13       13378
#> 585 2021-09-14       10882
#> 586 2021-09-15       13657
#> 587 2021-09-16       10253
#> 588 2021-09-17       11070
#> 589 2021-09-18        6026
#> 590 2021-09-19        5324
#> 591 2021-09-20       12383
#> 592 2021-09-21       10896
#> 593 2021-09-22        8482
#> 594 2021-09-23        8556
#> 595 2021-09-24        8979
#> 596 2021-09-25        5082
#> 597 2021-09-26        4232
#> 598 2021-09-27       10781
#> 599 2021-09-28        9897
#> 600 2021-09-29       10155
#> 601 2021-09-30        8756
#> 602 2021-10-01        8889
#> 603 2021-10-02        5537
#> 604 2021-10-03        4834
#> 605 2021-10-04       10978
#> 606 2021-10-05       10756
#> 607 2021-10-06        9847
#> 608 2021-10-07        9126
#> 609 2021-10-08        8613
#> 610 2021-10-09        5803
#> 611 2021-10-10        4890
#> 612 2021-10-11        9882
#> 613 2021-10-12        8598
#> 614 2021-10-13        8297
#> 615 2021-10-14        9017
#> 616 2021-10-15        9994
#> 617 2021-10-16        4880
#> 618 2021-10-17        4078
#> 619 2021-10-18       10324
#> 620 2021-10-19        8862
#> 621 2021-10-20        8388
#> 622 2021-10-21        8355
#> 623 2021-10-22        8646
#> 624 2021-10-23        5711
#> 625 2021-10-24        4441
#> 626 2021-10-25       10846
#> 627 2021-10-26        9149
#> 628 2021-10-27        9277
#> 629 2021-10-28        8861
#> 630 2021-10-29        8130
#> 631 2021-10-30        5540
#> 632 2021-10-31        4448
#> 633 2021-11-01        9780
#> 634 2021-11-02       10389
#> 635 2021-11-03       10648
#> 636 2021-11-04        9821
#> 637 2021-11-05        8805
#> 638 2021-11-06        5921
#> 639 2021-11-07        5541
#> 640 2021-11-08       10607
#> 641 2021-11-09       10967
#> 642 2021-11-10       11443
#> 643 2021-11-11        8792
#> 644 2021-11-12        9617
#> 645 2021-11-13        7237
#> 646 2021-11-14        4855
#> 647 2021-11-15       11915
#> 648 2021-11-16       11290
#> 649 2021-11-17       10584
#> 650 2021-11-18        9316
#> 651 2021-11-19        7937
#> 652 2021-11-20        5166
#> 653 2021-11-21        5410
#> 654 2021-11-22        9798
#> 655 2021-11-23        9429
#> 656 2021-11-24        8416
#> 657 2021-11-25        1320
#> 658 2021-11-26        7053
#> 659 2021-11-27        6905
#> 660 2021-11-28        6187
#> 661 2021-11-29       13768
#> 662 2021-11-30       11291
#> 663 2021-12-01       11855
#> 664 2021-12-02       10064
#> 665 2021-12-03        9607
#> 666 2021-12-04        7050
#> 667 2021-12-05        4762
#> 668 2021-12-06       12074
#> 669 2021-12-07       10855
#> 670 2021-12-08       11200
#> 671 2021-12-09        9874
#> 672 2021-12-10        8883
#> 673 2021-12-11        6037
#> 674 2021-12-12        4700
#> 675 2021-12-13       11059
#> 676 2021-12-14       12886
#> 677 2021-12-15       15059
#> 678 2021-12-16       12453
#> 679 2021-12-17       12488
#> 680 2021-12-18        8138
#> 681 2021-12-19        5988
#> 682 2021-12-20       14531
#> 683 2021-12-21       18023
#> 684 2021-12-22       19722
#> 685 2021-12-23       15041
#> 686 2021-12-24        5616
#> 687 2021-12-25        2648
#> 688 2021-12-26       10143
#> 689 2021-12-27       21329
#> 690 2021-12-28       20838
#> 691 2021-12-29       21657
#> 692 2021-12-30       20194
#> 693 2021-12-31        7586
#> 694 2022-01-01        4483
#> 695 2022-01-02       10766
#> 696 2022-01-03       23193
#> 697 2022-01-04       26175
#> 698 2022-01-05       18291
#> 699 2022-01-06       15949
#> 700 2022-01-07       15026
#> 701 2022-01-08        9578
#> 702 2022-01-09        8626
#> 703 2022-01-10       19597
#> 704 2022-01-11       18805
#> 705 2022-01-12       16609
#> 706 2022-01-13       15508
#> 707 2022-01-14       15040
#> 708 2022-01-15       12521
#> 709 2022-01-16        7772
#> 710 2022-01-17       13682
#> 711 2022-01-18       20420
#> 712 2022-01-19       20398
#> 713 2022-01-20       15171
#> 714 2022-01-21       13195
#> 715 2022-01-22        8767
#> 716 2022-01-23        7469
#> 717 2022-01-24       16406
#> 718 2022-01-25       14974
#> 719 2022-01-26       14083
#> 720 2022-01-27       14616
#> 721 2022-01-28       11309
#> 722 2022-01-29        8259
#> 723 2022-01-30        7182
#> 724 2022-01-31       16718
#> 725 2022-02-01       14999
#> 726 2022-02-02       13704
#> 727 2022-02-03       11191
#> 728 2022-02-04       10092
#> 729 2022-02-05        6242
#> 730 2022-02-06        6134
#> 731 2022-02-07       15683
#> 732 2022-02-08       13411
#> 733 2022-02-09       12065
#> 734 2022-02-10       10738
#> 735 2022-02-11        9036
#> 736 2022-02-12        6158
#> 737 2022-02-13        4487
#> 738 2022-02-14        9575
#> 739 2022-02-15       10035
#> 740 2022-02-16        9349
#> 741 2022-02-17        7550
#> 742 2022-02-18        6606
#> 743 2022-02-19        4087
#> 744 2022-02-20        3031
#> 745 2022-02-21        5961
#> 746 2022-02-22        7938
#> 747 2022-02-23        5755
#> 748 2022-02-24        4889
#> 749 2022-02-25        4514
#> 750 2022-02-26        2947
#> 751 2022-02-27        2285
#> 752 2022-02-28        4754
#> 753 2022-03-01        4549
#> 754 2022-03-02        3557
#> 755 2022-03-03        3032
#> 756 2022-03-04        2991
#> 757 2022-03-05        1917
#> 758 2022-03-06        1639
#> 759 2022-03-07        3950
#> 760 2022-03-08        2964
#> 761 2022-03-09        2818
#> 762 2022-03-10        2712
#> 763 2022-03-11        2355
#> 764 2022-03-12        1419
#> 765 2022-03-13        1541
#> 766 2022-03-14        3412
#> 767 2022-03-15        3118
#> 768 2022-03-16        2805
#> 769 2022-03-17        2575
#> 770 2022-03-18        2415
#> 771 2022-03-19        1887
#> 772 2022-03-20        1850
#> 773 2022-03-21        3637
#> 774 2022-03-22        3168
#> 775 2022-03-23        3167
#> 776 2022-03-24        2888
#> 777 2022-03-25        2693
#> 778 2022-03-26        2179
#> 779 2022-03-27        2218
#> 780 2022-03-28        3845
#> 781 2022-03-29        3419
#> 782 2022-03-30        3585
#> 783 2022-03-31        3304
#> 784 2022-04-01        3270
#> 785 2022-04-02        2464
#> 786 2022-04-03        2356
#> 787 2022-04-04        5243
#> 788 2022-04-05        4540
#> 789 2022-04-06        4304
#> 790 2022-04-07        4538
#> 791 2022-04-08        4597
#> 792 2022-04-09        2745
#> 793 2022-04-10        2801
#> 794 2022-04-11        5707
#> 795 2022-04-12        4412
#> 796 2022-04-13        4121
#> 797 2022-04-14        4184
#> 798 2022-04-15        5215
#> 799 2022-04-16        3373
#> 800 2022-04-17        1549
#> 801 2022-04-18        6012
#> 802 2022-04-19        5018
#> 803 2022-04-20        4790
#> 804 2022-04-21        4605
#> 805 2022-04-22        4757
#> 806 2022-04-23        3361
#> 807 2022-04-24        3555
#> 808 2022-04-25        6325
#> 809 2022-04-26        5646
#> 810 2022-04-27        4892
#> 811 2022-04-28        4161
#> 812 2022-04-29        4456
#> 813 2022-04-30        3587
#> 814 2022-05-01        3114
#> 815 2022-05-02        6721
#> 816 2022-05-03        6173
#> 817 2022-05-04        6007
#> 818 2022-05-05        6061
#> 819 2022-05-06        5555
#> 820 2022-05-07        4258
#> 821 2022-05-08        3788
#> 822 2022-05-09        7996
#> 823 2022-05-10        6640
#> 824 2022-05-11        6347
#> 825 2022-05-12        6229
#> 826 2022-05-13        5948
#> 827 2022-05-14        4652
#> 828 2022-05-15        4665
#> 829 2022-05-16        7725
#> 830 2022-05-17        6800
#> 831 2022-05-18        6450
#> 832 2022-05-19        5277
#> 833 2022-05-20        5436
#> 834 2022-05-21        4515
#> 835 2022-05-22        4424
#> 836 2022-05-23        7782
#> 837 2022-05-24        7347
#> 838 2022-05-25        6492
#> 839 2022-05-26        5266
#> 840 2022-05-27        5749
#> 841 2022-05-28        3920
#> 842 2022-05-29        4366
#> 843 2022-05-30        2151
#> 844 2022-05-31        9393
#> 845 2022-06-01        6439
#> 846 2022-06-02        6023
#> 847 2022-06-03        5579
#> 848 2022-06-04        3994
#> 849 2022-06-05        3944
#> 850 2022-06-06        7654
#> 851 2022-06-07        7161
#> 852 2022-06-08        6198
#> 853 2022-06-09        5910
#> 854 2022-06-10        5212
#> 855 2022-06-11        3513
#> 856 2022-06-12        3316
#> 857 2022-06-13        5955
#> 858 2022-06-14        5001
#> 859 2022-06-15        5634
#> 860 2022-06-16        5074
#> 861 2022-06-17        4795
#> 862 2022-06-18        3314
#> 863 2022-06-19        2513
#> 864 2022-06-20        4869
#> 865 2022-06-21        5408
#> 866 2022-06-22        5396
#> 867 2022-06-23        4784
#> 868 2022-06-24        4446
#> 869 2022-06-25        3097
#> 870 2022-06-26        3250
#> 871 2022-06-27        6468
#> 872 2022-06-28        4956
#> 873 2022-06-29        4518
#> 874 2022-06-30        4117
#> 875 2022-07-01        4822
#> 876 2022-07-02        2984
#> 877 2022-07-03        2965
#> 878 2022-07-04        1541
#> 879 2022-07-05        8079
#> 880 2022-07-06        5471
#> 881 2022-07-07        4389
#> 882 2022-07-08        4632
#> 883 2022-07-09        3575
#> 884 2022-07-10        3379
#> 885 2022-07-11        4637
#> 886 2022-07-12        3291
#> 887 2022-07-13        2985
#> 888 2022-07-14        2963
#> 889 2022-07-15        2513
#> 890 2022-07-16        1993
#> 891 2022-07-17        1779
#> 892 2022-07-18        2173
#> 893 2022-07-19        1310
#> 894 2022-07-20        1078
#> 895 2022-07-21         924
#> 896 2022-07-22         877
#> 897 2022-07-23         435
#> 898 2022-07-24         220
#> 899 2022-07-25         278
#> 900 2022-07-26         104
#> 901 2022-07-27          18
#> 902 2022-07-28          12

# Returns the total number of global sequences in the outbreak.info API.
getSeqCounts(cumulative=TRUE)
#> Retrieving data... 
#>           key    value
#> 1 total_count 11907184