Lookup sublineages for a given Pango lineage or WHO name

Retrieve all sublineages of a given Pango lineage or WHO designation. Based on a phylogeny maintained by the [Pango team](https://github.com/cov-lineages/lineages-website/blob/master/data/lineages.yml) to grab all sublineages associated with a lineage, and a curated list of Variants of Concern, Variants of Interest, Variants Under Monitoring, and de-escalated variants maintained by the [outbreak.info team](https://outbreak.info/situation-reports)

lookupSublineages(lineage, returnQueryString = FALSE)

Arguments

lineage

String with the name of a Pango lineage or WHO name

returnQueryString

Boolean to return a query string to be piped into functions like getPrevalence (collapses vector by `"OR"`)

Value

list containing the sublineages associated with that lineage or WHO designation

Examples

# WHO-designated lineages
lookupSublineages("Delta")
#>   [1] "AY.1"       "AY.2"       "AY.3"       "AY.3.1"     "AY.3.2"    
#>   [6] "AY.3.3"     "AY.3.4"     "AY.4"       "AY.4.1"     "AY.4.2"    
#>  [11] "AY.4.2.1"   "AY.4.2.2"   "AY.4.2.3"   "AY.4.2.4"   "AY.4.2.5"  
#>  [16] "AY.4.3"     "AY.4.4"     "AY.4.5"     "AY.4.6"     "AY.4.7"    
#>  [21] "AY.4.8"     "AY.4.9"     "AY.4.10"    "AY.4.11"    "AY.4.12"   
#>  [26] "AY.4.13"    "AY.4.14"    "AY.4.15"    "AY.4.16"    "AY.4.17"   
#>  [31] "AY.5"       "AY.5.1"     "AY.5.2"     "AY.5.3"     "AY.5.4"    
#>  [36] "AY.5.5"     "AY.5.6"     "AY.5.7"     "AY.6"       "AY.7"      
#>  [41] "AY.7.1"     "AY.7.2"     "AY.8"       "AY.9"       "AY.9.1"    
#>  [46] "AY.9.2"     "AY.9.2.1"   "AY.9.2.2"   "AY.10"      "AY.11"     
#>  [51] "AY.12"      "AY.13"      "AY.14"      "AY.15"      "AY.16"     
#>  [56] "AY.16.1"    "AY.17"      "AY.18"      "AY.19"      "AY.20"     
#>  [61] "AY.20.1"    "AY.21"      "AY.22"      "AY.23"      "AY.23.1"   
#>  [66] "AY.23.2"    "AY.24"      "AY.24.1"    "AY.25"      "AY.25.1"   
#>  [71] "AY.25.1.1"  "AY.25.1.2"  "AY.25.2"    "AY.25.3"    "AY.26"     
#>  [76] "AY.26.1"    "AY.27"      "AY.28"      "AY.29"      "AY.29.1"   
#>  [81] "AY.29.2"    "AY.30"      "AY.31"      "AY.32"      "AY.33"     
#>  [86] "AY.33.1"    "AY.33.2"    "AY.34"      "AY.34.1"    "AY.34.1.1" 
#>  [91] "AY.34.2"    "AY.35"      "AY.36"      "AY.36.1"    "AY.37"     
#>  [96] "AY.38"      "AY.39"      "AY.39.1"    "AY.39.1.1"  "AY.39.1.2" 
#> [101] "AY.39.1.3"  "AY.39.1.4"  "AY.39.2"    "AY.39.3"    "AY.39.4"   
#> [106] "AY.40"      "AY.41"      "AY.42"      "AY.42.1"    "AY.43"     
#> [111] "AY.43.1"    "AY.43.2"    "AY.43.3"    "AY.43.4"    "AY.43.5"   
#> [116] "AY.43.6"    "AY.43.7"    "AY.43.8"    "AY.43.9"    "AY.44"     
#> [121] "AY.45"      "AY.46"      "AY.46.1"    "AY.46.2"    "AY.46.3"   
#> [126] "AY.46.4"    "AY.46.5"    "AY.46.6"    "AY.46.6.1"  "AY.47"     
#> [131] "AY.48"      "AY.49"      "AY.50"      "AY.51"      "AY.52"     
#> [136] "AY.53"      "AY.54"      "AY.55"      "AY.56"      "AY.57"     
#> [141] "AY.58"      "AY.59"      "AY.60"      "AY.61"      "AY.62"     
#> [146] "AY.63"      "AY.64"      "AY.65"      "AY.66"      "AY.67"     
#> [151] "AY.68"      "AY.69"      "AY.70"      "AY.71"      "AY.72"     
#> [156] "AY.73"      "AY.74"      "AY.75"      "AY.75.1"    "AY.75.2"   
#> [161] "AY.75.3"    "AY.76"      "AY.77"      "AY.78"      "AY.79"     
#> [166] "AY.80"      "AY.81"      "AY.82"      "AY.83"      "AY.84"     
#> [171] "AY.85"      "AY.86"      "AY.87"      "AY.88"      "AY.89"     
#> [176] "AY.90"      "AY.91"      "AY.91.1"    "AY.92"      "AY.93"     
#> [181] "AY.94"      "AY.95"      "AY.96"      "AY.97"      "AY.98"     
#> [186] "AY.98.1"    "AY.99"      "AY.99.1"    "AY.99.2"    "AY.100"    
#> [191] "AY.101"     "AY.102"     "AY.102.1"   "AY.102.2"   "AY.103"    
#> [196] "AY.103.1"   "AY.103.2"   "AY.104"     "AY.105"     "AY.106"    
#> [201] "AY.107"     "AY.108"     "AY.109"     "AY.110"     "AY.111"    
#> [206] "AY.112"     "AY.112.1"   "AY.112.2"   "AY.112.3"   "AY.113"    
#> [211] "AY.114"     "AY.115"     "AY.116"     "AY.116.1"   "AY.117"    
#> [216] "AY.118"     "AY.119"     "AY.119.1"   "AY.119.2"   "AY.120"    
#> [221] "AY.120.1"   "AY.120.2"   "AY.120.2.1" "AY.121"     "AY.121.1"  
#> [226] "AY.122"     "AY.122.1"   "AY.122.2"   "AY.122.3"   "AY.122.4"  
#> [231] "AY.122.5"   "AY.122.6"   "AY.123"     "AY.123.1"   "AY.124"    
#> [236] "AY.124.1"   "AY.124.1.1" "AY.125"     "AY.125.1"   "AY.126"    
#> [241] "AY.127"     "AY.127.1"   "AY.127.2"   "AY.127.3"   "AY.128"    
#> [246] "AY.129"     "AY.130"     "AY.131"     "AY.132"     "AY.133"    
#> [251] "AY.134"     "B.1.617.2" 
lookupSublineages("epsilon")
#> [1] "B.1.427"   "B.1.429"   "B.1.429.1"
lookupSublineages("epsilon", returnQueryString = TRUE)
#> [1] "B.1.427 OR B.1.429 OR B.1.429.1"
# Pango lineage
lookupSublineages("B.1.1.7")
#> [1] "B.1.1.7" "Q.1"     "Q.2"     "Q.3"     "Q.4"     "Q.5"     "Q.6"    
#> [8] "Q.7"     "Q.8"    
#' # Not a recognized lineage
lookupSublineages("VOC-21APR-02")
#> Warning: Lineage VOC-21APR-02 not recognized as a Pango lineage
#> Error in invokeRestart("muffleWarning"): no 'restart' 'muffleWarning' found