Retrieve cumulative prevalence of a lineage

Retrieve cumulative prevalence of a PANGO lineage by the immediate administrative level of a location

getCumulativeBySubadmin(
  pangolin_lineage,
  location = NULL,
  mutations = NULL,
  ndays = NULL
)

Arguments

pangolin_lineage:
PANGO lineage name

location:
(optional) a location name (if not specified, cumulative prevalence at the country level globally returned)

mutations:
(optional) a `vector` of mutation(s)

ndays:
(optional) an `integer` specifying number of days from current date to calculative cumuative counts (if not specified, no limit on window)

Value

dataframe

Examples

getCumulativeBySubadmin(pangolin_lineage="B.1.1.7", location="California")

#> Retrieving data... 

#>          date                             name              id total_count
#> 35 2020-10-29  Napa Solano Yolo Marin Counties  USA_US-CA_None           6
#> 42 2020-12-23                          Redding  USA_US-CA_None           2
#> 2  2020-12-25                       Alexandria  USA_US-CA_None           1
#> 37 2020-12-28                         New York  USA_US-CA_None           1
#> 69 2021-01-08                            UNKNO  USA_US-CA_None           1
#> 30 2021-01-20                          Modesto  USA_US-CA_None           1
#> 58 2021-02-02                           Sierra USA_US-CA_06091           1
#> 75 2021-02-19                      unavailable  USA_US-CA_None           1
#> 31 2021-03-09                            Modoc USA_US-CA_06049           1
#> 11 2021-03-16                          Douglas  USA_US-CA_None           1
#> 66 2021-04-10                          Trinity USA_US-CA_06105           2
#> 72 2021-04-20                           Washoe  USA_US-CA_None           1
#> 6  2021-04-28                Californialaveras  USA_US-CA_None           1
#> 25 2021-04-29                         Maricopa  USA_US-CA_None           1
#> 17 2021-04-30                             Inyo USA_US-CA_06027           1
#> 53 2021-05-02                       San Mateo*  USA_US-CA_None           1
#> 7  2021-05-07                           Colusa USA_US-CA_06011           1
#> 32 2021-05-19                             Mono USA_US-CA_06051           1
#> 27 2021-05-28                         Mariposa USA_US-CA_06043           1
#> 41 2021-06-02                           Plumas USA_US-CA_06063           1
#> 21 2021-06-04                           Lassen USA_US-CA_06035           3
#> 10 2021-06-07                        Del Norte USA_US-CA_06015           1
#> 20 2021-06-07                             Lake USA_US-CA_06033           5
#> 22 2021-06-08                  Long Beach City  USA_US-CA_None           1
#> 19 2021-06-09                            Kings USA_US-CA_06031           1
#> 28 2021-06-09                        Mendocino USA_US-CA_06045           1
#> 65 2021-06-09                           Tehama USA_US-CA_06103           2
#> 54 2021-06-10                    Santa Barbara USA_US-CA_06083           6
#> 57 2021-06-10                           Shasta USA_US-CA_06089           1
#> 4  2021-06-11                            Butte USA_US-CA_06007           2
#> 18 2021-06-11                             Kern USA_US-CA_06029           1
#> 36 2021-06-11                           Nevada USA_US-CA_06057           2
#> 68 2021-06-11                         Tuolumne USA_US-CA_06109           1
#> 71 2021-06-11                          Ventura USA_US-CA_06111           1
#> 14 2021-06-12                            Glenn USA_US-CA_06021           1
#> 59 2021-06-12                         Siskiyou USA_US-CA_06093           1
#> 44 2021-06-21                        SANTA CLA  USA_US-CA_None           2
#> 3  2021-07-06                           Amador USA_US-CA_06005           1
#> 5  2021-07-06                        Calaveras USA_US-CA_06009           1
#> 64 2021-07-08                           Sutter USA_US-CA_06101           1
#> 24 2021-07-12                           Madera USA_US-CA_06039           2
#> 34 2021-07-12                             Napa USA_US-CA_06055           1
#> 74 2021-07-12                             Yuba USA_US-CA_06115           1
#> 61 2021-07-13                           Sonoma USA_US-CA_06097           1
#> 9  2021-07-14                            Davis  USA_US-CA_None           4
#> 12 2021-07-15                        El Dorado USA_US-CA_06017           1
#> 13 2021-07-15                           Fresno USA_US-CA_06019           3
#> 29 2021-07-15                           Merced USA_US-CA_06047           1
#> 60 2021-07-15                           Solano USA_US-CA_06095           1
#> 63 2021-07-15                       Stanislaus USA_US-CA_06099           3
#> 73 2021-07-15                             Yolo USA_US-CA_06113           2
#> 45 2021-07-16                       Sacramento USA_US-CA_06067           2
#> 52 2021-07-16                        San Mateo USA_US-CA_06081           1
#> 8  2021-07-22                     Contra Costa USA_US-CA_06013           1
#> 46 2021-07-29                       San Benito USA_US-CA_06069           1
#> 26 2021-08-02                            Marin USA_US-CA_06041           1
#> 40 2021-08-02                           Placer USA_US-CA_06061           1
#> 67 2021-08-03                           Tulare USA_US-CA_06107           1
#> 15 2021-08-05                         Humboldt USA_US-CA_06023           1
#> 51 2021-08-05                  San Luis Obispo USA_US-CA_06079           6
#> 47 2021-08-11                   San Bernardino USA_US-CA_06071           1
#> 50 2021-08-13                      San Joaquin USA_US-CA_06077           1
#> 62 2021-08-15      Southern San Joaquin Valley  USA_US-CA_None          17
#> 16 2021-08-16                         Imperial USA_US-CA_06025           1
#> 43 2021-08-16                        Riverside USA_US-CA_06065           1
#> 38 2021-08-17 North America / USA / California  USA_US-CA_None         494
#> 49 2021-08-17                    San Francisco USA_US-CA_06075          14
#> 23 2021-08-22                      Los Angeles USA_US-CA_06037           1
#> 33 2021-08-23                         Monterey USA_US-CA_06053           1
#> 48 2021-08-23                        San Diego USA_US-CA_06073           1
#> 70 2021-08-23                          Unknown  USA_US-CA_None           1
#> 1  2021-08-25                          Alameda USA_US-CA_06001           1
#> 56 2021-08-25                       Santa Cruz USA_US-CA_06087          18
#> 39 2021-08-26                           Orange USA_US-CA_06059           1
#> 55 2021-08-26                      Santa Clara USA_US-CA_06085           7
#>    lineage_count cum_total_count cum_lineage_count proportion
#> 35             0              27                 0 0.00000000
#> 42             0               2                 0 0.00000000
#> 2              0               1                 0 0.00000000
#> 37             0               1                 0 0.00000000
#> 69             0               1                 0 0.00000000
#> 30             0               1                 0 0.00000000
#> 58             0               1                 0 0.00000000
#> 75             0               4                 0 0.00000000
#> 31             0               4                 0 0.00000000
#> 11             0               2                 1 0.50000000
#> 66             2               3                 2 0.66666667
#> 72             0               6                 0 0.00000000
#> 6              0               6                 0 0.00000000
#> 25             0               1                 0 0.00000000
#> 17             1              10                 1 0.10000000
#> 53             1               1                 1 1.00000000
#> 7              0              14                 2 0.14285714
#> 32             1             132                 7 0.05303030
#> 27             1              39                 7 0.17948718
#> 41             0               6                 1 0.16666667
#> 21             0              11                 1 0.09090909
#> 10             0              66                15 0.22727273
#> 20             2             225                15 0.06666667
#> 22             0              75                 2 0.02666667
#> 19             0             592                59 0.09966216
#> 28             1             325                51 0.15692308
#> 65             0             180                13 0.07222222
#> 54             0            1098               149 0.13570128
#> 57             1             277                51 0.18411552
#> 4              0             325                89 0.27384615
#> 18             1             520                30 0.05769231
#> 36             0             317                75 0.23659306
#> 68             1             101                 9 0.08910891
#> 71             0             398                70 0.17587940
#> 14             1             119                10 0.08403361
#> 59             0              28                 8 0.28571429
#> 44             0              60                17 0.28333333
#> 3              0              54                 8 0.14814815
#> 5              0              80                12 0.15000000
#> 64             0             199                 9 0.04522613
#> 24             0             704                39 0.05539773
#> 34             0             193                 8 0.04145078
#> 74             0             113                 6 0.05309735
#> 61             0             569                33 0.05799649
#> 9              0             224                34 0.15178571
#> 12             1             214                58 0.27102804
#> 13             0            1785               209 0.11708683
#> 29             0             654                45 0.06880734
#> 60             0             491               101 0.20570265
#> 63             0             762                55 0.07217848
#> 73             0             137                10 0.07299270
#> 45             0            1302               243 0.18663594
#> 52             0             324                30 0.09259259
#> 8              0            2357               224 0.09503606
#> 46             0             225                 6 0.02666667
#> 26             0             399                11 0.02756892
#> 40             0             348                64 0.18390805
#> 67             0             765                31 0.04052288
#> 15             0            1581               221 0.13978495
#> 51             0            1434               217 0.15132497
#> 47             0            1008                69 0.06845238
#> 50             0             696                78 0.11206897
#> 62             0             681                96 0.14096916
#> 16             0             593                43 0.07251265
#> 43             0            2077                91 0.04381319
#> 38             0           41461              4704 0.11345602
#> 49             0            2343                68 0.02902262
#> 23             0            6496               597 0.09190271
#> 33             0            1205                44 0.03651452
#> 48             0           19014              1302 0.06847586
#> 70             0           31072              3589 0.11550592
#> 1              0            4438               283 0.06376746
#> 56             0            1596                32 0.02005013
#> 39             0            2507                46 0.01834862
#> 55             0            6714               327 0.04870420
#>    proportion_ci_lower proportion_ci_upper query_key
#> 35        1.801870e-05          0.08805083   B.1.1.7
#> 42        2.169085e-04          0.66682175   B.1.1.7
#> 2         3.855810e-04          0.85325368   B.1.1.7
#> 37        3.855810e-04          0.85325368   B.1.1.7
#> 69        3.855810e-04          0.85325368   B.1.1.7
#> 30        3.855810e-04          0.85325368   B.1.1.7
#> 58        3.855810e-04          0.85325368   B.1.1.7
#> 75        1.153328e-04          0.44476262   B.1.1.7
#> 31        1.153328e-04          0.44476262   B.1.1.7
#> 11        6.083028e-02          0.93916972   B.1.1.7
#> 66        1.767361e-01          0.96125238   B.1.1.7
#> 72        7.849985e-05          0.33038891   B.1.1.7
#> 6         7.849985e-05          0.33038891   B.1.1.7
#> 25        3.855810e-04          0.85325368   B.1.1.7
#> 17        1.101167e-02          0.38131477   B.1.1.7
#> 53        1.467463e-01          0.99961442   B.1.1.7
#> 7         3.092213e-02          0.38489432   B.1.1.7
#> 32        2.403017e-02          0.10128549   B.1.1.7
#> 27        8.405528e-02          0.32041900   B.1.1.7
#> 41        1.862013e-02          0.55805677   B.1.1.7
#> 21        9.991460e-03          0.35297580   B.1.1.7
#> 10        1.392333e-01          0.33862720   B.1.1.7
#> 20        3.948495e-02          0.10488938   B.1.1.7
#> 22        5.581996e-03          0.08279463   B.1.1.7
#> 19        7.748514e-02          0.12574083   B.1.1.7
#> 28        1.204940e-01          0.19942927   B.1.1.7
#> 65        4.108569e-02          0.11695423   B.1.1.7
#> 54        1.164119e-01          0.15690799   B.1.1.7
#> 57        1.418809e-01          0.23291002   B.1.1.7
#> 4         2.275237e-01          0.32417611   B.1.1.7
#> 18        4.005974e-02          0.08022698   B.1.1.7
#> 36        1.923425e-01          0.28562803   B.1.1.7
#> 68        4.501016e-02          0.15629529   B.1.1.7
#> 71        1.408921e-01          0.21556006   B.1.1.7
#> 14        4.402591e-02          0.14393286   B.1.1.7
#> 59        1.453633e-01          0.46783688   B.1.1.7
#> 44        1.814556e-01          0.40554610   B.1.1.7
#> 3         7.261794e-02          0.26016415   B.1.1.7
#> 5         8.469949e-02          0.24006484   B.1.1.7
#> 64        2.261020e-02          0.08090527   B.1.1.7
#> 24        4.029186e-02          0.07414727   B.1.1.7
#> 34        1.978854e-02          0.07668240   B.1.1.7
#> 74        2.246356e-02          0.10615389   B.1.1.7
#> 61        4.099304e-02          0.07947853   B.1.1.7
#> 9         1.093854e-01          0.20311295   B.1.1.7
#> 12        2.148351e-01          0.33336662   B.1.1.7
#> 13        1.027938e-01          0.13262028   B.1.1.7
#> 29        5.128955e-02          0.09012906   B.1.1.7
#> 60        1.717408e-01          0.24312201   B.1.1.7
#> 63        5.541120e-02          0.09218667   B.1.1.7
#> 73        3.814483e-02          0.12560626   B.1.1.7
#> 45        1.661812e-01          0.20848205   B.1.1.7
#> 52        6.464051e-02          0.12777897   B.1.1.7
#> 8         8.369343e-02          0.10737242   B.1.1.7
#> 46        1.120555e-02          0.05413297   B.1.1.7
#> 26        1.473375e-02          0.04720385   B.1.1.7
#> 40        1.459113e-01          0.22713637   B.1.1.7
#> 67        2.824014e-02          0.05627044   B.1.1.7
#> 15        1.233562e-01          0.15753102   B.1.1.7
#> 51        1.334874e-01          0.17056824   B.1.1.7
#> 47        5.408936e-02          0.08528832   B.1.1.7
#> 50        9.024986e-02          0.13710569   B.1.1.7
#> 62        1.163748e-01          0.16860724   B.1.1.7
#> 16        5.369930e-02          0.09548401   B.1.1.7
#> 43        3.563712e-02          0.05325910   B.1.1.7
#> 38        1.104303e-01          0.11653574   B.1.1.7
#> 49        2.279506e-02          0.03641224   B.1.1.7
#> 23        8.505924e-02          0.09910972   B.1.1.7
#> 33        2.701067e-02          0.04823978   B.1.1.7
#> 48        6.495155e-02          0.07213153   B.1.1.7
#> 70        1.119878e-01          0.11909567   B.1.1.7
#> 1         5.686180e-02          0.07124183   B.1.1.7
#> 56        1.401393e-02          0.02782292   B.1.1.7
#> 39        1.363527e-02          0.02417232   B.1.1.7
#> 55        4.374871e-02          0.05404865   B.1.1.7