Retrieve cumulative prevalence of a lineage

Retrieve cumulative prevalence of a PANGO lineage by the immediate administrative level of a location

getCumulativeBySubadmin(
  pangolin_lineage,
  location = NULL,
  mutations = NULL,
  ndays = NULL
)

Arguments

pangolin_lineage

PANGO lineage name

location

(optional) a location name (if not specified, cumulative prevalence at the country level globally returned)

mutations

(optional) a `vector` of mutation(s)

ndays

(optional) an `integer` specifying number of days from current date to calculative cumuative counts (if not specified, no limit on window)

Value

dataframe

Examples

# Worldwide prevalence of B.1.1.7 by country
head(getCumulativeBySubadmin(pangolin_lineage="B.1.1.7"))
#> Retrieving data... 
#>           date              name  id total_count lineage_count cum_total_count
#> 46  2020-03-24              Cuba CUB           1             0               2
#> 79  2021-02-05     Guinea-Bissau GNB           3             2              48
#> 171 2021-03-05           Somalia SOM           1             1              30
#> 200 2021-03-26           Vanuatu VUT           1             0               2
#> 203 2021-04-19 Wallis and Futuna WLF           1             1              10
#> 62  2021-05-04     Faroe Islands FRO           1             1              42
#>     cum_lineage_count proportion proportion_ci_lower proportion_ci_upper
#> 46                  0 0.00000000        0.0002169085           0.6668218
#> 79                 32 0.66666667        0.5266543998           0.7872517
#> 171                 7 0.23333333        0.1109209072           0.4043999
#> 200                 1 0.50000000        0.0608302759           0.9391697
#> 203                10 1.00000000        0.7828037325           0.9999521
#> 62                  2 0.04761905        0.0100261415           0.1441004
#>     lineage
#> 46  B.1.1.7
#> 79  B.1.1.7
#> 171 B.1.1.7
#> 200 B.1.1.7
#> 203 B.1.1.7
#> 62  B.1.1.7

# County-level prevalence of B.1.1.7
head(getCumulativeBySubadmin(pangolin_lineage="B.1.1.7", location="California"))
#> Retrieving data... 
#>          date   name              id total_count lineage_count cum_total_count
#> 26 2021-03-09  Modoc USA_US-CA_06049           1             0               5
#> 3  2021-09-14 Alpine USA_US-CA_06003           1             0               2
#> 19 2021-09-14 Lassen USA_US-CA_06035         113             0             136
#> 47 2021-12-01 Sierra USA_US-CA_06091           1             0               8
#> 33 2021-12-10 Plumas USA_US-CA_06063           1             0              10
#> 15 2021-12-25   Inyo USA_US-CA_06027           1             0              75
#>    cum_lineage_count proportion proportion_ci_lower proportion_ci_upper lineage
#> 26                 0 0.00000000        0.0000934206          0.37937714 B.1.1.7
#> 3                  0 0.00000000        0.0002169085          0.66682175 B.1.1.7
#> 19                 2 0.01470588        0.0030681956          0.04633378 B.1.1.7
#> 47                 0 0.00000000        0.0000594903          0.26221725 B.1.1.7
#> 33                 1 0.10000000        0.0110116738          0.38131477 B.1.1.7
#> 15                 2 0.02666667        0.0055819955          0.08279463 B.1.1.7