Retrieve cumulative prevalence of a lineage

Retrieve cumulative prevalence of a PANGO lineage by the immediate administrative level of a location

getCumulativeBySubadmin(
  pangolin_lineage,
  location = NULL,
  mutations = NULL,
  ndays = NULL
)

Arguments

pangolin_lineage:
PANGO lineage name

location:
(optional) a location name (if not specified, cumulative prevalence at the country level globally returned)

mutations:
(optional) a `vector` of mutation(s)

ndays:
(optional) an `integer` specifying number of days from current date to calculative cumuative counts (if not specified, no limit on window)

Value

dataframe

Examples

# Worldwide prevalence of B.1.1.7 by country
getCumulativeBySubadmin(pangolin_lineage="B.1.1.7")

#> Retrieving data... 

#>           date                             name   id total_count lineage_count
#> 157 2020-03-13                 Saint-Barthélemy  BLM           1             0
#> 158 2020-04-07                     Saint-Martin  MAF           1             0
#> 180 2020-04-18                      Timor-Leste  TLS           1             0
#> 43  2021-01-10                          Comoros  COM           3             0
#> 22  2021-01-12                          Bermuda  BMU           1             0
#> 67  2021-01-14                 French Polynesia  PYF           1             0
#> 104 2021-01-18                          Lesotho  LSO           8             0
#> 20  2021-01-20                           Belize  BLZ           1             1
#> 1   2021-01-21                      Afghanistan  AFG           8             0
#> 162 2021-01-31                     Sierra Leone  SLE           1             0
#> 80  2021-02-05                    Guinea-Bissau  GNB           3             2
#> 33  2021-02-08                       Cabo Verde None           1             0
#> 31  2021-02-15                     Burkina Faso  BFA           3             0
#> 85  2021-02-26                          Hungary  HUN           1             1
#> 140 2021-02-28                           Panama  PAN           2             0
#> 111 2021-03-01                       Madagascar  MDG           3             0
#> 21  2021-03-03                            Benin  BEN           3             1
#> 167 2021-03-05                          Somalia  SOM           1             1
#> 123 2021-03-09                         Mongolia  MNG           1             0
#> 37  2021-03-10                   Canary Islands None           3             3
#> 175 2021-03-23                        Swaziland  SWZ           2             0
#> 55  2021-04-01                         Dominica  DMA           1             1
#> 60  2021-04-01                Equatorial Guinea  GNQ           1             0
#> 133 2021-04-01                            Niger  NER           6             0
#> 115 2021-04-05                             Mali  MLI           1             0
#> 139 2021-04-09                        Palestina  PSE          19            16
#> 14  2021-04-14                          Bahamas  BHS           3             3
#> 35  2021-04-17                         Cameroon  CMR           2             0
#> 195 2021-04-19        Wallis and Futuna Islands None           1             1
#> 39  2021-04-26         Central African Republic  CAF           1             0
#> 159 2021-04-26                     Saudi Arabia  SAU           1             0
#> 62  2021-05-01                         Ethiopia  ETH           1             0
#> 63  2021-05-04                    Faroe Islands  FRO           1             1
#> 47  2021-05-05                             Cuba  CUB          19             0
#> 94  2021-05-05                          Jamaica  JAM           3             3
#> 97  2021-05-05                       Kazakhstan  KAZ           1             0
#> 89  2021-05-08                             Iran  IRN           2             2
#> 144 2021-05-10                      Philippines  PHL           4             3
#> 51  2021-05-11                    Côte d'Ivoire  CIV           1             0
#> 59  2021-05-15                      El Salvador  SLV           4             0
#> 90  2021-05-15                             Iraq  IRQ           1             0
#> 54  2021-05-20                         Djibouti  DJI           1             0
#> 82  2021-05-20                            Haiti  HTI          13             0
#> 18  2021-05-24                          Belarus  BLR           4             2
#> 83  2021-05-27                         Honduras  HND           6             0
#> 102 2021-05-28                           Latvia  LVA           2             2
#> 183 2021-05-28                          Tunisia  TUN           2             0
#> 196 2021-05-29                           Zambia  ZMB           4             0
#> 191 2021-05-30                          Uruguay  URY           1             0
#> 32  2021-05-31                          Burundi  BDI           3             0
#> 23  2021-06-02                          Bolivia  BOL           4             0
#> 154 2021-06-06            Saint Kitts and Nevis  KNA           3             0
#> 106 2021-06-07                            Libya  LBY           2             0
#> 129 2021-06-15                          Namibia  NAM           7             0
#> 81  2021-06-21                           Guyana  GUY           2             0
#> 188 2021-06-23             United Arab Emirates  ARE           2             0
#> 79  2021-06-25                           Guinea  GIN           1             0
#> 56  2021-06-29               Dominican Republic  DOM           3             0
#> 45  2021-07-01                           Crimea None          21             0
#> 121 2021-07-06                          Moldova  MDA           1             0
#> 161 2021-07-06                           Serbia  SRB           1             0
#> 193 2021-07-07                        Venezuela  VEN           1             0
#> 125 2021-07-08                       Montserrat  MSR           1             0
#> 126 2021-07-08                          Morocco  MAR           1             1
#> 142 2021-07-08                         Paraguay  PRY           1             0
#> 153 2021-07-09                           Rwanda  RWA           1             0
#> 5   2021-07-10                           Angola  AGO           4             0
#> 105 2021-07-10                          Liberia  LBR           1             0
#> 116 2021-07-10                            Malta  MLT           8             0
#> 52  2021-07-12 Democratic Republic of the Congo  COD           7             0
#> 69  2021-07-12                           Gambia  GMB           3             0
#> 130 2021-07-12                            Nepal  NPL           1             0
#> 185 2021-07-12         Turks and Caicos Islands  TCA           4             0
#> 2   2021-07-13                          Albania  ALB           2             1
#> 77  2021-07-13                             Guam  GUM           1             1
#> 187 2021-07-13                          Ukraine  UKR           2             0
#> 13  2021-07-14                       Azerbaijan  AZE           1             0
#> 128 2021-07-15                          Myanmar  MMR          13             0
#> 149 2021-07-15                Republic of Congo  COG           1             0
#> 103 2021-07-16                          Lebanon  LBN           1             0
#> 127 2021-07-16                       Mozambique  MOZ           1             0
#> 6   2021-07-18                         Anguilla  AIA           2             0
#> 186 2021-07-18                           Uganda  UGA           1             0
#> 4   2021-07-19                          Andorra  AND           1             0
#> 41  2021-07-19                            China  CHN           1             0
#> 61  2021-07-21                          Estonia  EST          24             1
#> 156 2021-07-21 Saint Vincent and the Grenadines  VCT           1             0
#> 122 2021-07-23                           Monaco  MCO           3             0
#> 192 2021-07-23                       Uzbekistan  UZB          17             0
#> 170 2021-07-24                      South Sudan  SSD           5             0
#>     cum_total_count cum_lineage_count  proportion proportion_ci_lower
#> 157               2                 0 0.000000000        2.169085e-04
#> 158              10                 0 0.000000000        4.789043e-05
#> 180              18                 0 0.000000000        2.690312e-05
#> 43                6                 0 0.000000000        7.849985e-05
#> 22               40                 2 0.050000000        1.053449e-02
#> 67               11                 0 0.000000000        4.363580e-05
#> 104              18                 0 0.000000000        2.690312e-05
#> 20               52                 1 0.019230769        2.082849e-03
#> 1                16                 0 0.000000000        3.021348e-05
#> 162              34                 0 0.000000000        1.433629e-05
#> 80               48                31 0.645833333        5.051968e-01
#> 33               39                 4 0.102564103        3.562299e-02
#> 31              256                 3 0.011718750        3.311255e-03
#> 85              435                29 0.066666667        4.605404e-02
#> 140             885                 0 0.000000000        5.546844e-07
#> 111             122                 0 0.000000000        4.016626e-06
#> 21               65                15 0.230769231        1.414856e-01
#> 167              30                 7 0.233333333        1.109209e-01
#> 123              27                 0 0.000000000        1.801870e-05
#> 37              358               110 0.307262570        2.611686e-01
#> 175              31                 0 0.000000000        1.571248e-05
#> 55                9                 4 0.444444444        1.729678e-01
#> 60              190                 1 0.005263158        5.684697e-04
#> 133              24                 1 0.041666667        4.533190e-03
#> 115              36                 0 0.000000000        1.354537e-05
#> 139             132                22 0.166666667        1.106315e-01
#> 14               61                26 0.426229508        3.079238e-01
#> 35              197                11 0.055837563        3.002529e-02
#> 195              10                10 1.000000000        7.828037e-01
#> 39               56                10 0.178571429        9.568161e-02
#> 159            1096                 4 0.003649635        1.233140e-03
#> 62               25                 3 0.120000000        3.498475e-02
#> 63               42                 2 0.047619048        1.002614e-02
#> 47              649                 0 0.000000000        7.563101e-07
#> 94              167               151 0.904191617        8.526240e-01
#> 97              355               162 0.456338028        4.050544e-01
#> 89              547                52 0.095063985        7.261821e-02
#> 144            5265               983 0.186704653        1.763527e-01
#> 51              236                33 0.139830508        1.000616e-01
#> 59              104                 3 0.028846154        8.189920e-03
#> 90              169                48 0.284023669        2.201225e-01
#> 54              305                79 0.259016393        2.122874e-01
#> 82               79                 1 0.012658228        1.369203e-03
#> 18               90                19 0.211111111        1.367321e-01
#> 83               72                 1 0.013888889        1.502686e-03
#> 102            5507              3123 0.567096423        5.539778e-01
#> 183             129                 6 0.046511628        1.964398e-02
#> 196             618                 2 0.003236246        6.730922e-04
#> 191             720                 0 0.000000000        6.817554e-07
#> 32                9                 1 0.111111111        1.226409e-02
#> 23               66                 0 0.000000000        7.411762e-06
#> 154              13                 0 0.000000000        3.705193e-05
#> 106              22                 1 0.045454545        4.949094e-03
#> 129             230                 3 0.013043478        3.686929e-03
#> 81               14                 0 0.000000000        3.445267e-05
#> 188            1857                19 0.010231556        6.380450e-03
#> 79              180                13 0.072222222        4.108569e-02
#> 56              281                16 0.056939502        3.425961e-02
#> 45               32                 0 0.000000000        1.522530e-05
#> 121              67                37 0.552238806        4.330680e-01
#> 161             293                45 0.153583618        1.157867e-01
#> 193             169                 6 0.035502959        1.495225e-02
#> 125               3                 2 0.666666667        1.767361e-01
#> 126             314                91 0.289808917        2.417271e-01
#> 142             224                 4 0.017857143        6.056823e-03
#> 153             483                 6 0.012422360        5.201211e-03
#> 5               860               139 0.161627907        1.381810e-01
#> 105              44                 0 0.000000000        1.109659e-05
#> 116             253               111 0.438735178        3.785737e-01
#> 52              649                 8 0.012326656        5.844352e-03
#> 69              609                58 0.095238095        7.384084e-02
#> 130             168                 3 0.017857143        5.054359e-03
#> 185              16                 5 0.312500000        1.305817e-01
#> 2                42                28 0.666666667        5.167779e-01
#> 77              194                95 0.489690722        4.199357e-01
#> 187             249                70 0.281124498        2.280502e-01
#> 13               14                 3 0.214285714        6.433508e-02
#> 128              74                 2 0.027027027        5.657991e-03
#> 149             221                27 0.122171946        8.396952e-02
#> 103            1019               791 0.776251227        7.499027e-01
#> 127             573                 1 0.001745201        1.883675e-04
#> 6                 8                 2 0.250000000        5.596689e-02
#> 186             626                17 0.027156550        1.651165e-02
#> 4                22                 7 0.318181818        1.550801e-01
#> 41              963                18 0.018691589        1.151618e-02
#> 61             4773              3166 0.663314477        6.498119e-01
#> 156              12                 0 0.000000000        4.007531e-05
#> 122              56                 3 0.053571429        1.531695e-02
#> 192              87                 2 0.022988506        4.807185e-03
#> 170              87                 1 0.011494253        1.243013e-03
#>     proportion_ci_upper query_key
#> 157         0.666821754   B.1.1.7
#> 158         0.217196268   B.1.1.7
#> 180         0.128558212   B.1.1.7
#> 43          0.330388909   B.1.1.7
#> 22          0.150865051   B.1.1.7
#> 67          0.199994569   B.1.1.7
#> 104         0.128558212   B.1.1.7
#> 20          0.086368335   B.1.1.7
#> 1           0.143184138   B.1.1.7
#> 162         0.070711631   B.1.1.7
#> 80          0.769468931   B.1.1.7
#> 33          0.225658999   B.1.1.7
#> 31          0.030940262   B.1.1.7
#> 85          0.093016879   B.1.1.7
#> 140         0.002833530   B.1.1.7
#> 111         0.020337816   B.1.1.7
#> 21          0.343414898   B.1.1.7
#> 167         0.404399913   B.1.1.7
#> 123         0.088050832   B.1.1.7
#> 37          0.356458627   B.1.1.7
#> 175         0.077230036   B.1.1.7
#> 55          0.745906132   B.1.1.7
#> 60          0.024332646   B.1.1.7
#> 133         0.178694919   B.1.1.7
#> 115         0.066944482   B.1.1.7
#> 139         0.237125549   B.1.1.7
#> 14          0.551352780   B.1.1.7
#> 35          0.094548747   B.1.1.7
#> 195         0.999952110   B.1.1.7
#> 39          0.293669242   B.1.1.7
#> 159         0.008654497   B.1.1.7
#> 62          0.286727541   B.1.1.7
#> 63          0.144100362   B.1.1.7
#> 47          0.003861516   B.1.1.7
#> 94          0.941913021   B.1.1.7
#> 97          0.508330402   B.1.1.7
#> 89          0.121779546   B.1.1.7
#> 144         0.197400934   B.1.1.7
#> 51          0.188365906   B.1.1.7
#> 59          0.074971389   B.1.1.7
#> 90          0.355250307   B.1.1.7
#> 54          0.310294062   B.1.1.7
#> 82          0.057629127   B.1.1.7
#> 18          0.303721845   B.1.1.7
#> 83          0.063070801   B.1.1.7
#> 102         0.580144553   B.1.1.7
#> 183         0.093342744   B.1.1.7
#> 196         0.010341075   B.1.1.7
#> 191         0.003481524   B.1.1.7
#> 32          0.414469277   B.1.1.7
#> 23          0.037205610   B.1.1.7
#> 154         0.172619445   B.1.1.7
#> 106         0.193437365   B.1.1.7
#> 129         0.034395983   B.1.1.7
#> 81          0.161553305   B.1.1.7
#> 188         0.015604398   B.1.1.7
#> 79          0.116954226   B.1.1.7
#> 56          0.088672100   B.1.1.7
#> 45          0.074927763   B.1.1.7
#> 121         0.667003719   B.1.1.7
#> 161         0.198182815   B.1.1.7
#> 193         0.071704393   B.1.1.7
#> 125         0.961252382   B.1.1.7
#> 126         0.341745226   B.1.1.7
#> 142         0.041894877   B.1.1.7
#> 153         0.025424562   B.1.1.7
#> 5           0.187347723   B.1.1.7
#> 105         0.055183597   B.1.1.7
#> 116         0.500289494   B.1.1.7
#> 52          0.023123700   B.1.1.7
#> 69          0.120470218   B.1.1.7
#> 130         0.046881973   B.1.1.7
#> 185         0.555704465   B.1.1.7
#> 2           0.794459865   B.1.1.7
#> 77          0.559750807   B.1.1.7
#> 187         0.339253319   B.1.1.7
#> 13          0.469023667   B.1.1.7
#> 128         0.083876400   B.1.1.7
#> 149         0.170187612   B.1.1.7
#> 103         0.801033045   B.1.1.7
#> 127         0.008127774   B.1.1.7
#> 6           0.591623511   B.1.1.7
#> 186         0.042151542   B.1.1.7
#> 4           0.525709375   B.1.1.7
#> 41          0.028748224   B.1.1.7
#> 61          0.676619073   B.1.1.7
#> 156         0.185305938   B.1.1.7
#> 122         0.136100539   B.1.1.7
#> 192         0.071697393   B.1.1.7
#> 170         0.052456430   B.1.1.7
#>  [ reached 'max' / getOption("max.print") -- omitted 107 rows ]

# County-level prevalence of B.1.1.7
getCumulativeBySubadmin(pangolin_lineage="B.1.1.7", location="California")

#> Retrieving data... 

#>          date                             name              id total_count
#> 35 2020-10-29  Napa Solano Yolo Marin Counties  USA_US-CA_None           6
#> 42 2020-12-23                          Redding  USA_US-CA_None           2
#> 2  2020-12-25                       Alexandria  USA_US-CA_None           1
#> 37 2020-12-28                         New York  USA_US-CA_None           1
#> 69 2021-01-08                            UNKNO  USA_US-CA_None           1
#> 30 2021-01-20                          Modesto  USA_US-CA_None           1
#> 58 2021-02-02                           Sierra USA_US-CA_06091           1
#> 75 2021-02-19                      unavailable  USA_US-CA_None           1
#> 31 2021-03-09                            Modoc USA_US-CA_06049           1
#> 11 2021-03-16                          Douglas  USA_US-CA_None           1
#> 66 2021-04-10                          Trinity USA_US-CA_06105           2
#> 72 2021-04-20                           Washoe  USA_US-CA_None           1
#> 6  2021-04-28                Californialaveras  USA_US-CA_None           1
#> 25 2021-04-29                         Maricopa  USA_US-CA_None           1
#> 17 2021-04-30                             Inyo USA_US-CA_06027           1
#> 53 2021-05-02                       San Mateo*  USA_US-CA_None           1
#> 7  2021-05-07                           Colusa USA_US-CA_06011           1
#> 32 2021-05-19                             Mono USA_US-CA_06051           1
#> 27 2021-05-28                         Mariposa USA_US-CA_06043           1
#> 41 2021-06-02                           Plumas USA_US-CA_06063           1
#> 21 2021-06-04                           Lassen USA_US-CA_06035           3
#> 10 2021-06-07                        Del Norte USA_US-CA_06015           1
#> 20 2021-06-07                             Lake USA_US-CA_06033           5
#> 22 2021-06-08                  Long Beach City  USA_US-CA_None           1
#> 19 2021-06-09                            Kings USA_US-CA_06031           1
#> 28 2021-06-09                        Mendocino USA_US-CA_06045           1
#> 65 2021-06-09                           Tehama USA_US-CA_06103           2
#> 54 2021-06-10                    Santa Barbara USA_US-CA_06083           6
#> 57 2021-06-10                           Shasta USA_US-CA_06089           1
#> 4  2021-06-11                            Butte USA_US-CA_06007           2
#> 18 2021-06-11                             Kern USA_US-CA_06029           1
#> 36 2021-06-11                           Nevada USA_US-CA_06057           2
#> 68 2021-06-11                         Tuolumne USA_US-CA_06109           1
#> 71 2021-06-11                          Ventura USA_US-CA_06111           1
#> 14 2021-06-12                            Glenn USA_US-CA_06021           1
#> 59 2021-06-12                         Siskiyou USA_US-CA_06093           1
#> 44 2021-06-21                        SANTA CLA  USA_US-CA_None           2
#> 3  2021-07-06                           Amador USA_US-CA_06005           1
#> 5  2021-07-06                        Calaveras USA_US-CA_06009           1
#> 64 2021-07-08                           Sutter USA_US-CA_06101           1
#> 24 2021-07-12                           Madera USA_US-CA_06039           2
#> 34 2021-07-12                             Napa USA_US-CA_06055           1
#> 74 2021-07-12                             Yuba USA_US-CA_06115           1
#> 61 2021-07-13                           Sonoma USA_US-CA_06097           1
#> 9  2021-07-14                            Davis  USA_US-CA_None           4
#> 12 2021-07-15                        El Dorado USA_US-CA_06017           1
#> 13 2021-07-15                           Fresno USA_US-CA_06019           3
#> 29 2021-07-15                           Merced USA_US-CA_06047           1
#> 60 2021-07-15                           Solano USA_US-CA_06095           1
#> 63 2021-07-15                       Stanislaus USA_US-CA_06099           3
#> 73 2021-07-15                             Yolo USA_US-CA_06113           2
#> 45 2021-07-16                       Sacramento USA_US-CA_06067           2
#> 52 2021-07-16                        San Mateo USA_US-CA_06081           1
#> 8  2021-07-22                     Contra Costa USA_US-CA_06013           1
#> 46 2021-07-29                       San Benito USA_US-CA_06069           1
#> 26 2021-08-02                            Marin USA_US-CA_06041           1
#> 40 2021-08-02                           Placer USA_US-CA_06061           1
#> 67 2021-08-03                           Tulare USA_US-CA_06107           1
#> 15 2021-08-05                         Humboldt USA_US-CA_06023           1
#> 51 2021-08-05                  San Luis Obispo USA_US-CA_06079           6
#> 47 2021-08-11                   San Bernardino USA_US-CA_06071           1
#> 50 2021-08-13                      San Joaquin USA_US-CA_06077           1
#> 62 2021-08-15      Southern San Joaquin Valley  USA_US-CA_None          17
#> 16 2021-08-16                         Imperial USA_US-CA_06025           1
#> 43 2021-08-16                        Riverside USA_US-CA_06065           1
#> 38 2021-08-17 North America / USA / California  USA_US-CA_None         494
#> 49 2021-08-17                    San Francisco USA_US-CA_06075          14
#> 23 2021-08-22                      Los Angeles USA_US-CA_06037           1
#> 33 2021-08-23                         Monterey USA_US-CA_06053           1
#> 48 2021-08-23                        San Diego USA_US-CA_06073           1
#> 70 2021-08-23                          Unknown  USA_US-CA_None           1
#> 1  2021-08-25                          Alameda USA_US-CA_06001           1
#> 56 2021-08-25                       Santa Cruz USA_US-CA_06087          18
#> 39 2021-08-26                           Orange USA_US-CA_06059           1
#> 55 2021-08-26                      Santa Clara USA_US-CA_06085           7
#>    lineage_count cum_total_count cum_lineage_count proportion
#> 35             0              27                 0 0.00000000
#> 42             0               2                 0 0.00000000
#> 2              0               1                 0 0.00000000
#> 37             0               1                 0 0.00000000
#> 69             0               1                 0 0.00000000
#> 30             0               1                 0 0.00000000
#> 58             0               1                 0 0.00000000
#> 75             0               4                 0 0.00000000
#> 31             0               4                 0 0.00000000
#> 11             0               2                 1 0.50000000
#> 66             2               3                 2 0.66666667
#> 72             0               6                 0 0.00000000
#> 6              0               6                 0 0.00000000
#> 25             0               1                 0 0.00000000
#> 17             1              10                 1 0.10000000
#> 53             1               1                 1 1.00000000
#> 7              0              14                 2 0.14285714
#> 32             1             132                 7 0.05303030
#> 27             1              39                 7 0.17948718
#> 41             0               6                 1 0.16666667
#> 21             0              11                 1 0.09090909
#> 10             0              66                15 0.22727273
#> 20             2             225                15 0.06666667
#> 22             0              75                 2 0.02666667
#> 19             0             592                59 0.09966216
#> 28             1             325                51 0.15692308
#> 65             0             180                13 0.07222222
#> 54             0            1098               149 0.13570128
#> 57             1             277                51 0.18411552
#> 4              0             325                89 0.27384615
#> 18             1             520                30 0.05769231
#> 36             0             317                75 0.23659306
#> 68             1             101                 9 0.08910891
#> 71             0             398                70 0.17587940
#> 14             1             119                10 0.08403361
#> 59             0              28                 8 0.28571429
#> 44             0              60                17 0.28333333
#> 3              0              54                 8 0.14814815
#> 5              0              80                12 0.15000000
#> 64             0             199                 9 0.04522613
#> 24             0             704                39 0.05539773
#> 34             0             193                 8 0.04145078
#> 74             0             113                 6 0.05309735
#> 61             0             569                33 0.05799649
#> 9              0             224                34 0.15178571
#> 12             1             214                58 0.27102804
#> 13             0            1785               209 0.11708683
#> 29             0             654                45 0.06880734
#> 60             0             491               101 0.20570265
#> 63             0             762                55 0.07217848
#> 73             0             137                10 0.07299270
#> 45             0            1302               243 0.18663594
#> 52             0             324                30 0.09259259
#> 8              0            2357               224 0.09503606
#> 46             0             225                 6 0.02666667
#> 26             0             399                11 0.02756892
#> 40             0             348                64 0.18390805
#> 67             0             765                31 0.04052288
#> 15             0            1581               221 0.13978495
#> 51             0            1434               217 0.15132497
#> 47             0            1008                69 0.06845238
#> 50             0             696                78 0.11206897
#> 62             0             681                96 0.14096916
#> 16             0             593                43 0.07251265
#> 43             0            2077                91 0.04381319
#> 38             0           41461              4704 0.11345602
#> 49             0            2343                68 0.02902262
#> 23             0            6496               597 0.09190271
#> 33             0            1205                44 0.03651452
#> 48             0           19014              1302 0.06847586
#> 70             0           31072              3589 0.11550592
#> 1              0            4438               283 0.06376746
#> 56             0            1596                32 0.02005013
#> 39             0            2507                46 0.01834862
#> 55             0            6714               327 0.04870420
#>    proportion_ci_lower proportion_ci_upper query_key
#> 35        1.801870e-05          0.08805083   B.1.1.7
#> 42        2.169085e-04          0.66682175   B.1.1.7
#> 2         3.855810e-04          0.85325368   B.1.1.7
#> 37        3.855810e-04          0.85325368   B.1.1.7
#> 69        3.855810e-04          0.85325368   B.1.1.7
#> 30        3.855810e-04          0.85325368   B.1.1.7
#> 58        3.855810e-04          0.85325368   B.1.1.7
#> 75        1.153328e-04          0.44476262   B.1.1.7
#> 31        1.153328e-04          0.44476262   B.1.1.7
#> 11        6.083028e-02          0.93916972   B.1.1.7
#> 66        1.767361e-01          0.96125238   B.1.1.7
#> 72        7.849985e-05          0.33038891   B.1.1.7
#> 6         7.849985e-05          0.33038891   B.1.1.7
#> 25        3.855810e-04          0.85325368   B.1.1.7
#> 17        1.101167e-02          0.38131477   B.1.1.7
#> 53        1.467463e-01          0.99961442   B.1.1.7
#> 7         3.092213e-02          0.38489432   B.1.1.7
#> 32        2.403017e-02          0.10128549   B.1.1.7
#> 27        8.405528e-02          0.32041900   B.1.1.7
#> 41        1.862013e-02          0.55805677   B.1.1.7
#> 21        9.991460e-03          0.35297580   B.1.1.7
#> 10        1.392333e-01          0.33862720   B.1.1.7
#> 20        3.948495e-02          0.10488938   B.1.1.7
#> 22        5.581996e-03          0.08279463   B.1.1.7
#> 19        7.748514e-02          0.12574083   B.1.1.7
#> 28        1.204940e-01          0.19942927   B.1.1.7
#> 65        4.108569e-02          0.11695423   B.1.1.7
#> 54        1.164119e-01          0.15690799   B.1.1.7
#> 57        1.418809e-01          0.23291002   B.1.1.7
#> 4         2.275237e-01          0.32417611   B.1.1.7
#> 18        4.005974e-02          0.08022698   B.1.1.7
#> 36        1.923425e-01          0.28562803   B.1.1.7
#> 68        4.501016e-02          0.15629529   B.1.1.7
#> 71        1.408921e-01          0.21556006   B.1.1.7
#> 14        4.402591e-02          0.14393286   B.1.1.7
#> 59        1.453633e-01          0.46783688   B.1.1.7
#> 44        1.814556e-01          0.40554610   B.1.1.7
#> 3         7.261794e-02          0.26016415   B.1.1.7
#> 5         8.469949e-02          0.24006484   B.1.1.7
#> 64        2.261020e-02          0.08090527   B.1.1.7
#> 24        4.029186e-02          0.07414727   B.1.1.7
#> 34        1.978854e-02          0.07668240   B.1.1.7
#> 74        2.246356e-02          0.10615389   B.1.1.7
#> 61        4.099304e-02          0.07947853   B.1.1.7
#> 9         1.093854e-01          0.20311295   B.1.1.7
#> 12        2.148351e-01          0.33336662   B.1.1.7
#> 13        1.027938e-01          0.13262028   B.1.1.7
#> 29        5.128955e-02          0.09012906   B.1.1.7
#> 60        1.717408e-01          0.24312201   B.1.1.7
#> 63        5.541120e-02          0.09218667   B.1.1.7
#> 73        3.814483e-02          0.12560626   B.1.1.7
#> 45        1.661812e-01          0.20848205   B.1.1.7
#> 52        6.464051e-02          0.12777897   B.1.1.7
#> 8         8.369343e-02          0.10737242   B.1.1.7
#> 46        1.120555e-02          0.05413297   B.1.1.7
#> 26        1.473375e-02          0.04720385   B.1.1.7
#> 40        1.459113e-01          0.22713637   B.1.1.7
#> 67        2.824014e-02          0.05627044   B.1.1.7
#> 15        1.233562e-01          0.15753102   B.1.1.7
#> 51        1.334874e-01          0.17056824   B.1.1.7
#> 47        5.408936e-02          0.08528832   B.1.1.7
#> 50        9.024986e-02          0.13710569   B.1.1.7
#> 62        1.163748e-01          0.16860724   B.1.1.7
#> 16        5.369930e-02          0.09548401   B.1.1.7
#> 43        3.563712e-02          0.05325910   B.1.1.7
#> 38        1.104303e-01          0.11653574   B.1.1.7
#> 49        2.279506e-02          0.03641224   B.1.1.7
#> 23        8.505924e-02          0.09910972   B.1.1.7
#> 33        2.701067e-02          0.04823978   B.1.1.7
#> 48        6.495155e-02          0.07213153   B.1.1.7
#> 70        1.119878e-01          0.11909567   B.1.1.7
#> 1         5.686180e-02          0.07124183   B.1.1.7
#> 56        1.401393e-02          0.02782292   B.1.1.7
#> 39        1.363527e-02          0.02417232   B.1.1.7
#> 55        4.374871e-02          0.05404865   B.1.1.7