Retrieve daily prevalence of a lineage by location

Retrieves the daily prevalence of a PANGO lineage(s) by location

getPrevalenceByLocation(
  pangolin_lineage = NULL,
  location = NULL,
  mutations = NULL,
  cumulative = NULL
)

Arguments

pangolin_lineage:
(optional) PANGO lineage name or vector of PANGO lineage names. Either `pangolin_lineage` or `mutations` needs to be specified. A list of lineages will return a long dataframe with `lineage` as a variable; if you want to calculate the prevalence of lineage1 or lineage2 together, enter the lineages separated by " OR ". For instance, to calculate the prevalence of Delta, you'll need to supply `"B.1.617.2 OR AY.1 OR AY.2 OR ..."` **Be sure to include the space around "OR" and it must be capitalized.**

mutations:
(optional) a `vector` of mutation(s). Either `pangolin_lineage` or `mutations` needs to be specified.

location:
(optional) a location name

cumulative:
(optional) `Boolean` (T/F), T returns cumulative prevalence since first day of detection

Value

dataframe

Examples

# lineage: P.1 in Brazil
getPrevalenceByLocation(pangolin_lineage = "P.1", location = "Brazil") %>% head()

#> Retrieving data... 
#> Retrieving data... 

#>         date total_count lineage_count total_count_rolling
#> 1 2020-09-11          17             1            14.66667
#> 2 2020-09-12           8             0            14.66667
#> 3 2020-09-13           2             0            12.85714
#> 4 2020-09-14          17             0            14.71429
#> 5 2020-09-15          35             1            16.28571
#> 6 2020-09-16          18             0            16.42857
#>   lineage_count_rolling  proportion proportion_ci_lower proportion_ci_upper
#> 1             0.1666667 0.011363636        3.219414e-05           0.1518167
#> 2             0.1666667 0.011363636        3.219414e-05           0.1518167
#> 3             0.1428571 0.011111111        3.705193e-05           0.1726194
#> 4             0.1428571 0.009708738        3.219414e-05           0.1518167
#> 5             0.2857143 0.017543860        3.021348e-05           0.1431841
#> 6             0.2857143 0.017391304        3.021348e-05           0.1431841
#>   lineage
#> 1     P.1
#> 2     P.1
#> 3     P.1
#> 4     P.1
#> 5     P.1
#> 6     P.1

# AY.4 and B.1.617.2 in Brazil (AY.4, AY.34, and B.1.617.2 separately)
getPrevalenceByLocation(pangolin_lineage = c("AY.4", "AY.34", "B.1.617.2"), location = "Brazil") %>% filter(date == "2021-09-01")

#> Retrieving data... 
#> Retrieving data... 
#> Retrieving data... 
#> Retrieving data... 
#> Retrieving data... 
#> Retrieving data... 

#>         date total_count lineage_count total_count_rolling
#> 1 2021-09-01          36             1            431.2857
#> 2 2021-09-01          36            19            431.2857
#>   lineage_count_rolling proportion proportion_ci_lower proportion_ci_upper
#> 1              10.28571 0.02384896          0.01199003          0.04076588
#> 2             255.57143 0.59258032          0.54710319          0.63957454
#>     lineage
#> 1      AY.4
#> 2 B.1.617.2

# AY.4 OR B.1.617.2 in Brazil (AY.4, AY.34 and B.1.617.2 combined)
getPrevalenceByLocation(pangolin_lineage = "AY.4 OR AY.34 OR B.1.617.2", location = "Brazil") %>% filter(date == "2021-09-01")

#> Retrieving data... 
#> Retrieving data... 

#>         date total_count lineage_count total_count_rolling
#> 1 2021-09-01          28            16                  33
#>   lineage_count_rolling proportion proportion_ci_lower proportion_ci_upper
#> 1              13.33333  0.4040404           0.2418367           0.5637582
#>                      lineage
#> 1 AY.4 OR AY.34 OR B.1.617.2

# S:E484K mutation prevalence worldwide
getPrevalenceByLocation(mutations = c("S:E484K")) %>% head()

#> Retrieving data... 

#>         date total_count lineage_count total_count_rolling
#> 1 2020-02-01          50             1            48.71429
#> 2 2020-02-02          48             0            48.14286
#> 3 2020-02-03          41             0            48.71429
#> 4 2020-02-04          34             0            48.42857
#> 5 2020-02-05          45             0            45.42857
#> 6 2020-02-06          33             0            41.28571
#>   lineage_count_rolling  proportion proportion_ci_lower proportion_ci_upper
#> 1             0.1428571 0.002932551        9.970067e-06          0.04972339
#> 2             0.1428571 0.002967359        1.017669e-05          0.05072726
#> 3             0.1428571 0.002932551        9.970067e-06          0.04972339
#> 4             0.1428571 0.002949853        1.017669e-05          0.05072726
#> 5             0.1428571 0.003144654        1.085137e-05          0.05399771
#> 6             0.1428571 0.003460208        1.190358e-05          0.05907571
#>   lineage
#> 1 S:E484K
#> 2 S:E484K
#> 3 S:E484K
#> 4 S:E484K
#> 5 S:E484K
#> 6 S:E484K

# B.1.1.7 + S:E484K mutation worldwide
getPrevalenceByLocation(pangolin_lineage = "B.1.1.7", mutations = c("S:E484K")) %>% head()

#> Retrieving data... 

#>         date total_count lineage_count total_count_rolling
#> 1 2020-12-17        4559             1            4817.286
#> 2 2020-12-18        4521             1            4887.429
#> 3 2020-12-19        2920             0            4957.429
#> 4 2020-12-20        2858             0            5040.429
#> 5 2020-12-21        8632             0            4856.857
#> 6 2020-12-22        4020             0            4533.000
#>   lineage_count_rolling   proportion proportion_ci_lower proportion_ci_upper
#> 1             0.1428571 2.965511e-05        1.019325e-07        0.0005213116
#> 2             0.2857143 5.845902e-05        1.004726e-07        0.0005138468
#> 3             0.2857143 5.763357e-05        9.905382e-08        0.0005065927
#> 4             0.2857143 5.668452e-05        9.742265e-08        0.0004982525
#> 5             0.2857143 5.882699e-05        1.010931e-07        0.0005170196
#> 6             0.2857143 6.302984e-05        1.083184e-07        0.0005539618
#>                   lineage
#> 1 (B.1.1.7) AND (S:E484K)
#> 2 (B.1.1.7) AND (S:E484K)
#> 3 (B.1.1.7) AND (S:E484K)
#> 4 (B.1.1.7) AND (S:E484K)
#> 5 (B.1.1.7) AND (S:E484K)
#> 6 (B.1.1.7) AND (S:E484K)