Retrieve sequence count data

Retrieves number of sequences by location, by date of sample collection. See getSubmissionDateByLocation to view similar information by date of sequence submission.

getSeqCounts(location = NULL, cumulative = NULL, subadmin = NULL)

Arguments

location

(optional) a location name (if not specified, global total counts returned)

cumulative

(optional) Boolean (T/F), T returns cumulative number of sequences to date

subadmin

(optional) Boolean (T/F), subadmin=T and cumulative=T returns cumulative number of sequences for next administrative level

Value

dataframe

Examples

# Retrieves the number of samples sequenced in the U.S. each day, by date of sample collection
getSeqCounts(location="United States")
#> Retrieving data... 
#>           date total_count
#> 1   2020-01-26           3
#> 2   2020-01-27           1
#> 3   2020-01-28           2
#> 4   2020-01-29           4
#> 5   2020-01-31           1
#> 6   2020-02-02           1
#> 7   2020-02-06           1
#> 8   2020-02-08           1
#> 9   2020-02-10           1
#> 10  2020-02-11           1
#> 11  2020-02-17           9
#> 12  2020-02-18           6
#> 13  2020-02-20           2
#> 14  2020-02-21           9
#> 15  2020-02-22           3
#> 16  2020-02-23           1
#> 17  2020-02-24           8
#> 18  2020-02-25           5
#> 19  2020-02-26           8
#> 20  2020-02-27           9
#> 21  2020-02-28          27
#> 22  2020-02-29          37
#> 23  2020-03-01          31
#> 24  2020-03-02          66
#> 25  2020-03-03          52
#> 26  2020-03-04          79
#> 27  2020-03-05         120
#> 28  2020-03-06          87
#> 29  2020-03-07          96
#> 30  2020-03-08         109
#> 31  2020-03-09         197
#> 32  2020-03-10         196
#> 33  2020-03-11         244
#> 34  2020-03-12         389
#> 35  2020-03-13         553
#> 36  2020-03-14         539
#> 37  2020-03-15        1008
#> 38  2020-03-16         953
#> 39  2020-03-17         795
#> 40  2020-03-18         859
#> 41  2020-03-19         778
#> 42  2020-03-20         645
#> 43  2020-03-21         497
#> 44  2020-03-22         403
#> 45  2020-03-23         658
#> 46  2020-03-24         767
#> 47  2020-03-25         795
#> 48  2020-03-26         762
#> 49  2020-03-27         747
#> 50  2020-03-28         512
#> 51  2020-03-29         383
#> 52  2020-03-30         809
#> 53  2020-03-31         712
#> 54  2020-04-01         885
#> 55  2020-04-02         769
#> 56  2020-04-03         784
#> 57  2020-04-04         485
#> 58  2020-04-05         392
#> 59  2020-04-06         724
#> 60  2020-04-07         578
#> 61  2020-04-08         683
#> 62  2020-04-09         560
#> 63  2020-04-10         567
#> 64  2020-04-11         406
#> 65  2020-04-12         256
#> 66  2020-04-13         710
#> 67  2020-04-14         477
#> 68  2020-04-15         857
#> 69  2020-04-16         438
#> 70  2020-04-17         467
#> 71  2020-04-18         332
#> 72  2020-04-19         220
#> 73  2020-04-20         548
#> 74  2020-04-21         578
#> 75  2020-04-22         452
#> 76  2020-04-23         452
#> 77  2020-04-24         466
#> 78  2020-04-25         255
#> 79  2020-04-26         189
#> 80  2020-04-27         557
#> 81  2020-04-28         530
#> 82  2020-04-29         618
#> 83  2020-04-30         666
#> 84  2020-05-01         578
#> 85  2020-05-02         310
#> 86  2020-05-03         192
#> 87  2020-05-04         593
#> 88  2020-05-05         561
#> 89  2020-05-06         429
#> 90  2020-05-07         509
#> 91  2020-05-08         451
#> 92  2020-05-09         292
#> 93  2020-05-10         176
#> 94  2020-05-11         484
#> 95  2020-05-12         461
#> 96  2020-05-13         508
#> 97  2020-05-14         625
#> 98  2020-05-15         503
#> 99  2020-05-16         187
#> 100 2020-05-17         154
#> 101 2020-05-18         373
#> 102 2020-05-19         412
#> 103 2020-05-20         390
#> 104 2020-05-21         464
#> 105 2020-05-22         462
#> 106 2020-05-23         181
#> 107 2020-05-24         175
#> 108 2020-05-25         178
#> 109 2020-05-26         496
#> 110 2020-05-27         480
#> 111 2020-05-28         371
#> 112 2020-05-29         471
#> 113 2020-05-30         228
#> 114 2020-05-31         189
#> 115 2020-06-01         446
#> 116 2020-06-02         380
#> 117 2020-06-03         484
#> 118 2020-06-04         456
#> 119 2020-06-05         442
#> 120 2020-06-06         190
#> 121 2020-06-07         149
#> 122 2020-06-08         474
#> 123 2020-06-09         469
#> 124 2020-06-10         447
#> 125 2020-06-11         433
#> 126 2020-06-12         370
#> 127 2020-06-13         385
#> 128 2020-06-14         244
#> 129 2020-06-15         708
#> 130 2020-06-16         684
#> 131 2020-06-17         488
#> 132 2020-06-18         735
#> 133 2020-06-19         664
#> 134 2020-06-20         391
#> 135 2020-06-21         442
#> 136 2020-06-22         779
#> 137 2020-06-23         653
#> 138 2020-06-24         776
#> 139 2020-06-25         718
#> 140 2020-06-26         695
#> 141 2020-06-27         350
#> 142 2020-06-28         362
#> 143 2020-06-29         839
#> 144 2020-06-30         735
#> 145 2020-07-01         856
#> 146 2020-07-02         870
#> 147 2020-07-03         558
#> 148 2020-07-04         336
#> 149 2020-07-05         353
#> 150 2020-07-06         829
#> 151 2020-07-07        1017
#> 152 2020-07-08        1156
#> 153 2020-07-09        1130
#> 154 2020-07-10         970
#> 155 2020-07-11         453
#> 156 2020-07-12         392
#> 157 2020-07-13         946
#> 158 2020-07-14         959
#> 159 2020-07-15         987
#> 160 2020-07-16         738
#> 161 2020-07-17         858
#> 162 2020-07-18         462
#> 163 2020-07-19         414
#> 164 2020-07-20         881
#> 165 2020-07-21         968
#> 166 2020-07-22         771
#> 167 2020-07-23         707
#> 168 2020-07-24         564
#> 169 2020-07-25         377
#> 170 2020-07-26         237
#> 171 2020-07-27         806
#> 172 2020-07-28         711
#> 173 2020-07-29         635
#> 174 2020-07-30         709
#> 175 2020-07-31         693
#> 176 2020-08-01         404
#> 177 2020-08-02         214
#> 178 2020-08-03         748
#> 179 2020-08-04         688
#> 180 2020-08-05         647
#> 181 2020-08-06         608
#> 182 2020-08-07         533
#> 183 2020-08-08         287
#> 184 2020-08-09         234
#> 185 2020-08-10         600
#> 186 2020-08-11         710
#> 187 2020-08-12         549
#> 188 2020-08-13         540
#> 189 2020-08-14         501
#> 190 2020-08-15         269
#> 191 2020-08-16         220
#> 192 2020-08-17         635
#> 193 2020-08-18         577
#> 194 2020-08-19         549
#> 195 2020-08-20         548
#> 196 2020-08-21         456
#> 197 2020-08-22         285
#> 198 2020-08-23         181
#> 199 2020-08-24         582
#> 200 2020-08-25         624
#> 201 2020-08-26         898
#> 202 2020-08-27         589
#> 203 2020-08-28         499
#> 204 2020-08-29         260
#> 205 2020-08-30         186
#> 206 2020-08-31         620
#> 207 2020-09-01         562
#> 208 2020-09-02         558
#> 209 2020-09-03         535
#> 210 2020-09-04         457
#> 211 2020-09-05         203
#> 212 2020-09-06         144
#> 213 2020-09-07         132
#> 214 2020-09-08         600
#> 215 2020-09-09         786
#> 216 2020-09-10         578
#> 217 2020-09-11         614
#> 218 2020-09-12         252
#> 219 2020-09-13         156
#> 220 2020-09-14         723
#> 221 2020-09-15         812
#> 222 2020-09-16         822
#> 223 2020-09-17         636
#> 224 2020-09-18         609
#> 225 2020-09-19         296
#> 226 2020-09-20         195
#> 227 2020-09-21         630
#> 228 2020-09-22         650
#> 229 2020-09-23        1042
#> 230 2020-09-24         762
#> 231 2020-09-25         588
#> 232 2020-09-26         383
#> 233 2020-09-27         217
#> 234 2020-09-28         741
#> 235 2020-09-29         746
#> 236 2020-09-30         764
#> 237 2020-10-01         657
#> 238 2020-10-02         642
#> 239 2020-10-03         277
#> 240 2020-10-04         298
#> 241 2020-10-05         970
#> 242 2020-10-06         955
#> 243 2020-10-07        1076
#> 244 2020-10-08        1264
#> 245 2020-10-09         936
#> 246 2020-10-10         481
#> 247 2020-10-11         274
#> 248 2020-10-12         642
#> 249 2020-10-13         845
#> 250 2020-10-14         851
#> 251 2020-10-15         880
#> 252 2020-10-16         881
#> 253 2020-10-17         521
#> 254 2020-10-18         252
#> 255 2020-10-19        1193
#> 256 2020-10-20         976
#> 257 2020-10-21         881
#> 258 2020-10-22        1074
#> 259 2020-10-23         770
#> 260 2020-10-24         454
#> 261 2020-10-25         311
#> 262 2020-10-26        1468
#> 263 2020-10-27        1174
#> 264 2020-10-28        1182
#> 265 2020-10-29        1312
#> 266 2020-10-30        1058
#> 267 2020-10-31         443
#> 268 2020-11-01         596
#> 269 2020-11-02        1608
#> 270 2020-11-03        1574
#> 271 2020-11-04        2088
#> 272 2020-11-05        1986
#> 273 2020-11-06        1749
#> 274 2020-11-07         785
#> 275 2020-11-08         571
#> 276 2020-11-09        1901
#> 277 2020-11-10        1668
#> 278 2020-11-11        1614
#> 279 2020-11-12        1419
#> 280 2020-11-13        1278
#> 281 2020-11-14         798
#> 282 2020-11-15         833
#> 283 2020-11-16        1830
#> 284 2020-11-17        1850
#> 285 2020-11-18        1461
#> 286 2020-11-19        1105
#> 287 2020-11-20        1116
#> 288 2020-11-21         429
#> 289 2020-11-22         341
#> 290 2020-11-23        1503
#> 291 2020-11-24        1461
#> 292 2020-11-25        1094
#> 293 2020-11-26         212
#> 294 2020-11-27         859
#> 295 2020-11-28         492
#> 296 2020-11-29         479
#> 297 2020-11-30        2445
#> 298 2020-12-01        2279
#> 299 2020-12-02        1912
#> 300 2020-12-03        1495
#> 301 2020-12-04        1281
#> 302 2020-12-05         436
#> 303 2020-12-06         353
#> 304 2020-12-07        1413
#> 305 2020-12-08        1345
#> 306 2020-12-09        2172
#> 307 2020-12-10        2824
#> 308 2020-12-11        2815
#> 309 2020-12-12        1762
#> 310 2020-12-13         868
#> 311 2020-12-14        2593
#> 312 2020-12-15        3316
#> 313 2020-12-16        2255
#> 314 2020-12-17        1335
#> 315 2020-12-18        1266
#> 316 2020-12-19         992
#> 317 2020-12-20         695
#> 318 2020-12-21        1807
#> 319 2020-12-22        1350
#> 320 2020-12-23        1333
#> 321 2020-12-24        1119
#> 322 2020-12-25         373
#> 323 2020-12-26        2010
#> 324 2020-12-27        1519
#> 325 2020-12-28        2701
#> 326 2020-12-29        2552
#> 327 2020-12-30        1929
#> 328 2020-12-31        1572
#> 329 2021-01-01         849
#> 330 2021-01-02        1588
#> 331 2021-01-03        1651
#> 332 2021-01-04        4956
#> 333 2021-01-05        4347
#> 334 2021-01-06        3111
#> 335 2021-01-07        2778
#> 336 2021-01-08        2299
#> 337 2021-01-09        2180
#> 338 2021-01-10        1233
#> 339 2021-01-11        3524
#> 340 2021-01-12        3609
#> 341 2021-01-13        3196
#> 342 2021-01-14        3547
#> 343 2021-01-15        4384
#> 344 2021-01-16        2749
#> 345 2021-01-17        1825
#> 346 2021-01-18        3403
#> 347 2021-01-19        4776
#> 348 2021-01-20        3073
#> 349 2021-01-21        2939
#> 350 2021-01-22        2214
#> 351 2021-01-23        1314
#> 352 2021-01-24        1329
#> 353 2021-01-25        2807
#> 354 2021-01-26        2812
#> 355 2021-01-27        3196
#> 356 2021-01-28        3134
#> 357 2021-01-29        2460
#> 358 2021-01-30        1728
#> 359 2021-01-31        1819
#> 360 2021-02-01        4062
#> 361 2021-02-02        4353
#> 362 2021-02-03        3852
#> 363 2021-02-04        2866
#> 364 2021-02-05        2508
#> 365 2021-02-06        1671
#> 366 2021-02-07        1426
#> 367 2021-02-08        3717
#> 368 2021-02-09        4354
#> 369 2021-02-10        4064
#> 370 2021-02-11        3441
#> 371 2021-02-12        2686
#> 372 2021-02-13        1915
#> 373 2021-02-14        1480
#> 374 2021-02-15        4013
#> 375 2021-02-16        3254
#> 376 2021-02-17        3075
#> 377 2021-02-18        2411
#> 378 2021-02-19        2435
#> 379 2021-02-20        1944
#> 380 2021-02-21        1500
#> 381 2021-02-22        3486
#> 382 2021-02-23        3148
#> 383 2021-02-24        3329
#> 384 2021-02-25        2973
#> 385 2021-02-26        2849
#> 386 2021-02-27        2137
#> 387 2021-02-28        1645
#> 388 2021-03-01        3604
#> 389 2021-03-02        3613
#> 390 2021-03-03        4020
#> 391 2021-03-04        3593
#> 392 2021-03-05        3097
#> 393 2021-03-06        2148
#> 394 2021-03-07        2017
#> 395 2021-03-08        4342
#> 396 2021-03-09        3793
#> 397 2021-03-10        3828
#> 398 2021-03-11        3363
#> 399 2021-03-12        4194
#> 400 2021-03-13        2809
#> 401 2021-03-14        2267
#> 402 2021-03-15        6419
#> 403 2021-03-16        4741
#> 404 2021-03-17        4533
#> 405 2021-03-18        4707
#> 406 2021-03-19        4356
#> 407 2021-03-20        3313
#> 408 2021-03-21        2934
#> 409 2021-03-22        5536
#> 410 2021-03-23        5204
#> 411 2021-03-24        5770
#> 412 2021-03-25        5342
#> 413 2021-03-26        4532
#> 414 2021-03-27        3658
#> 415 2021-03-28        3020
#> 416 2021-03-29        6562
#> 417 2021-03-30        5766
#> 418 2021-03-31        6624
#> 419 2021-04-01        6974
#> 420 2021-04-02        5660
#> 421 2021-04-03        4108
#> 422 2021-04-04        2772
#> 423 2021-04-05        7378
#> 424 2021-04-06        6462
#> 425 2021-04-07        6876
#> 426 2021-04-08        6346
#> 427 2021-04-09        6683
#> 428 2021-04-10        4882
#> 429 2021-04-11        3869
#> 430 2021-04-12        7457
#> 431 2021-04-13        6944
#> 432 2021-04-14        7001
#> 433 2021-04-15        8017
#> 434 2021-04-16        5975
#> 435 2021-04-17        4142
#> 436 2021-04-18        2895
#> 437 2021-04-19        6874
#> 438 2021-04-20        6033
#> 439 2021-04-21        5827
#> 440 2021-04-22        5594
#> 441 2021-04-23        5100
#> 442 2021-04-24        3209
#> 443 2021-04-25        2850
#> 444 2021-04-26        6627
#> 445 2021-04-27        6033
#> 446 2021-04-28        5824
#> 447 2021-04-29        5719
#> 448 2021-04-30        5362
#> 449 2021-05-01        3316
#> 450 2021-05-02        2608
#> 451 2021-05-03        6657
#> 452 2021-05-04        5226
#> 453 2021-05-05        5464
#> 454 2021-05-06        4527
#> 455 2021-05-07        4125
#> 456 2021-05-08        2321
#> 457 2021-05-09        1938
#> 458 2021-05-10        4968
#> 459 2021-05-11        4323
#> 460 2021-05-12        3946
#> 461 2021-05-13        3538
#> 462 2021-05-14        2859
#> 463 2021-05-15        2773
#> 464 2021-05-16        1514
#> 465 2021-05-17        3998
#> 466 2021-05-18        3373
#> 467 2021-05-19        3144
#> 468 2021-05-20        2852
#> 469 2021-05-21        2444
#> 470 2021-05-22        1405
#> 471 2021-05-23        1000
#> 472 2021-05-24        2733
#> 473 2021-05-25        2366
#> 474 2021-05-26        2149
#> 475 2021-05-27        2030
#> 476 2021-05-28        1720
#> 477 2021-05-29         888
#> 478 2021-05-30         750
#> 479 2021-05-31         543
#> 480 2021-06-01        2448
#> 481 2021-06-02        1913
#> 482 2021-06-03        1813
#> 483 2021-06-04        1600
#> 484 2021-06-05         948
#> 485 2021-06-06         842
#> 486 2021-06-07        1976
#> 487 2021-06-08        1827
#> 488 2021-06-09        1751
#> 489 2021-06-10        1807
#> 490 2021-06-11        1739
#> 491 2021-06-12        1049
#> 492 2021-06-13        1018
#> 493 2021-06-14        2166
#> 494 2021-06-15        2043
#> 495 2021-06-16        1783
#> 496 2021-06-17        1733
#> 497 2021-06-18        1723
#> 498 2021-06-19        1110
#> 499 2021-06-20         890
#> 500 2021-06-21        2368
#> 501 2021-06-22        2108
#> 502 2021-06-23        2015
#> 503 2021-06-24        2020
#> 504 2021-06-25        1875
#> 505 2021-06-26        1270
#> 506 2021-06-27        1257
#> 507 2021-06-28        2983
#> 508 2021-06-29        2642
#> 509 2021-06-30        2486
#> 510 2021-07-01        2641
#> 511 2021-07-02        2457
#> 512 2021-07-03        1512
#> 513 2021-07-04         676
#> 514 2021-07-05        2462
#> 515 2021-07-06        4389
#> 516 2021-07-07        4411
#> 517 2021-07-08        4217
#> 518 2021-07-09        4100
#> 519 2021-07-10        3175
#> 520 2021-07-11        2686
#> 521 2021-07-12        6010
#> 522 2021-07-13        5241
#> 523 2021-07-14        5347
#> 524 2021-07-15        7228
#> 525 2021-07-16        5814
#> 526 2021-07-17        4342
#> 527 2021-07-18        3809
#> 528 2021-07-19        8542
#> 529 2021-07-20        8610
#> 530 2021-07-21        8725
#> 531 2021-07-22        8514
#> 532 2021-07-23        8445
#> 533 2021-07-24        6354
#> 534 2021-07-25        5018
#> 535 2021-07-26       12018
#> 536 2021-07-27       10869
#> 537 2021-07-28       11103
#> 538 2021-07-29       12173
#> 539 2021-07-30       11560
#> 540 2021-07-31        8156
#> 541 2021-08-01        6904
#> 542 2021-08-02       14154
#> 543 2021-08-03       14066
#> 544 2021-08-04       12969
#> 545 2021-08-05       13178
#> 546 2021-08-06       13018
#> 547 2021-08-07        8242
#> 548 2021-08-08        5725
#> 549 2021-08-09       13768
#> 550 2021-08-10       12710
#> 551 2021-08-11       11431
#> 552 2021-08-12       11050
#> 553 2021-08-13        9567
#> 554 2021-08-14        6096
#> 555 2021-08-15        8638
#> 556 2021-08-16       13069
#> 557 2021-08-17       12088
#> 558 2021-08-18       11283
#> 559 2021-08-19       10274
#> 560 2021-08-20       10579
#> 561 2021-08-21        5882
#> 562 2021-08-22        4117
#> 563 2021-08-23       12179
#> 564 2021-08-24       12551
#> 565 2021-08-25       11820
#> 566 2021-08-26       12291
#> 567 2021-08-27       10310
#> 568 2021-08-28        5451
#> 569 2021-08-29        4199
#> 570 2021-08-30       13047
#> 571 2021-08-31       14378
#> 572 2021-09-01       12511
#> 573 2021-09-02       12447
#> 574 2021-09-03       10606
#> 575 2021-09-04        6211
#> 576 2021-09-05        5295
#> 577 2021-09-06        4782
#> 578 2021-09-07       16771
#> 579 2021-09-08       13589
#> 580 2021-09-09       13385
#> 581 2021-09-10       12658
#> 582 2021-09-11        7039
#> 583 2021-09-12        6168
#> 584 2021-09-13       13428
#> 585 2021-09-14       10925
#> 586 2021-09-15       13679
#> 587 2021-09-16       10274
#> 588 2021-09-17       11108
#> 589 2021-09-18        6035
#> 590 2021-09-19        5332
#> 591 2021-09-20       12413
#> 592 2021-09-21       10938
#> 593 2021-09-22        8517
#> 594 2021-09-23        8597
#> 595 2021-09-24        9002
#> 596 2021-09-25        5088
#> 597 2021-09-26        4240
#> 598 2021-09-27       10813
#> 599 2021-09-28        9912
#> 600 2021-09-29       10186
#> 601 2021-09-30        8776
#> 602 2021-10-01        8913
#> 603 2021-10-02        5549
#> 604 2021-10-03        4853
#> 605 2021-10-04       11002
#> 606 2021-10-05       10798
#> 607 2021-10-06        9873
#> 608 2021-10-07        9157
#> 609 2021-10-08        8643
#> 610 2021-10-09        5811
#> 611 2021-10-10        4903
#> 612 2021-10-11        9925
#> 613 2021-10-12        8631
#> 614 2021-10-13        8320
#> 615 2021-10-14        9043
#> 616 2021-10-15       10021
#> 617 2021-10-16        4898
#> 618 2021-10-17        4090
#> 619 2021-10-18       10336
#> 620 2021-10-19        8886
#> 621 2021-10-20        8416
#> 622 2021-10-21        8372
#> 623 2021-10-22        8667
#> 624 2021-10-23        5717
#> 625 2021-10-24        4456
#> 626 2021-10-25       10885
#> 627 2021-10-26        9175
#> 628 2021-10-27        9302
#> 629 2021-10-28        8881
#> 630 2021-10-29        8151
#> 631 2021-10-30        5546
#> 632 2021-10-31        4453
#> 633 2021-11-01        9794
#> 634 2021-11-02       10413
#> 635 2021-11-03       10666
#> 636 2021-11-04        9832
#> 637 2021-11-05        8820
#> 638 2021-11-06        5967
#> 639 2021-11-07        5556
#> 640 2021-11-08       10634
#> 641 2021-11-09       11038
#> 642 2021-11-10       11550
#> 643 2021-11-11        8822
#> 644 2021-11-12        9646
#> 645 2021-11-13        7263
#> 646 2021-11-14        4865
#> 647 2021-11-15       11989
#> 648 2021-11-16       11368
#> 649 2021-11-17       10609
#> 650 2021-11-18        9333
#> 651 2021-11-19        7957
#> 652 2021-11-20        5176
#> 653 2021-11-21        5425
#> 654 2021-11-22        9834
#> 655 2021-11-23        9470
#> 656 2021-11-24        8457
#> 657 2021-11-25        1323
#> 658 2021-11-26        7074
#> 659 2021-11-27        6917
#> 660 2021-11-28        6206
#> 661 2021-11-29       13824
#> 662 2021-11-30       11339
#> 663 2021-12-01       11924
#> 664 2021-12-02       10124
#> 665 2021-12-03        9635
#> 666 2021-12-04        7073
#> 667 2021-12-05        4780
#> 668 2021-12-06       12117
#> 669 2021-12-07       10911
#> 670 2021-12-08       11250
#> 671 2021-12-09        9924
#> 672 2021-12-10        8925
#> 673 2021-12-11        6068
#> 674 2021-12-12        4730
#> 675 2021-12-13       11141
#> 676 2021-12-14       12864
#> 677 2021-12-15       15151
#> 678 2021-12-16       12530
#> 679 2021-12-17       12557
#> 680 2021-12-18        8196
#> 681 2021-12-19        6027
#> 682 2021-12-20       14773
#> 683 2021-12-21       18266
#> 684 2021-12-22       19917
#> 685 2021-12-23       15185
#> 686 2021-12-24        5688
#> 687 2021-12-25        2659
#> 688 2021-12-26       10221
#> 689 2021-12-27       21562
#> 690 2021-12-28       21007
#> 691 2021-12-29       21785
#> 692 2021-12-30       20276
#> 693 2021-12-31        7622
#> 694 2022-01-01        4512
#> 695 2022-01-02       10829
#> 696 2022-01-03       23456
#> 697 2022-01-04       26468
#> 698 2022-01-05       18694
#> 699 2022-01-06       16793
#> 700 2022-01-07       15180
#> 701 2022-01-08        9604
#> 702 2022-01-09        8661
#> 703 2022-01-10       19798
#> 704 2022-01-11       19022
#> 705 2022-01-12       16802
#> 706 2022-01-13       15685
#> 707 2022-01-14       15210
#> 708 2022-01-15       12675
#> 709 2022-01-16        7810
#> 710 2022-01-17       13751
#> 711 2022-01-18       20686
#> 712 2022-01-19       20568
#> 713 2022-01-20       15361
#> 714 2022-01-21       13408
#> 715 2022-01-22        8838
#> 716 2022-01-23        7523
#> 717 2022-01-24       16546
#> 718 2022-01-25       15094
#> 719 2022-01-26       14204
#> 720 2022-01-27       14718
#> 721 2022-01-28       11447
#> 722 2022-01-29        8298
#> 723 2022-01-30        7219
#> 724 2022-01-31       16852
#> 725 2022-02-01       15101
#> 726 2022-02-02       13773
#> 727 2022-02-03       11239
#> 728 2022-02-04       10167
#> 729 2022-02-05        6264
#> 730 2022-02-06        6163
#> 731 2022-02-07       15773
#> 732 2022-02-08       13517
#> 733 2022-02-09       12125
#> 734 2022-02-10       10775
#> 735 2022-02-11        9094
#> 736 2022-02-12        6182
#> 737 2022-02-13        4505
#> 738 2022-02-14        9626
#> 739 2022-02-15       10099
#> 740 2022-02-16        9405
#> 741 2022-02-17        7581
#> 742 2022-02-18        6661
#> 743 2022-02-19        4106
#> 744 2022-02-20        3082
#> 745 2022-02-21        5983
#> 746 2022-02-22        7942
#> 747 2022-02-23        5705
#> 748 2022-02-24        4867
#> 749 2022-02-25        4498
#> 750 2022-02-26        2948
#> 751 2022-02-27        2329
#> 752 2022-02-28        4796
#> 753 2022-03-01        4585
#> 754 2022-03-02        3593
#> 755 2022-03-03        3063
#> 756 2022-03-04        3016
#> 757 2022-03-05        1933
#> 758 2022-03-06        1647
#> 759 2022-03-07        3997
#> 760 2022-03-08        2998
#> 761 2022-03-09        2848
#> 762 2022-03-10        2721
#> 763 2022-03-11        2373
#> 764 2022-03-12        1428
#> 765 2022-03-13        1554
#> 766 2022-03-14        3437
#> 767 2022-03-15        3185
#> 768 2022-03-16        2828
#> 769 2022-03-17        2605
#> 770 2022-03-18        2440
#> 771 2022-03-19        1900
#> 772 2022-03-20        1862
#> 773 2022-03-21        3699
#> 774 2022-03-22        3198
#> 775 2022-03-23        3191
#> 776 2022-03-24        2914
#> 777 2022-03-25        2716
#> 778 2022-03-26        2201
#> 779 2022-03-27        2266
#> 780 2022-03-28        3879
#> 781 2022-03-29        3451
#> 782 2022-03-30        3635
#> 783 2022-03-31        3327
#> 784 2022-04-01        3298
#> 785 2022-04-02        2477
#> 786 2022-04-03        2370
#> 787 2022-04-04        5296
#> 788 2022-04-05        4595
#> 789 2022-04-06        4555
#> 790 2022-04-07        4578
#> 791 2022-04-08        4736
#> 792 2022-04-09        3287
#> 793 2022-04-10        3398
#> 794 2022-04-11        6487
#> 795 2022-04-12        5171
#> 796 2022-04-13        4794
#> 797 2022-04-14        4875
#> 798 2022-04-15        5886
#> 799 2022-04-16        3805
#> 800 2022-04-17        1993
#> 801 2022-04-18        6564
#> 802 2022-04-19        5681
#> 803 2022-04-20        5424
#> 804 2022-04-21        5215
#> 805 2022-04-22        5390
#> 806 2022-04-23        3827
#> 807 2022-04-24        3974
#> 808 2022-04-25        6957
#> 809 2022-04-26        6034
#> 810 2022-04-27        5184
#> 811 2022-04-28        4323
#> 812 2022-04-29        4683
#> 813 2022-04-30        3724
#> 814 2022-05-01        3374
#> 815 2022-05-02        6914
#> 816 2022-05-03        6386
#> 817 2022-05-04        6204
#> 818 2022-05-05        6294
#> 819 2022-05-06        5826
#> 820 2022-05-07        4471
#> 821 2022-05-08        3964
#> 822 2022-05-09        8280
#> 823 2022-05-10        7195
#> 824 2022-05-11        6758
#> 825 2022-05-12        6573
#> 826 2022-05-13        6225
#> 827 2022-05-14        4854
#> 828 2022-05-15        4821
#> 829 2022-05-16        8119
#> 830 2022-05-17        7181
#> 831 2022-05-18        6890
#> 832 2022-05-19        5650
#> 833 2022-05-20        5814
#> 834 2022-05-21        4751
#> 835 2022-05-22        4667
#> 836 2022-05-23        8261
#> 837 2022-05-24        7776
#> 838 2022-05-25        6827
#> 839 2022-05-26        5838
#> 840 2022-05-27        6047
#> 841 2022-05-28        4040
#> 842 2022-05-29        4553
#> 843 2022-05-30        2439
#> 844 2022-05-31       10083
#> 845 2022-06-01        7025
#> 846 2022-06-02        6561
#> 847 2022-06-03        6068
#> 848 2022-06-04        4239
#> 849 2022-06-05        4230
#> 850 2022-06-06        8272
#> 851 2022-06-07        7606
#> 852 2022-06-08        6780
#> 853 2022-06-09        6424
#> 854 2022-06-10        5687
#> 855 2022-06-11        3886
#> 856 2022-06-12        3701
#> 857 2022-06-13        6740
#> 858 2022-06-14        5655
#> 859 2022-06-15        6890
#> 860 2022-06-16        5772
#> 861 2022-06-17        5509
#> 862 2022-06-18        3684
#> 863 2022-06-19        2797
#> 864 2022-06-20        5668
#> 865 2022-06-21        6254
#> 866 2022-06-22        6246
#> 867 2022-06-23        5758
#> 868 2022-06-24        5169
#> 869 2022-06-25        3592
#> 870 2022-06-26        3739
#> 871 2022-06-27        7813
#> 872 2022-06-28        6371
#> 873 2022-06-29        5907
#> 874 2022-06-30        5862
#> 875 2022-07-01        6027
#> 876 2022-07-02        3807
#> 877 2022-07-03        3669
#> 878 2022-07-04        2365
#> 879 2022-07-05        9954
#> 880 2022-07-06        7497
#> 881 2022-07-07        6506
#> 882 2022-07-08        6413
#> 883 2022-07-09        4688
#> 884 2022-07-10        5055
#> 885 2022-07-11        8384
#> 886 2022-07-12        6521
#> 887 2022-07-13        6269
#> 888 2022-07-14        6292
#> 889 2022-07-15        6332
#> 890 2022-07-16        5532
#> 891 2022-07-17        5351
#> 892 2022-07-18        8236
#> 893 2022-07-19        6785
#> 894 2022-07-20        5503
#> 895 2022-07-21        5175
#> 896 2022-07-22        6082
#> 897 2022-07-23        3759
#> 898 2022-07-24        4055
#> 899 2022-07-25        7766
#> 900 2022-07-26        5916
#> 901 2022-07-27        5110
#> 902 2022-07-28        4931
#> 903 2022-07-29        4844
#> 904 2022-07-30        3766
#> 905 2022-07-31        4177
#> 906 2022-08-01        8055
#> 907 2022-08-02        5837
#> 908 2022-08-03        5131
#> 909 2022-08-04        4746
#> 910 2022-08-05        4112
#> 911 2022-08-06        2751
#> 912 2022-08-07        2742
#> 913 2022-08-08        5385
#> 914 2022-08-09        4220
#> 915 2022-08-10        4283
#> 916 2022-08-11        4849
#> 917 2022-08-12        4618
#> 918 2022-08-13        3515
#> 919 2022-08-14        3142
#> 920 2022-08-15        5307
#> 921 2022-08-16        4490
#> 922 2022-08-17        4423
#> 923 2022-08-18        4119
#> 924 2022-08-19        4301
#> 925 2022-08-20        3083
#> 926 2022-08-21        3341
#> 927 2022-08-22        5634
#> 928 2022-08-23        4130
#> 929 2022-08-24        3679
#> 930 2022-08-25        3582
#> 931 2022-08-26        4150
#> 932 2022-08-27        3845
#> 933 2022-08-28        3300
#> 934 2022-08-29        5297
#> 935 2022-08-30        4122
#> 936 2022-08-31        3501
#> 937 2022-09-01        3288
#> 938 2022-09-02        2889
#> 939 2022-09-03        2677
#> 940 2022-09-04        3005
#> 941 2022-09-05         966
#> 942 2022-09-06        5424
#> 943 2022-09-07        3121
#> 944 2022-09-08        2573
#> 945 2022-09-09        2281
#> 946 2022-09-10        2629
#> 947 2022-09-11        2230
#> 948 2022-09-12        4668
#> 949 2022-09-13        3871
#> 950 2022-09-14        2753
#> 951 2022-09-15        2976
#> 952 2022-09-16        3076
#> 953 2022-09-17        2376
#> 954 2022-09-18        1998
#> 955 2022-09-19        3623
#> 956 2022-09-20        2626
#> 957 2022-09-21        2499
#> 958 2022-09-22        1994
#> 959 2022-09-23        1784
#> 960 2022-09-24        1281
#> 961 2022-09-25        1044
#> 962 2022-09-26        1950
#> 963 2022-09-27        1023
#> 964 2022-09-28         673
#> 965 2022-09-29         555
#> 966 2022-09-30         563
#> 967 2022-10-01         343
#> 968 2022-10-02          82
#> 969 2022-10-03          93
#> 970 2022-10-04         114
#> 971 2022-10-05          73
#> 972 2022-10-06          40
#> 973 2022-10-07          33
#> 974 2022-10-08          33
#> 975 2022-10-09          20
#> 976 2022-10-10           6
#> 977 2022-10-11           2
#> 978 2022-10-12           5
#> 979 2022-10-13           8

# Returns the total number of global sequences in the outbreak.info API.
getSeqCounts(cumulative=TRUE)
#> Retrieving data... 
#>           key    value
#> 1 total_count 13097778