Retrieve sequence count data

Retrieves number of sequences by location, by date of sample collection. See getSubmissionDateByLocation to view similar information by date of sequence submission.

getSeqCounts(location = NULL, cumulative = NULL, subadmin = NULL)

Arguments

location

(optional) a location name (if not specified, global total counts returned)

cumulative

(optional) Boolean (T/F), T returns cumulative number of sequences to date

subadmin

(optional) Boolean (T/F), subadmin=T and cumulative=T returns cumulative number of sequences for next administrative level

Value

dataframe

Examples

# Retrieves the number of samples sequenced in the U.S. each day, by date of sample collection
getSeqCounts(location="United States")
#> Retrieving data... 
#>           date total_count
#> 1   2020-01-26           3
#> 2   2020-01-27           1
#> 3   2020-01-28           2
#> 4   2020-01-29           4
#> 5   2020-01-31           1
#> 6   2020-02-02           1
#> 7   2020-02-06           1
#> 8   2020-02-08           1
#> 9   2020-02-10           1
#> 10  2020-02-11           1
#> 11  2020-02-17           9
#> 12  2020-02-18           6
#> 13  2020-02-20           2
#> 14  2020-02-21           8
#> 15  2020-02-22           3
#> 16  2020-02-23           1
#> 17  2020-02-24           7
#> 18  2020-02-25           4
#> 19  2020-02-26           7
#> 20  2020-02-27           9
#> 21  2020-02-28          27
#> 22  2020-02-29          37
#> 23  2020-03-01          27
#> 24  2020-03-02          50
#> 25  2020-03-03          44
#> 26  2020-03-04          70
#> 27  2020-03-05         107
#> 28  2020-03-06          85
#> 29  2020-03-07          96
#> 30  2020-03-08         107
#> 31  2020-03-09         197
#> 32  2020-03-10         196
#> 33  2020-03-11         243
#> 34  2020-03-12         389
#> 35  2020-03-13         553
#> 36  2020-03-14         539
#> 37  2020-03-15        1008
#> 38  2020-03-16         953
#> 39  2020-03-17         795
#> 40  2020-03-18         859
#> 41  2020-03-19         763
#> 42  2020-03-20         645
#> 43  2020-03-21         497
#> 44  2020-03-22         403
#> 45  2020-03-23         658
#> 46  2020-03-24         767
#> 47  2020-03-25         795
#> 48  2020-03-26         762
#> 49  2020-03-27         747
#> 50  2020-03-28         512
#> 51  2020-03-29         383
#> 52  2020-03-30         809
#> 53  2020-03-31         712
#> 54  2020-04-01         885
#> 55  2020-04-02         769
#> 56  2020-04-03         784
#> 57  2020-04-04         485
#> 58  2020-04-05         392
#> 59  2020-04-06         724
#> 60  2020-04-07         578
#> 61  2020-04-08         683
#> 62  2020-04-09         560
#> 63  2020-04-10         567
#> 64  2020-04-11         406
#> 65  2020-04-12         256
#> 66  2020-04-13         709
#> 67  2020-04-14         477
#> 68  2020-04-15         857
#> 69  2020-04-16         437
#> 70  2020-04-17         467
#> 71  2020-04-18         332
#> 72  2020-04-19         220
#> 73  2020-04-20         548
#> 74  2020-04-21         575
#> 75  2020-04-22         451
#> 76  2020-04-23         448
#> 77  2020-04-24         464
#> 78  2020-04-25         255
#> 79  2020-04-26         189
#> 80  2020-04-27         557
#> 81  2020-04-28         530
#> 82  2020-04-29         618
#> 83  2020-04-30         666
#> 84  2020-05-01         577
#> 85  2020-05-02         309
#> 86  2020-05-03         192
#> 87  2020-05-04         593
#> 88  2020-05-05         561
#> 89  2020-05-06         429
#> 90  2020-05-07         509
#> 91  2020-05-08         450
#> 92  2020-05-09         292
#> 93  2020-05-10         176
#> 94  2020-05-11         484
#> 95  2020-05-12         461
#> 96  2020-05-13         508
#> 97  2020-05-14         625
#> 98  2020-05-15         503
#> 99  2020-05-16         187
#> 100 2020-05-17         154
#> 101 2020-05-18         373
#> 102 2020-05-19         412
#> 103 2020-05-20         390
#> 104 2020-05-21         464
#> 105 2020-05-22         462
#> 106 2020-05-23         181
#> 107 2020-05-24         173
#> 108 2020-05-25         176
#> 109 2020-05-26         495
#> 110 2020-05-27         477
#> 111 2020-05-28         369
#> 112 2020-05-29         469
#> 113 2020-05-30         225
#> 114 2020-05-31         181
#> 115 2020-06-01         442
#> 116 2020-06-02         373
#> 117 2020-06-03         478
#> 118 2020-06-04         452
#> 119 2020-06-05         433
#> 120 2020-06-06         187
#> 121 2020-06-07         142
#> 122 2020-06-08         465
#> 123 2020-06-09         461
#> 124 2020-06-10         438
#> 125 2020-06-11         426
#> 126 2020-06-12         357
#> 127 2020-06-13         378
#> 128 2020-06-14         238
#> 129 2020-06-15         705
#> 130 2020-06-16         674
#> 131 2020-06-17         482
#> 132 2020-06-18         728
#> 133 2020-06-19         656
#> 134 2020-06-20         383
#> 135 2020-06-21         436
#> 136 2020-06-22         775
#> 137 2020-06-23         647
#> 138 2020-06-24         770
#> 139 2020-06-25         713
#> 140 2020-06-26         694
#> 141 2020-06-27         345
#> 142 2020-06-28         361
#> 143 2020-06-29         837
#> 144 2020-06-30         730
#> 145 2020-07-01         851
#> 146 2020-07-02         870
#> 147 2020-07-03         558
#> 148 2020-07-04         336
#> 149 2020-07-05         351
#> 150 2020-07-06         825
#> 151 2020-07-07        1014
#> 152 2020-07-08        1153
#> 153 2020-07-09        1129
#> 154 2020-07-10         969
#> 155 2020-07-11         452
#> 156 2020-07-12         390
#> 157 2020-07-13         942
#> 158 2020-07-14         955
#> 159 2020-07-15         984
#> 160 2020-07-16         735
#> 161 2020-07-17         855
#> 162 2020-07-18         461
#> 163 2020-07-19         410
#> 164 2020-07-20         879
#> 165 2020-07-21         968
#> 166 2020-07-22         769
#> 167 2020-07-23         702
#> 168 2020-07-24         559
#> 169 2020-07-25         370
#> 170 2020-07-26         234
#> 171 2020-07-27         802
#> 172 2020-07-28         708
#> 173 2020-07-29         631
#> 174 2020-07-30         703
#> 175 2020-07-31         692
#> 176 2020-08-01         400
#> 177 2020-08-02         213
#> 178 2020-08-03         744
#> 179 2020-08-04         686
#> 180 2020-08-05         644
#> 181 2020-08-06         606
#> 182 2020-08-07         533
#> 183 2020-08-08         287
#> 184 2020-08-09         234
#> 185 2020-08-10         600
#> 186 2020-08-11         710
#> 187 2020-08-12         549
#> 188 2020-08-13         540
#> 189 2020-08-14         501
#> 190 2020-08-15         269
#> 191 2020-08-16         220
#> 192 2020-08-17         635
#> 193 2020-08-18         575
#> 194 2020-08-19         549
#> 195 2020-08-20         548
#> 196 2020-08-21         456
#> 197 2020-08-22         285
#> 198 2020-08-23         181
#> 199 2020-08-24         580
#> 200 2020-08-25         623
#> 201 2020-08-26         898
#> 202 2020-08-27         589
#> 203 2020-08-28         499
#> 204 2020-08-29         260
#> 205 2020-08-30         186
#> 206 2020-08-31         620
#> 207 2020-09-01         562
#> 208 2020-09-02         558
#> 209 2020-09-03         535
#> 210 2020-09-04         457
#> 211 2020-09-05         203
#> 212 2020-09-06         144
#> 213 2020-09-07         132
#> 214 2020-09-08         600
#> 215 2020-09-09         785
#> 216 2020-09-10         570
#> 217 2020-09-11         614
#> 218 2020-09-12         252
#> 219 2020-09-13         156
#> 220 2020-09-14         723
#> 221 2020-09-15         812
#> 222 2020-09-16         821
#> 223 2020-09-17         635
#> 224 2020-09-18         609
#> 225 2020-09-19         296
#> 226 2020-09-20         195
#> 227 2020-09-21         630
#> 228 2020-09-22         650
#> 229 2020-09-23        1035
#> 230 2020-09-24         761
#> 231 2020-09-25         583
#> 232 2020-09-26         383
#> 233 2020-09-27         217
#> 234 2020-09-28         741
#> 235 2020-09-29         746
#> 236 2020-09-30         764
#> 237 2020-10-01         657
#> 238 2020-10-02         642
#> 239 2020-10-03         277
#> 240 2020-10-04         298
#> 241 2020-10-05         968
#> 242 2020-10-06         955
#> 243 2020-10-07        1076
#> 244 2020-10-08        1264
#> 245 2020-10-09         936
#> 246 2020-10-10         480
#> 247 2020-10-11         274
#> 248 2020-10-12         642
#> 249 2020-10-13         845
#> 250 2020-10-14         850
#> 251 2020-10-15         879
#> 252 2020-10-16         881
#> 253 2020-10-17         521
#> 254 2020-10-18         252
#> 255 2020-10-19        1193
#> 256 2020-10-20         976
#> 257 2020-10-21         880
#> 258 2020-10-22        1074
#> 259 2020-10-23         770
#> 260 2020-10-24         454
#> 261 2020-10-25         311
#> 262 2020-10-26        1466
#> 263 2020-10-27        1171
#> 264 2020-10-28        1182
#> 265 2020-10-29        1305
#> 266 2020-10-30        1058
#> 267 2020-10-31         443
#> 268 2020-11-01         596
#> 269 2020-11-02        1608
#> 270 2020-11-03        1572
#> 271 2020-11-04        2083
#> 272 2020-11-05        1986
#> 273 2020-11-06        1749
#> 274 2020-11-07         785
#> 275 2020-11-08         571
#> 276 2020-11-09        1900
#> 277 2020-11-10        1666
#> 278 2020-11-11        1614
#> 279 2020-11-12        1419
#> 280 2020-11-13        1278
#> 281 2020-11-14         798
#> 282 2020-11-15         833
#> 283 2020-11-16        1830
#> 284 2020-11-17        1850
#> 285 2020-11-18        1461
#> 286 2020-11-19        1105
#> 287 2020-11-20        1116
#> 288 2020-11-21         429
#> 289 2020-11-22         341
#> 290 2020-11-23        1494
#> 291 2020-11-24        1461
#> 292 2020-11-25        1094
#> 293 2020-11-26         212
#> 294 2020-11-27         859
#> 295 2020-11-28         492
#> 296 2020-11-29         478
#> 297 2020-11-30        2444
#> 298 2020-12-01        2278
#> 299 2020-12-02        1912
#> 300 2020-12-03        1477
#> 301 2020-12-04        1261
#> 302 2020-12-05         436
#> 303 2020-12-06         353
#> 304 2020-12-07        1388
#> 305 2020-12-08        1311
#> 306 2020-12-09        2167
#> 307 2020-12-10        2765
#> 308 2020-12-11        2782
#> 309 2020-12-12        1760
#> 310 2020-12-13         867
#> 311 2020-12-14        2590
#> 312 2020-12-15        3314
#> 313 2020-12-16        2248
#> 314 2020-12-17        1329
#> 315 2020-12-18        1262
#> 316 2020-12-19         992
#> 317 2020-12-20         694
#> 318 2020-12-21        1797
#> 319 2020-12-22        1347
#> 320 2020-12-23        1331
#> 321 2020-12-24        1116
#> 322 2020-12-25         373
#> 323 2020-12-26        2006
#> 324 2020-12-27        1519
#> 325 2020-12-28        2695
#> 326 2020-12-29        2549
#> 327 2020-12-30        1929
#> 328 2020-12-31        1572
#> 329 2021-01-01         848
#> 330 2021-01-02        1588
#> 331 2021-01-03        1649
#> 332 2021-01-04        4946
#> 333 2021-01-05        4298
#> 334 2021-01-06        3105
#> 335 2021-01-07        2767
#> 336 2021-01-08        2291
#> 337 2021-01-09        2170
#> 338 2021-01-10        1230
#> 339 2021-01-11        3524
#> 340 2021-01-12        3608
#> 341 2021-01-13        3194
#> 342 2021-01-14        3546
#> 343 2021-01-15        4382
#> 344 2021-01-16        2748
#> 345 2021-01-17        1825
#> 346 2021-01-18        3402
#> 347 2021-01-19        4775
#> 348 2021-01-20        3071
#> 349 2021-01-21        2936
#> 350 2021-01-22        2213
#> 351 2021-01-23        1314
#> 352 2021-01-24        1328
#> 353 2021-01-25        2804
#> 354 2021-01-26        2812
#> 355 2021-01-27        3195
#> 356 2021-01-28        3134
#> 357 2021-01-29        2458
#> 358 2021-01-30        1728
#> 359 2021-01-31        1816
#> 360 2021-02-01        4062
#> 361 2021-02-02        4352
#> 362 2021-02-03        3848
#> 363 2021-02-04        2864
#> 364 2021-02-05        2487
#> 365 2021-02-06        1669
#> 366 2021-02-07        1426
#> 367 2021-02-08        3716
#> 368 2021-02-09        4354
#> 369 2021-02-10        4062
#> 370 2021-02-11        3441
#> 371 2021-02-12        2686
#> 372 2021-02-13        1915
#> 373 2021-02-14        1479
#> 374 2021-02-15        4013
#> 375 2021-02-16        3253
#> 376 2021-02-17        3074
#> 377 2021-02-18        2409
#> 378 2021-02-19        2435
#> 379 2021-02-20        1944
#> 380 2021-02-21        1500
#> 381 2021-02-22        3484
#> 382 2021-02-23        3146
#> 383 2021-02-24        3327
#> 384 2021-02-25        2973
#> 385 2021-02-26        2849
#> 386 2021-02-27        2136
#> 387 2021-02-28        1645
#> 388 2021-03-01        3604
#> 389 2021-03-02        3611
#> 390 2021-03-03        4019
#> 391 2021-03-04        3593
#> 392 2021-03-05        3095
#> 393 2021-03-06        2148
#> 394 2021-03-07        2017
#> 395 2021-03-08        4340
#> 396 2021-03-09        3793
#> 397 2021-03-10        3827
#> 398 2021-03-11        3362
#> 399 2021-03-12        4192
#> 400 2021-03-13        2809
#> 401 2021-03-14        2265
#> 402 2021-03-15        6419
#> 403 2021-03-16        4741
#> 404 2021-03-17        4533
#> 405 2021-03-18        4707
#> 406 2021-03-19        4353
#> 407 2021-03-20        3311
#> 408 2021-03-21        2934
#> 409 2021-03-22        5536
#> 410 2021-03-23        5204
#> 411 2021-03-24        5770
#> 412 2021-03-25        5342
#> 413 2021-03-26        4532
#> 414 2021-03-27        3658
#> 415 2021-03-28        3020
#> 416 2021-03-29        6561
#> 417 2021-03-30        5765
#> 418 2021-03-31        6624
#> 419 2021-04-01        6974
#> 420 2021-04-02        5658
#> 421 2021-04-03        4108
#> 422 2021-04-04        2772
#> 423 2021-04-05        7378
#> 424 2021-04-06        6462
#> 425 2021-04-07        6876
#> 426 2021-04-08        6346
#> 427 2021-04-09        6683
#> 428 2021-04-10        4882
#> 429 2021-04-11        3869
#> 430 2021-04-12        7453
#> 431 2021-04-13        6944
#> 432 2021-04-14        7001
#> 433 2021-04-15        8083
#> 434 2021-04-16        5974
#> 435 2021-04-17        4142
#> 436 2021-04-18        2895
#> 437 2021-04-19        6872
#> 438 2021-04-20        6033
#> 439 2021-04-21        5827
#> 440 2021-04-22        5594
#> 441 2021-04-23        5099
#> 442 2021-04-24        3209
#> 443 2021-04-25        2850
#> 444 2021-04-26        6623
#> 445 2021-04-27        6033
#> 446 2021-04-28        5824
#> 447 2021-04-29        5719
#> 448 2021-04-30        5362
#> 449 2021-05-01        3316
#> 450 2021-05-02        2608
#> 451 2021-05-03        6657
#> 452 2021-05-04        5226
#> 453 2021-05-05        5464
#> 454 2021-05-06        4527
#> 455 2021-05-07        4125
#> 456 2021-05-08        2321
#> 457 2021-05-09        1938
#> 458 2021-05-10        4967
#> 459 2021-05-11        4323
#> 460 2021-05-12        3946
#> 461 2021-05-13        3538
#> 462 2021-05-14        2859
#> 463 2021-05-15        2773
#> 464 2021-05-16        1512
#> 465 2021-05-17        3998
#> 466 2021-05-18        3373
#> 467 2021-05-19        3144
#> 468 2021-05-20        2852
#> 469 2021-05-21        2444
#> 470 2021-05-22        1405
#> 471 2021-05-23        1000
#> 472 2021-05-24        2733
#> 473 2021-05-25        2366
#> 474 2021-05-26        2149
#> 475 2021-05-27        2030
#> 476 2021-05-28        1720
#> 477 2021-05-29         888
#> 478 2021-05-30         750
#> 479 2021-05-31         543
#> 480 2021-06-01        2448
#> 481 2021-06-02        1913
#> 482 2021-06-03        1813
#> 483 2021-06-04        1600
#> 484 2021-06-05         948
#> 485 2021-06-06         842
#> 486 2021-06-07        1976
#> 487 2021-06-08        1827
#> 488 2021-06-09        1751
#> 489 2021-06-10        1807
#> 490 2021-06-11        1739
#> 491 2021-06-12        1049
#> 492 2021-06-13        1018
#> 493 2021-06-14        2166
#> 494 2021-06-15        2043
#> 495 2021-06-16        1783
#> 496 2021-06-17        1733
#> 497 2021-06-18        1723
#> 498 2021-06-19        1109
#> 499 2021-06-20         890
#> 500 2021-06-21        2368
#> 501 2021-06-22        2108
#> 502 2021-06-23        2015
#> 503 2021-06-24        2020
#> 504 2021-06-25        1875
#> 505 2021-06-26        1270
#> 506 2021-06-27        1257
#> 507 2021-06-28        2983
#> 508 2021-06-29        2642
#> 509 2021-06-30        2486
#> 510 2021-07-01        2641
#> 511 2021-07-02        2457
#> 512 2021-07-03        1512
#> 513 2021-07-04         676
#> 514 2021-07-05        2462
#> 515 2021-07-06        4389
#> 516 2021-07-07        4411
#> 517 2021-07-08        4217
#> 518 2021-07-09        4100
#> 519 2021-07-10        3175
#> 520 2021-07-11        2686
#> 521 2021-07-12        6010
#> 522 2021-07-13        5241
#> 523 2021-07-14        5347
#> 524 2021-07-15        7228
#> 525 2021-07-16        5814
#> 526 2021-07-17        4341
#> 527 2021-07-18        3809
#> 528 2021-07-19        8539
#> 529 2021-07-20        8607
#> 530 2021-07-21        8721
#> 531 2021-07-22        8513
#> 532 2021-07-23        8444
#> 533 2021-07-24        6350
#> 534 2021-07-25        5014
#> 535 2021-07-26       12016
#> 536 2021-07-27       10868
#> 537 2021-07-28       11098
#> 538 2021-07-29       12171
#> 539 2021-07-30       11555
#> 540 2021-07-31        8155
#> 541 2021-08-01        6902
#> 542 2021-08-02       14153
#> 543 2021-08-03       14061
#> 544 2021-08-04       12959
#> 545 2021-08-05       13170
#> 546 2021-08-06       13011
#> 547 2021-08-07        8242
#> 548 2021-08-08        5721
#> 549 2021-08-09       13756
#> 550 2021-08-10       12704
#> 551 2021-08-11       11425
#> 552 2021-08-12       11049
#> 553 2021-08-13        9561
#> 554 2021-08-14        6091
#> 555 2021-08-15        8638
#> 556 2021-08-16       13062
#> 557 2021-08-17       12087
#> 558 2021-08-18       11281
#> 559 2021-08-19       10270
#> 560 2021-08-20       10578
#> 561 2021-08-21        5882
#> 562 2021-08-22        4112
#> 563 2021-08-23       12176
#> 564 2021-08-24       12547
#> 565 2021-08-25       11815
#> 566 2021-08-26       12288
#> 567 2021-08-27       10304
#> 568 2021-08-28        5449
#> 569 2021-08-29        4197
#> 570 2021-08-30       13011
#> 571 2021-08-31       14374
#> 572 2021-09-01       12509
#> 573 2021-09-02       12445
#> 574 2021-09-03       10605
#> 575 2021-09-04        6208
#> 576 2021-09-05        5291
#> 577 2021-09-06        4781
#> 578 2021-09-07       16767
#> 579 2021-09-08       13583
#> 580 2021-09-09       13375
#> 581 2021-09-10       12655
#> 582 2021-09-11        7038
#> 583 2021-09-12        6168
#> 584 2021-09-13       13426
#> 585 2021-09-14       10922
#> 586 2021-09-15       13676
#> 587 2021-09-16       10264
#> 588 2021-09-17       11104
#> 589 2021-09-18        6032
#> 590 2021-09-19        5330
#> 591 2021-09-20       12403
#> 592 2021-09-21       10921
#> 593 2021-09-22        8508
#> 594 2021-09-23        8567
#> 595 2021-09-24        8989
#> 596 2021-09-25        5086
#> 597 2021-09-26        4239
#> 598 2021-09-27       10806
#> 599 2021-09-28        9908
#> 600 2021-09-29       10185
#> 601 2021-09-30        8773
#> 602 2021-10-01        8906
#> 603 2021-10-02        5547
#> 604 2021-10-03        4848
#> 605 2021-10-04       10998
#> 606 2021-10-05       10794
#> 607 2021-10-06        9870
#> 608 2021-10-07        9153
#> 609 2021-10-08        8640
#> 610 2021-10-09        5810
#> 611 2021-10-10        4901
#> 612 2021-10-11        9924
#> 613 2021-10-12        8630
#> 614 2021-10-13        8317
#> 615 2021-10-14        9039
#> 616 2021-10-15       10017
#> 617 2021-10-16        4896
#> 618 2021-10-17        4089
#> 619 2021-10-18       10336
#> 620 2021-10-19        8884
#> 621 2021-10-20        8412
#> 622 2021-10-21        8371
#> 623 2021-10-22        8667
#> 624 2021-10-23        5716
#> 625 2021-10-24        4455
#> 626 2021-10-25       10884
#> 627 2021-10-26        9172
#> 628 2021-10-27        9300
#> 629 2021-10-28        8878
#> 630 2021-10-29        8151
#> 631 2021-10-30        5546
#> 632 2021-10-31        4453
#> 633 2021-11-01        9794
#> 634 2021-11-02       10411
#> 635 2021-11-03       10666
#> 636 2021-11-04        9831
#> 637 2021-11-05        8819
#> 638 2021-11-06        5966
#> 639 2021-11-07        5551
#> 640 2021-11-08       10632
#> 641 2021-11-09       11037
#> 642 2021-11-10       11550
#> 643 2021-11-11        8821
#> 644 2021-11-12        9644
#> 645 2021-11-13        7263
#> 646 2021-11-14        4865
#> 647 2021-11-15       11983
#> 648 2021-11-16       11366
#> 649 2021-11-17       10607
#> 650 2021-11-18        9331
#> 651 2021-11-19        7955
#> 652 2021-11-20        5175
#> 653 2021-11-21        5422
#> 654 2021-11-22        9833
#> 655 2021-11-23        9465
#> 656 2021-11-24        8455
#> 657 2021-11-25        1323
#> 658 2021-11-26        7072
#> 659 2021-11-27        6914
#> 660 2021-11-28        6204
#> 661 2021-11-29       13819
#> 662 2021-11-30       11333
#> 663 2021-12-01       11921
#> 664 2021-12-02       10122
#> 665 2021-12-03        9633
#> 666 2021-12-04        7071
#> 667 2021-12-05        4777
#> 668 2021-12-06       12106
#> 669 2021-12-07       10903
#> 670 2021-12-08       11240
#> 671 2021-12-09        9915
#> 672 2021-12-10        8914
#> 673 2021-12-11        6057
#> 674 2021-12-12        4720
#> 675 2021-12-13       11140
#> 676 2021-12-14       12862
#> 677 2021-12-15       15138
#> 678 2021-12-16       12523
#> 679 2021-12-17       12557
#> 680 2021-12-18        8196
#> 681 2021-12-19        6027
#> 682 2021-12-20       14750
#> 683 2021-12-21       18251
#> 684 2021-12-22       19904
#> 685 2021-12-23       15182
#> 686 2021-12-24        5677
#> 687 2021-12-25        2659
#> 688 2021-12-26       10188
#> 689 2021-12-27       21520
#> 690 2021-12-28       20962
#> 691 2021-12-29       21736
#> 692 2021-12-30       20268
#> 693 2021-12-31        7618
#> 694 2022-01-01        4499
#> 695 2022-01-02       10814
#> 696 2022-01-03       23420
#> 697 2022-01-04       26439
#> 698 2022-01-05       18664
#> 699 2022-01-06       16732
#> 700 2022-01-07       15153
#> 701 2022-01-08        9601
#> 702 2022-01-09        8651
#> 703 2022-01-10       19770
#> 704 2022-01-11       18999
#> 705 2022-01-12       16792
#> 706 2022-01-13       15681
#> 707 2022-01-14       15172
#> 708 2022-01-15       12675
#> 709 2022-01-16        7810
#> 710 2022-01-17       13746
#> 711 2022-01-18       20603
#> 712 2022-01-19       20543
#> 713 2022-01-20       15321
#> 714 2022-01-21       13351
#> 715 2022-01-22        8807
#> 716 2022-01-23        7511
#> 717 2022-01-24       16494
#> 718 2022-01-25       15051
#> 719 2022-01-26       14172
#> 720 2022-01-27       14673
#> 721 2022-01-28       11396
#> 722 2022-01-29        8285
#> 723 2022-01-30        7207
#> 724 2022-01-31       16819
#> 725 2022-02-01       15094
#> 726 2022-02-02       13750
#> 727 2022-02-03       11222
#> 728 2022-02-04       10165
#> 729 2022-02-05        6261
#> 730 2022-02-06        6162
#> 731 2022-02-07       15771
#> 732 2022-02-08       13513
#> 733 2022-02-09       12118
#> 734 2022-02-10       10772
#> 735 2022-02-11        9090
#> 736 2022-02-12        6180
#> 737 2022-02-13        4504
#> 738 2022-02-14        9617
#> 739 2022-02-15       10090
#> 740 2022-02-16        9401
#> 741 2022-02-17        7574
#> 742 2022-02-18        6652
#> 743 2022-02-19        4105
#> 744 2022-02-20        3076
#> 745 2022-02-21        5976
#> 746 2022-02-22        7940
#> 747 2022-02-23        5695
#> 748 2022-02-24        4863
#> 749 2022-02-25        4489
#> 750 2022-02-26        2943
#> 751 2022-02-27        2326
#> 752 2022-02-28        4777
#> 753 2022-03-01        4574
#> 754 2022-03-02        3577
#> 755 2022-03-03        3053
#> 756 2022-03-04        3008
#> 757 2022-03-05        1933
#> 758 2022-03-06        1646
#> 759 2022-03-07        3996
#> 760 2022-03-08        2996
#> 761 2022-03-09        2846
#> 762 2022-03-10        2720
#> 763 2022-03-11        2372
#> 764 2022-03-12        1426
#> 765 2022-03-13        1554
#> 766 2022-03-14        3435
#> 767 2022-03-15        3179
#> 768 2022-03-16        2827
#> 769 2022-03-17        2600
#> 770 2022-03-18        2440
#> 771 2022-03-19        1900
#> 772 2022-03-20        1862
#> 773 2022-03-21        3692
#> 774 2022-03-22        3194
#> 775 2022-03-23        3189
#> 776 2022-03-24        2910
#> 777 2022-03-25        2713
#> 778 2022-03-26        2199
#> 779 2022-03-27        2265
#> 780 2022-03-28        3875
#> 781 2022-03-29        3446
#> 782 2022-03-30        3628
#> 783 2022-03-31        3324
#> 784 2022-04-01        3293
#> 785 2022-04-02        2475
#> 786 2022-04-03        2369
#> 787 2022-04-04        5283
#> 788 2022-04-05        4578
#> 789 2022-04-06        4546
#> 790 2022-04-07        4567
#> 791 2022-04-08        4722
#> 792 2022-04-09        3248
#> 793 2022-04-10        3351
#> 794 2022-04-11        6473
#> 795 2022-04-12        5148
#> 796 2022-04-13        4780
#> 797 2022-04-14        4858
#> 798 2022-04-15        5874
#> 799 2022-04-16        3790
#> 800 2022-04-17        1966
#> 801 2022-04-18        6483
#> 802 2022-04-19        5456
#> 803 2022-04-20        5068
#> 804 2022-04-21        4871
#> 805 2022-04-22        5016
#> 806 2022-04-23        3397
#> 807 2022-04-24        3617
#> 808 2022-04-25        6394
#> 809 2022-04-26        5698
#> 810 2022-04-27        4965
#> 811 2022-04-28        4215
#> 812 2022-04-29        4526
#> 813 2022-04-30        3625
#> 814 2022-05-01        3292
#> 815 2022-05-02        6806
#> 816 2022-05-03        6245
#> 817 2022-05-04        6066
#> 818 2022-05-05        6147
#> 819 2022-05-06        5639
#> 820 2022-05-07        4317
#> 821 2022-05-08        3916
#> 822 2022-05-09        8137
#> 823 2022-05-10        7055
#> 824 2022-05-11        6629
#> 825 2022-05-12        6377
#> 826 2022-05-13        6090
#> 827 2022-05-14        4695
#> 828 2022-05-15        4739
#> 829 2022-05-16        7903
#> 830 2022-05-17        6957
#> 831 2022-05-18        6669
#> 832 2022-05-19        5472
#> 833 2022-05-20        5681
#> 834 2022-05-21        4642
#> 835 2022-05-22        4596
#> 836 2022-05-23        8092
#> 837 2022-05-24        7647
#> 838 2022-05-25        6707
#> 839 2022-05-26        5516
#> 840 2022-05-27        5984
#> 841 2022-05-28        4017
#> 842 2022-05-29        4511
#> 843 2022-05-30        2389
#> 844 2022-05-31        9872
#> 845 2022-06-01        6820
#> 846 2022-06-02        6361
#> 847 2022-06-03        5903
#> 848 2022-06-04        4127
#> 849 2022-06-05        4115
#> 850 2022-06-06        8094
#> 851 2022-06-07        7464
#> 852 2022-06-08        6613
#> 853 2022-06-09        6264
#> 854 2022-06-10        5533
#> 855 2022-06-11        3751
#> 856 2022-06-12        3591
#> 857 2022-06-13        6542
#> 858 2022-06-14        5494
#> 859 2022-06-15        6690
#> 860 2022-06-16        5613
#> 861 2022-06-17        5259
#> 862 2022-06-18        3562
#> 863 2022-06-19        2707
#> 864 2022-06-20        5548
#> 865 2022-06-21        6111
#> 866 2022-06-22        6115
#> 867 2022-06-23        5507
#> 868 2022-06-24        4997
#> 869 2022-06-25        3490
#> 870 2022-06-26        3687
#> 871 2022-06-27        7638
#> 872 2022-06-28        6096
#> 873 2022-06-29        5651
#> 874 2022-06-30        5575
#> 875 2022-07-01        5946
#> 876 2022-07-02        3770
#> 877 2022-07-03        3638
#> 878 2022-07-04        2299
#> 879 2022-07-05        9819
#> 880 2022-07-06        7351
#> 881 2022-07-07        6341
#> 882 2022-07-08        6321
#> 883 2022-07-09        4606
#> 884 2022-07-10        5001
#> 885 2022-07-11        8210
#> 886 2022-07-12        6398
#> 887 2022-07-13        6136
#> 888 2022-07-14        6193
#> 889 2022-07-15        6188
#> 890 2022-07-16        5444
#> 891 2022-07-17        5274
#> 892 2022-07-18        8042
#> 893 2022-07-19        6599
#> 894 2022-07-20        5316
#> 895 2022-07-21        5005
#> 896 2022-07-22        5912
#> 897 2022-07-23        3659
#> 898 2022-07-24        3982
#> 899 2022-07-25        7378
#> 900 2022-07-26        5668
#> 901 2022-07-27        4938
#> 902 2022-07-28        4797
#> 903 2022-07-29        4730
#> 904 2022-07-30        3676
#> 905 2022-07-31        4080
#> 906 2022-08-01        7766
#> 907 2022-08-02        5579
#> 908 2022-08-03        4907
#> 909 2022-08-04        4542
#> 910 2022-08-05        3868
#> 911 2022-08-06        2613
#> 912 2022-08-07        2604
#> 913 2022-08-08        5038
#> 914 2022-08-09        3902
#> 915 2022-08-10        3980
#> 916 2022-08-11        4613
#> 917 2022-08-12        4394
#> 918 2022-08-13        3399
#> 919 2022-08-14        3037
#> 920 2022-08-15        5050
#> 921 2022-08-16        4258
#> 922 2022-08-17        4163
#> 923 2022-08-18        3889
#> 924 2022-08-19        4189
#> 925 2022-08-20        3016
#> 926 2022-08-21        3268
#> 927 2022-08-22        5393
#> 928 2022-08-23        3866
#> 929 2022-08-24        3450
#> 930 2022-08-25        3264
#> 931 2022-08-26        3939
#> 932 2022-08-27        3695
#> 933 2022-08-28        3188
#> 934 2022-08-29        4782
#> 935 2022-08-30        3480
#> 936 2022-08-31        2895
#> 937 2022-09-01        2719
#> 938 2022-09-02        2492
#> 939 2022-09-03        2525
#> 940 2022-09-04        2835
#> 941 2022-09-05         731
#> 942 2022-09-06        4588
#> 943 2022-09-07        2407
#> 944 2022-09-08        1900
#> 945 2022-09-09        1662
#> 946 2022-09-10        2121
#> 947 2022-09-11        1593
#> 948 2022-09-12        2331
#> 949 2022-09-13        1303
#> 950 2022-09-14         503
#> 951 2022-09-15         423
#> 952 2022-09-16         250
#> 953 2022-09-17         152
#> 954 2022-09-18          71
#> 955 2022-09-19          80
#> 956 2022-09-20          21
#> 957 2022-09-21          52
#> 958 2022-09-22           4
#> 959 2022-09-23           3
#> 960 2022-09-24           1

# Returns the total number of global sequences in the outbreak.info API.
getSeqCounts(cumulative=TRUE)
#> Retrieving data... 
#>           key    value
#> 1 total_count 12885411