Retrieve sequence count data

Retrieves number of sequences by location, by date of sample collection. See getSubmissionDateByLocation to view similar information by date of sequence submission.

getSeqCounts(location = NULL, cumulative = NULL, subadmin = NULL)

Arguments

location

(optional) a location name (if not specified, global total counts returned)

cumulative

(optional) Boolean (T/F), T returns cumulative number of sequences to date

subadmin

(optional) Boolean (T/F), subadmin=T and cumulative=T returns cumulative number of sequences for next administrative level

Value

dataframe

Examples

# Retrieves the number of samples sequenced in the U.S. each day, by date of sample collection
getSeqCounts(location="United States")
#> Retrieving data... 
#>           date total_count
#> 1   2020-01-26           3
#> 2   2020-01-27           1
#> 3   2020-01-28           2
#> 4   2020-01-29           4
#> 5   2020-01-31           1
#> 6   2020-02-02           1
#> 7   2020-02-06           1
#> 8   2020-02-08           1
#> 9   2020-02-10           1
#> 10  2020-02-11           1
#> 11  2020-02-17           9
#> 12  2020-02-18           6
#> 13  2020-02-20           2
#> 14  2020-02-21           8
#> 15  2020-02-22           3
#> 16  2020-02-23           1
#> 17  2020-02-24           7
#> 18  2020-02-25           4
#> 19  2020-02-26           7
#> 20  2020-02-27           9
#> 21  2020-02-28          27
#> 22  2020-02-29          37
#> 23  2020-03-01          26
#> 24  2020-03-02          49
#> 25  2020-03-03          43
#> 26  2020-03-04          68
#> 27  2020-03-05         104
#> 28  2020-03-06          84
#> 29  2020-03-07          94
#> 30  2020-03-08          98
#> 31  2020-03-09         192
#> 32  2020-03-10         190
#> 33  2020-03-11         240
#> 34  2020-03-12         381
#> 35  2020-03-13         550
#> 36  2020-03-14         531
#> 37  2020-03-15         999
#> 38  2020-03-16         933
#> 39  2020-03-17         770
#> 40  2020-03-18         828
#> 41  2020-03-19         684
#> 42  2020-03-20         593
#> 43  2020-03-21         459
#> 44  2020-03-22         383
#> 45  2020-03-23         626
#> 46  2020-03-24         720
#> 47  2020-03-25         731
#> 48  2020-03-26         700
#> 49  2020-03-27         670
#> 50  2020-03-28         472
#> 51  2020-03-29         349
#> 52  2020-03-30         746
#> 53  2020-03-31         672
#> 54  2020-04-01         832
#> 55  2020-04-02         703
#> 56  2020-04-03         733
#> 57  2020-04-04         447
#> 58  2020-04-05         372
#> 59  2020-04-06         644
#> 60  2020-04-07         531
#> 61  2020-04-08         605
#> 62  2020-04-09         451
#> 63  2020-04-10         510
#> 64  2020-04-11         399
#> 65  2020-04-12         253
#> 66  2020-04-13         700
#> 67  2020-04-14         468
#> 68  2020-04-15         835
#> 69  2020-04-16         413
#> 70  2020-04-17         457
#> 71  2020-04-18         327
#> 72  2020-04-19         208
#> 73  2020-04-20         531
#> 74  2020-04-21         548
#> 75  2020-04-22         433
#> 76  2020-04-23         434
#> 77  2020-04-24         453
#> 78  2020-04-25         249
#> 79  2020-04-26         180
#> 80  2020-04-27         531
#> 81  2020-04-28         509
#> 82  2020-04-29         594
#> 83  2020-04-30         640
#> 84  2020-05-01         542
#> 85  2020-05-02         278
#> 86  2020-05-03         167
#> 87  2020-05-04         551
#> 88  2020-05-05         496
#> 89  2020-05-06         395
#> 90  2020-05-07         426
#> 91  2020-05-08         368
#> 92  2020-05-09         243
#> 93  2020-05-10         129
#> 94  2020-05-11         420
#> 95  2020-05-12         413
#> 96  2020-05-13         453
#> 97  2020-05-14         572
#> 98  2020-05-15         435
#> 99  2020-05-16         155
#> 100 2020-05-17         119
#> 101 2020-05-18         334
#> 102 2020-05-19         372
#> 103 2020-05-20         354
#> 104 2020-05-21         398
#> 105 2020-05-22         406
#> 106 2020-05-23         148
#> 107 2020-05-24         129
#> 108 2020-05-25         146
#> 109 2020-05-26         437
#> 110 2020-05-27         379
#> 111 2020-05-28         273
#> 112 2020-05-29         378
#> 113 2020-05-30         171
#> 114 2020-05-31         155
#> 115 2020-06-01         336
#> 116 2020-06-02         315
#> 117 2020-06-03         376
#> 118 2020-06-04         359
#> 119 2020-06-05         355
#> 120 2020-06-06         149
#> 121 2020-06-07         119
#> 122 2020-06-08         396
#> 123 2020-06-09         387
#> 124 2020-06-10         365
#> 125 2020-06-11         365
#> 126 2020-06-12         306
#> 127 2020-06-13         354
#> 128 2020-06-14         216
#> 129 2020-06-15         641
#> 130 2020-06-16         601
#> 131 2020-06-17         425
#> 132 2020-06-18         673
#> 133 2020-06-19         634
#> 134 2020-06-20         369
#> 135 2020-06-21         420
#> 136 2020-06-22         748
#> 137 2020-06-23         625
#> 138 2020-06-24         745
#> 139 2020-06-25         679
#> 140 2020-06-26         681
#> 141 2020-06-27         343
#> 142 2020-06-28         354
#> 143 2020-06-29         825
#> 144 2020-06-30         707
#> 145 2020-07-01         833
#> 146 2020-07-02         855
#> 147 2020-07-03         542
#> 148 2020-07-04         327
#> 149 2020-07-05         321
#> 150 2020-07-06         776
#> 151 2020-07-07         948
#> 152 2020-07-08        1061
#> 153 2020-07-09        1016
#> 154 2020-07-10         822
#> 155 2020-07-11         407
#> 156 2020-07-12         375
#> 157 2020-07-13         909
#> 158 2020-07-14         929
#> 159 2020-07-15         973
#> 160 2020-07-16         726
#> 161 2020-07-17         845
#> 162 2020-07-18         458
#> 163 2020-07-19         406
#> 164 2020-07-20         810
#> 165 2020-07-21         925
#> 166 2020-07-22         752
#> 167 2020-07-23         688
#> 168 2020-07-24         542
#> 169 2020-07-25         352
#> 170 2020-07-26         227
#> 171 2020-07-27         772
#> 172 2020-07-28         657
#> 173 2020-07-29         578
#> 174 2020-07-30         631
#> 175 2020-07-31         585
#> 176 2020-08-01         353
#> 177 2020-08-02         179
#> 178 2020-08-03         683
#> 179 2020-08-04         611
#> 180 2020-08-05         570
#> 181 2020-08-06         536
#> 182 2020-08-07         478
#> 183 2020-08-08         243
#> 184 2020-08-09         203
#> 185 2020-08-10         515
#> 186 2020-08-11         592
#> 187 2020-08-12         486
#> 188 2020-08-13         439
#> 189 2020-08-14         432
#> 190 2020-08-15         217
#> 191 2020-08-16         175
#> 192 2020-08-17         505
#> 193 2020-08-18         464
#> 194 2020-08-19         443
#> 195 2020-08-20         407
#> 196 2020-08-21         352
#> 197 2020-08-22         232
#> 198 2020-08-23         121
#> 199 2020-08-24         425
#> 200 2020-08-25         498
#> 201 2020-08-26         744
#> 202 2020-08-27         472
#> 203 2020-08-28         405
#> 204 2020-08-29         205
#> 205 2020-08-30         145
#> 206 2020-08-31         498
#> 207 2020-09-01         501
#> 208 2020-09-02         478
#> 209 2020-09-03         437
#> 210 2020-09-04         378
#> 211 2020-09-05         185
#> 212 2020-09-06         111
#> 213 2020-09-07         111
#> 214 2020-09-08         498
#> 215 2020-09-09         715
#> 216 2020-09-10         460
#> 217 2020-09-11         494
#> 218 2020-09-12         188
#> 219 2020-09-13          90
#> 220 2020-09-14         549
#> 221 2020-09-15         632
#> 222 2020-09-16         622
#> 223 2020-09-17         451
#> 224 2020-09-18         436
#> 225 2020-09-19         189
#> 226 2020-09-20         129
#> 227 2020-09-21         401
#> 228 2020-09-22         432
#> 229 2020-09-23         797
#> 230 2020-09-24         536
#> 231 2020-09-25         434
#> 232 2020-09-26         278
#> 233 2020-09-27         139
#> 234 2020-09-28         503
#> 235 2020-09-29         587
#> 236 2020-09-30         572
#> 237 2020-10-01         539
#> 238 2020-10-02         544
#> 239 2020-10-03         207
#> 240 2020-10-04         225
#> 241 2020-10-05         712
#> 242 2020-10-06         662
#> 243 2020-10-07         761
#> 244 2020-10-08         957
#> 245 2020-10-09         624
#> 246 2020-10-10         308
#> 247 2020-10-11         161
#> 248 2020-10-12         512
#> 249 2020-10-13         703
#> 250 2020-10-14         615
#> 251 2020-10-15         794
#> 252 2020-10-16         766
#> 253 2020-10-17         462
#> 254 2020-10-18         225
#> 255 2020-10-19        1075
#> 256 2020-10-20         761
#> 257 2020-10-21         787
#> 258 2020-10-22         877
#> 259 2020-10-23         653
#> 260 2020-10-24         333
#> 261 2020-10-25         245
#> 262 2020-10-26        1272
#> 263 2020-10-27         993
#> 264 2020-10-28         999
#> 265 2020-10-29        1170
#> 266 2020-10-30         927
#> 267 2020-10-31         371
#> 268 2020-11-01         510
#> 269 2020-11-02        1454
#> 270 2020-11-03        1399
#> 271 2020-11-04        1935
#> 272 2020-11-05        1801
#> 273 2020-11-06        1612
#> 274 2020-11-07         709
#> 275 2020-11-08         480
#> 276 2020-11-09        1632
#> 277 2020-11-10        1492
#> 278 2020-11-11        1457
#> 279 2020-11-12        1269
#> 280 2020-11-13        1173
#> 281 2020-11-14         686
#> 282 2020-11-15         745
#> 283 2020-11-16        1625
#> 284 2020-11-17        1599
#> 285 2020-11-18        1399
#> 286 2020-11-19        1035
#> 287 2020-11-20        1020
#> 288 2020-11-21         383
#> 289 2020-11-22         288
#> 290 2020-11-23        1342
#> 291 2020-11-24        1324
#> 292 2020-11-25         998
#> 293 2020-11-26         181
#> 294 2020-11-27         781
#> 295 2020-11-28         412
#> 296 2020-11-29         417
#> 297 2020-11-30        2263
#> 298 2020-12-01        2168
#> 299 2020-12-02        1805
#> 300 2020-12-03        1326
#> 301 2020-12-04        1157
#> 302 2020-12-05         393
#> 303 2020-12-06         318
#> 304 2020-12-07        1310
#> 305 2020-12-08        1251
#> 306 2020-12-09        2067
#> 307 2020-12-10        2668
#> 308 2020-12-11        2722
#> 309 2020-12-12        1714
#> 310 2020-12-13         832
#> 311 2020-12-14        2526
#> 312 2020-12-15        3241
#> 313 2020-12-16        2199
#> 314 2020-12-17        1266
#> 315 2020-12-18        1197
#> 316 2020-12-19         950
#> 317 2020-12-20         664
#> 318 2020-12-21        1743
#> 319 2020-12-22        1297
#> 320 2020-12-23        1276
#> 321 2020-12-24        1062
#> 322 2020-12-25         343
#> 323 2020-12-26        1861
#> 324 2020-12-27        1398
#> 325 2020-12-28        2594
#> 326 2020-12-29        2500
#> 327 2020-12-30        1849
#> 328 2020-12-31        1482
#> 329 2021-01-01         801
#> 330 2021-01-02        1469
#> 331 2021-01-03        1548
#> 332 2021-01-04        4722
#> 333 2021-01-05        4101
#> 334 2021-01-06        3024
#> 335 2021-01-07        2662
#> 336 2021-01-08        2189
#> 337 2021-01-09        2087
#> 338 2021-01-10        1187
#> 339 2021-01-11        3408
#> 340 2021-01-12        3506
#> 341 2021-01-13        3111
#> 342 2021-01-14        3469
#> 343 2021-01-15        4250
#> 344 2021-01-16        2639
#> 345 2021-01-17        1740
#> 346 2021-01-18        3254
#> 347 2021-01-19        4564
#> 348 2021-01-20        2941
#> 349 2021-01-21        2813
#> 350 2021-01-22        2142
#> 351 2021-01-23        1244
#> 352 2021-01-24        1281
#> 353 2021-01-25        2665
#> 354 2021-01-26        2716
#> 355 2021-01-27        3075
#> 356 2021-01-28        2980
#> 357 2021-01-29        2336
#> 358 2021-01-30        1664
#> 359 2021-01-31        1764
#> 360 2021-02-01        3883
#> 361 2021-02-02        4196
#> 362 2021-02-03        3746
#> 363 2021-02-04        2799
#> 364 2021-02-05        2431
#> 365 2021-02-06        1629
#> 366 2021-02-07        1396
#> 367 2021-02-08        3642
#> 368 2021-02-09        4140
#> 369 2021-02-10        3853
#> 370 2021-02-11        3314
#> 371 2021-02-12        2607
#> 372 2021-02-13        1872
#> 373 2021-02-14        1447
#> 374 2021-02-15        3855
#> 375 2021-02-16        3142
#> 376 2021-02-17        2992
#> 377 2021-02-18        2361
#> 378 2021-02-19        2387
#> 379 2021-02-20        1879
#> 380 2021-02-21        1460
#> 381 2021-02-22        3357
#> 382 2021-02-23        3058
#> 383 2021-02-24        3234
#> 384 2021-02-25        2866
#> 385 2021-02-26        2761
#> 386 2021-02-27        2085
#> 387 2021-02-28        1604
#> 388 2021-03-01        3503
#> 389 2021-03-02        3531
#> 390 2021-03-03        3928
#> 391 2021-03-04        3521
#> 392 2021-03-05        3022
#> 393 2021-03-06        2121
#> 394 2021-03-07        1975
#> 395 2021-03-08        4226
#> 396 2021-03-09        3702
#> 397 2021-03-10        3765
#> 398 2021-03-11        3275
#> 399 2021-03-12        4113
#> 400 2021-03-13        2754
#> 401 2021-03-14        2219
#> 402 2021-03-15        6240
#> 403 2021-03-16        4651
#> 404 2021-03-17        4421
#> 405 2021-03-18        4601
#> 406 2021-03-19        4278
#> 407 2021-03-20        3275
#> 408 2021-03-21        2893
#> 409 2021-03-22        5367
#> 410 2021-03-23        5099
#> 411 2021-03-24        5697
#> 412 2021-03-25        5288
#> 413 2021-03-26        4467
#> 414 2021-03-27        3612
#> 415 2021-03-28        2974
#> 416 2021-03-29        6451
#> 417 2021-03-30        5711
#> 418 2021-03-31        6563
#> 419 2021-04-01        6916
#> 420 2021-04-02        5595
#> 421 2021-04-03        4059
#> 422 2021-04-04        2745
#> 423 2021-04-05        7298
#> 424 2021-04-06        6399
#> 425 2021-04-07        6814
#> 426 2021-04-08        6291
#> 427 2021-04-09        6600
#> 428 2021-04-10        4826
#> 429 2021-04-11        3841
#> 430 2021-04-12        7368
#> 431 2021-04-13        6874
#> 432 2021-04-14        6945
#> 433 2021-04-15        7999
#> 434 2021-04-16        5919
#> 435 2021-04-17        4111
#> 436 2021-04-18        2872
#> 437 2021-04-19        6786
#> 438 2021-04-20        5981
#> 439 2021-04-21        5769
#> 440 2021-04-22        5542
#> 441 2021-04-23        5047
#> 442 2021-04-24        3179
#> 443 2021-04-25        2826
#> 444 2021-04-26        6542
#> 445 2021-04-27        5918
#> 446 2021-04-28        5732
#> 447 2021-04-29        5668
#> 448 2021-04-30        5311
#> 449 2021-05-01        3268
#> 450 2021-05-02        2588
#> 451 2021-05-03        6582
#> 452 2021-05-04        5165
#> 453 2021-05-05        5420
#> 454 2021-05-06        4493
#> 455 2021-05-07        4073
#> 456 2021-05-08        2298
#> 457 2021-05-09        1920
#> 458 2021-05-10        4933
#> 459 2021-05-11        4298
#> 460 2021-05-12        3908
#> 461 2021-05-13        3517
#> 462 2021-05-14        2844
#> 463 2021-05-15        2741
#> 464 2021-05-16        1482
#> 465 2021-05-17        3933
#> 466 2021-05-18        3336
#> 467 2021-05-19        3098
#> 468 2021-05-20        2805
#> 469 2021-05-21        2419
#> 470 2021-05-22        1389
#> 471 2021-05-23         995
#> 472 2021-05-24        2709
#> 473 2021-05-25        2342
#> 474 2021-05-26        2135
#> 475 2021-05-27        2010
#> 476 2021-05-28        1694
#> 477 2021-05-29         879
#> 478 2021-05-30         745
#> 479 2021-05-31         537
#> 480 2021-06-01        2426
#> 481 2021-06-02        1890
#> 482 2021-06-03        1797
#> 483 2021-06-04        1571
#> 484 2021-06-05         931
#> 485 2021-06-06         832
#> 486 2021-06-07        1961
#> 487 2021-06-08        1808
#> 488 2021-06-09        1730
#> 489 2021-06-10        1787
#> 490 2021-06-11        1724
#> 491 2021-06-12        1040
#> 492 2021-06-13        1015
#> 493 2021-06-14        2159
#> 494 2021-06-15        2037
#> 495 2021-06-16        1767
#> 496 2021-06-17        1724
#> 497 2021-06-18        1704
#> 498 2021-06-19        1101
#> 499 2021-06-20         885
#> 500 2021-06-21        2353
#> 501 2021-06-22        2096
#> 502 2021-06-23        2004
#> 503 2021-06-24        2010
#> 504 2021-06-25        1859
#> 505 2021-06-26        1266
#> 506 2021-06-27        1250
#> 507 2021-06-28        2967
#> 508 2021-06-29        2628
#> 509 2021-06-30        2474
#> 510 2021-07-01        2610
#> 511 2021-07-02        2442
#> 512 2021-07-03        1503
#> 513 2021-07-04         672
#> 514 2021-07-05        2454
#> 515 2021-07-06        4365
#> 516 2021-07-07        4377
#> 517 2021-07-08        4185
#> 518 2021-07-09        4067
#> 519 2021-07-10        3140
#> 520 2021-07-11        2677
#> 521 2021-07-12        5968
#> 522 2021-07-13        5213
#> 523 2021-07-14        5311
#> 524 2021-07-15        7185
#> 525 2021-07-16        5780
#> 526 2021-07-17        4310
#> 527 2021-07-18        3788
#> 528 2021-07-19        8491
#> 529 2021-07-20        8567
#> 530 2021-07-21        8682
#> 531 2021-07-22        8476
#> 532 2021-07-23        8395
#> 533 2021-07-24        6321
#> 534 2021-07-25        4983
#> 535 2021-07-26       11952
#> 536 2021-07-27       10811
#> 537 2021-07-28       11017
#> 538 2021-07-29       12116
#> 539 2021-07-30       11507
#> 540 2021-07-31        8124
#> 541 2021-08-01        6871
#> 542 2021-08-02       14082
#> 543 2021-08-03       14003
#> 544 2021-08-04       12869
#> 545 2021-08-05       13083
#> 546 2021-08-06       12917
#> 547 2021-08-07        8187
#> 548 2021-08-08        5670
#> 549 2021-08-09       13613
#> 550 2021-08-10       12568
#> 551 2021-08-11       11135
#> 552 2021-08-12       10861
#> 553 2021-08-13        9366
#> 554 2021-08-14        5960
#> 555 2021-08-15        8531
#> 556 2021-08-16       12850
#> 557 2021-08-17       11823
#> 558 2021-08-18       11038
#> 559 2021-08-19       10107
#> 560 2021-08-20       10492
#> 561 2021-08-21        5822
#> 562 2021-08-22        4083
#> 563 2021-08-23       12089
#> 564 2021-08-24       12475
#> 565 2021-08-25       11740
#> 566 2021-08-26       12218
#> 567 2021-08-27       10127
#> 568 2021-08-28        5356
#> 569 2021-08-29        4122
#> 570 2021-08-30       12802
#> 571 2021-08-31       14269
#> 572 2021-09-01       12374
#> 573 2021-09-02       12341
#> 574 2021-09-03       10527
#> 575 2021-09-04        6075
#> 576 2021-09-05        5257
#> 577 2021-09-06        4754
#> 578 2021-09-07       16601
#> 579 2021-09-08       13369
#> 580 2021-09-09       13143
#> 581 2021-09-10       12446
#> 582 2021-09-11        6947
#> 583 2021-09-12        6082
#> 584 2021-09-13       13252
#> 585 2021-09-14       10744
#> 586 2021-09-15       13535
#> 587 2021-09-16       10136
#> 588 2021-09-17       10943
#> 589 2021-09-18        5946
#> 590 2021-09-19        5267
#> 591 2021-09-20       12212
#> 592 2021-09-21       10717
#> 593 2021-09-22        8340
#> 594 2021-09-23        8455
#> 595 2021-09-24        8927
#> 596 2021-09-25        5042
#> 597 2021-09-26        4206
#> 598 2021-09-27       10698
#> 599 2021-09-28        9828
#> 600 2021-09-29       10120
#> 601 2021-09-30        8720
#> 602 2021-10-01        8844
#> 603 2021-10-02        5475
#> 604 2021-10-03        4785
#> 605 2021-10-04       10881
#> 606 2021-10-05       10693
#> 607 2021-10-06        9772
#> 608 2021-10-07        9092
#> 609 2021-10-08        8554
#> 610 2021-10-09        5752
#> 611 2021-10-10        4871
#> 612 2021-10-11        9828
#> 613 2021-10-12        8581
#> 614 2021-10-13        8285
#> 615 2021-10-14        8970
#> 616 2021-10-15        9960
#> 617 2021-10-16        4823
#> 618 2021-10-17        4069
#> 619 2021-10-18       10313
#> 620 2021-10-19        8853
#> 621 2021-10-20        8372
#> 622 2021-10-21        8338
#> 623 2021-10-22        8636
#> 624 2021-10-23        5683
#> 625 2021-10-24        4441
#> 626 2021-10-25       10826
#> 627 2021-10-26        9143
#> 628 2021-10-27        9264
#> 629 2021-10-28        8850
#> 630 2021-10-29        8116
#> 631 2021-10-30        5506
#> 632 2021-10-31        4441
#> 633 2021-11-01        9772
#> 634 2021-11-02       10386
#> 635 2021-11-03       10642
#> 636 2021-11-04        9803
#> 637 2021-11-05        8791
#> 638 2021-11-06        5894
#> 639 2021-11-07        5535
#> 640 2021-11-08       10580
#> 641 2021-11-09       10939
#> 642 2021-11-10       11420
#> 643 2021-11-11        8781
#> 644 2021-11-12        9609
#> 645 2021-11-13        7199
#> 646 2021-11-14        4849
#> 647 2021-11-15       11882
#> 648 2021-11-16       11250
#> 649 2021-11-17       10542
#> 650 2021-11-18        9287
#> 651 2021-11-19        7892
#> 652 2021-11-20        5096
#> 653 2021-11-21        5364
#> 654 2021-11-22        9737
#> 655 2021-11-23        9401
#> 656 2021-11-24        8382
#> 657 2021-11-25        1301
#> 658 2021-11-26        6991
#> 659 2021-11-27        6791
#> 660 2021-11-28        6116
#> 661 2021-11-29       13669
#> 662 2021-11-30       11275
#> 663 2021-12-01       11826
#> 664 2021-12-02       10048
#> 665 2021-12-03        9583
#> 666 2021-12-04        7040
#> 667 2021-12-05        4751
#> 668 2021-12-06       12034
#> 669 2021-12-07       10831
#> 670 2021-12-08       11189
#> 671 2021-12-09        9858
#> 672 2021-12-10        8874
#> 673 2021-12-11        6030
#> 674 2021-12-12        4697
#> 675 2021-12-13       11046
#> 676 2021-12-14       12837
#> 677 2021-12-15       14954
#> 678 2021-12-16       12408
#> 679 2021-12-17       12382
#> 680 2021-12-18        8072
#> 681 2021-12-19        5891
#> 682 2021-12-20       14477
#> 683 2021-12-21       18004
#> 684 2021-12-22       19572
#> 685 2021-12-23       14950
#> 686 2021-12-24        5570
#> 687 2021-12-25        2640
#> 688 2021-12-26        9963
#> 689 2021-12-27       21141
#> 690 2021-12-28       20645
#> 691 2021-12-29       21343
#> 692 2021-12-30       19932
#> 693 2021-12-31        7440
#> 694 2022-01-01        4443
#> 695 2022-01-02       10656
#> 696 2022-01-03       22720
#> 697 2022-01-04       24903
#> 698 2022-01-05       17745
#> 699 2022-01-06       15362
#> 700 2022-01-07       14648
#> 701 2022-01-08        9507
#> 702 2022-01-09        8451
#> 703 2022-01-10       19106
#> 704 2022-01-11       18494
#> 705 2022-01-12       16172
#> 706 2022-01-13       15227
#> 707 2022-01-14       14791
#> 708 2022-01-15       12401
#> 709 2022-01-16        7701
#> 710 2022-01-17       13500
#> 711 2022-01-18       20175
#> 712 2022-01-19       20018
#> 713 2022-01-20       14921
#> 714 2022-01-21       13016
#> 715 2022-01-22        8731
#> 716 2022-01-23        7432
#> 717 2022-01-24       16156
#> 718 2022-01-25       14788
#> 719 2022-01-26       13933
#> 720 2022-01-27       14308
#> 721 2022-01-28       10915
#> 722 2022-01-29        8140
#> 723 2022-01-30        6945
#> 724 2022-01-31       15539
#> 725 2022-02-01       14011
#> 726 2022-02-02       12620
#> 727 2022-02-03       10957
#> 728 2022-02-04        9989
#> 729 2022-02-05        6165
#> 730 2022-02-06        6035
#> 731 2022-02-07       14813
#> 732 2022-02-08       12747
#> 733 2022-02-09       11548
#> 734 2022-02-10       10403
#> 735 2022-02-11        8600
#> 736 2022-02-12        5844
#> 737 2022-02-13        4278
#> 738 2022-02-14        9125
#> 739 2022-02-15        9667
#> 740 2022-02-16        8897
#> 741 2022-02-17        7183
#> 742 2022-02-18        6289
#> 743 2022-02-19        3994
#> 744 2022-02-20        2986
#> 745 2022-02-21        5858
#> 746 2022-02-22        7678
#> 747 2022-02-23        5384
#> 748 2022-02-24        4572
#> 749 2022-02-25        4299
#> 750 2022-02-26        2888
#> 751 2022-02-27        2267
#> 752 2022-02-28        4686
#> 753 2022-03-01        4519
#> 754 2022-03-02        3469
#> 755 2022-03-03        2922
#> 756 2022-03-04        2939
#> 757 2022-03-05        1911
#> 758 2022-03-06        1628
#> 759 2022-03-07        3916
#> 760 2022-03-08        2940
#> 761 2022-03-09        2797
#> 762 2022-03-10        2696
#> 763 2022-03-11        2332
#> 764 2022-03-12        1409
#> 765 2022-03-13        1535
#> 766 2022-03-14        3376
#> 767 2022-03-15        3017
#> 768 2022-03-16        2776
#> 769 2022-03-17        2528
#> 770 2022-03-18        2363
#> 771 2022-03-19        1863
#> 772 2022-03-20        1839
#> 773 2022-03-21        3615
#> 774 2022-03-22        3107
#> 775 2022-03-23        3148
#> 776 2022-03-24        2874
#> 777 2022-03-25        2670
#> 778 2022-03-26        2158
#> 779 2022-03-27        2199
#> 780 2022-03-28        3799
#> 781 2022-03-29        3369
#> 782 2022-03-30        3532
#> 783 2022-03-31        3257
#> 784 2022-04-01        3239
#> 785 2022-04-02        2442
#> 786 2022-04-03        2311
#> 787 2022-04-04        5175
#> 788 2022-04-05        4496
#> 789 2022-04-06        4236
#> 790 2022-04-07        4486
#> 791 2022-04-08        4531
#> 792 2022-04-09        2722
#> 793 2022-04-10        2762
#> 794 2022-04-11        5641
#> 795 2022-04-12        4356
#> 796 2022-04-13        4068
#> 797 2022-04-14        4141
#> 798 2022-04-15        5159
#> 799 2022-04-16        3343
#> 800 2022-04-17        1513
#> 801 2022-04-18        5755
#> 802 2022-04-19        4903
#> 803 2022-04-20        4552
#> 804 2022-04-21        4279
#> 805 2022-04-22        4430
#> 806 2022-04-23        3150
#> 807 2022-04-24        3418
#> 808 2022-04-25        5700
#> 809 2022-04-26        5061
#> 810 2022-04-27        4622
#> 811 2022-04-28        3965
#> 812 2022-04-29        4276
#> 813 2022-04-30        3356
#> 814 2022-05-01        2914
#> 815 2022-05-02        6097
#> 816 2022-05-03        5538
#> 817 2022-05-04        5375
#> 818 2022-05-05        5407
#> 819 2022-05-06        5039
#> 820 2022-05-07        3973
#> 821 2022-05-08        3551
#> 822 2022-05-09        7305
#> 823 2022-05-10        6099
#> 824 2022-05-11        5582
#> 825 2022-05-12        5529
#> 826 2022-05-13        5385
#> 827 2022-05-14        4295
#> 828 2022-05-15        4259
#> 829 2022-05-16        6621
#> 830 2022-05-17        6104
#> 831 2022-05-18        5750
#> 832 2022-05-19        4713
#> 833 2022-05-20        4832
#> 834 2022-05-21        4161
#> 835 2022-05-22        4064
#> 836 2022-05-23        6745
#> 837 2022-05-24        6496
#> 838 2022-05-25        5473
#> 839 2022-05-26        3745
#> 840 2022-05-27        4235
#> 841 2022-05-28        3126
#> 842 2022-05-29        3568
#> 843 2022-05-30        1259
#> 844 2022-05-31        5814
#> 845 2022-06-01        3133
#> 846 2022-06-02        2971
#> 847 2022-06-03        2277
#> 848 2022-06-04        1382
#> 849 2022-06-05         711
#> 850 2022-06-06        1420
#> 851 2022-06-07        1258
#> 852 2022-06-08        1082
#> 853 2022-06-09         686
#> 854 2022-06-10         566
#> 855 2022-06-11         359
#> 856 2022-06-12         160
#> 857 2022-06-13         197
#> 858 2022-06-14         102
#> 859 2022-06-15          12
#> 860 2022-06-16          16
#> 861 2022-06-17           2
#> 862 2022-06-20           1

# Returns the total number of global sequences in the outbreak.info API.
getSeqCounts(cumulative=TRUE)
#> Retrieving data... 
#>           key    value
#> 1 total_count 11134519