Lookup sublineages for a given Pango lineage or WHO name

Retrieve all sublineages of a given Pango lineage or WHO designation. Based on a phylogeny maintained by the [Pango team](https://github.com/cov-lineages/lineages-website/blob/master/data/lineages.yml) to grab all sublineages associated with a lineage, and a curated list of Variants of Concern, Variants of Interest, Variants Under Monitoring, and de-escalated variants maintained by the [outbreak.info team](https://outbreak.info/situation-reports)

lookupSublineages(lineage, returnQueryString = FALSE)

Arguments

lineage
String with the name of a Pango lineage or WHO name

returnQueryString
Boolean to return a query string to be piped into functions like getPrevalence (collapses vector by `"OR"`)

Value

list containing the sublineages associated with that lineage or WHO designation

Examples

# WHO-designated lineages
lookupSublineages("Delta")
#>   [1] "AY.1"       "AY.2"       "AY.3"       "AY.3.1"     "AY.4"      
#>   [6] "AY.4.1"     "AY.4.2"     "AY.4.2.1"   "AY.4.3"     "AY.4.4"    
#>  [11] "AY.4.5"     "AY.5"       "AY.5.1"     "AY.5.2"     "AY.5.3"    
#>  [16] "AY.5.4"     "AY.6"       "AY.7"       "AY.7.1"     "AY.7.2"    
#>  [21] "AY.8"       "AY.9"       "AY.9.1"     "AY.9.2"     "AY.9.2.1"  
#>  [26] "AY.10"      "AY.11"      "AY.12"      "AY.13"      "AY.14"     
#>  [31] "AY.15"      "AY.16"      "AY.16.1"    "AY.17"      "AY.18"     
#>  [36] "AY.19"      "AY.20"      "AY.21"      "AY.22"      "AY.23"     
#>  [41] "AY.23.1"    "AY.24"      "AY.25"      "AY.26"      "AY.27"     
#>  [46] "AY.28"      "AY.29"      "AY.29.1"    "AY.30"      "AY.31"     
#>  [51] "AY.32"      "AY.33"      "AY.34"      "AY.34.1"    "AY.35"     
#>  [56] "AY.36"      "AY.37"      "AY.38"      "AY.39"      "AY.39.1"   
#>  [61] "AY.39.1.1"  "AY.39.2"    "AY.40"      "AY.41"      "AY.42"     
#>  [66] "AY.43"      "AY.44"      "AY.45"      "AY.46"      "AY.46.1"   
#>  [71] "AY.46.2"    "AY.46.3"    "AY.46.4"    "AY.46.5"    "AY.46.6"   
#>  [76] "AY.47"      "AY.48"      "AY.49"      "AY.50"      "AY.51"     
#>  [81] "AY.52"      "AY.53"      "AY.54"      "AY.55"      "AY.56"     
#>  [86] "AY.57"      "AY.58"      "AY.59"      "AY.60"      "AY.61"     
#>  [91] "AY.62"      "AY.63"      "AY.64"      "AY.65"      "AY.66"     
#>  [96] "AY.67"      "AY.68"      "AY.69"      "AY.70"      "AY.71"     
#> [101] "AY.72"      "AY.73"      "AY.74"      "AY.75"      "AY.75.1"   
#> [106] "AY.76"      "AY.77"      "AY.78"      "AY.79"      "AY.80"     
#> [111] "AY.81"      "AY.82"      "AY.83"      "AY.84"      "AY.85"     
#> [116] "AY.86"      "AY.87"      "AY.88"      "AY.89"      "AY.90"     
#> [121] "AY.91"      "AY.91.1"    "AY.92"      "AY.93"      "AY.94"     
#> [126] "AY.95"      "AY.96"      "AY.97"      "AY.98"      "AY.98.1"   
#> [131] "AY.99"      "AY.99.1"    "AY.99.2"    "AY.100"     "AY.101"    
#> [136] "AY.102"     "AY.103"     "AY.104"     "AY.105"     "AY.106"    
#> [141] "AY.107"     "AY.108"     "AY.109"     "AY.110"     "AY.111"    
#> [146] "AY.112"     "AY.113"     "AY.114"     "AY.115"     "AY.116"    
#> [151] "AY.116.1"   "AY.117"     "AY.118"     "AY.119"     "AY.120"    
#> [156] "AY.120.1"   "AY.120.2"   "AY.120.2.1" "AY.121"     "B.1.617.2" 
lookupSublineages("epsilon")
#> [1] "B.1.427"   "B.1.429"   "B.1.429.1"
lookupSublineages("epsilon", returnQueryString = TRUE)
#> [1] "B.1.427 OR B.1.429 OR B.1.429.1"
# Pango lineage
lookupSublineages("B.1.1.7")
#> [1] "B.1.1.7" "Q.1"     "Q.2"     "Q.3"     "Q.4"     "Q.5"     "Q.6"    
#> [8] "Q.7"     "Q.8"    
#' # Not a recognized lineage
lookupSublineages("VOC-21APR-02")
#> Warning: Lineage VOC-21APR-02 not recognized as a Pango lineage
#> Error in invokeRestart("muffleWarning"): no 'restart' 'muffleWarning' found